% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{addColonyToColonies}
\alias{addColonyToColonies}
\title{Add a colony to colonies}
\usage{
addColonyToColonies(colonies, colony)
}
\arguments{
\item{colonies}{Colonies, set of colonies that will be expanded}

\item{colony}{Colony, colony that will be added}
}
\value{
Expanded \code{\link{Colonies-class}} object

Expanded Colonies object
}
\description{
Adds a colony to a set of colonies, for example, to add a new colony to an apiary.
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 6, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybees
founderDrones <- createFounderDrones(pop = basePop[1:3], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[4], fathers = founderDrones[1:10])
colony2 <- createColony(queen = basePop[5], fathers = founderDrones[11:20])
colony3 <- createColony(queen = basePop[6], fathers = founderDrones[21:30])
apiary <- c(colony1, colony2)

# Add colony to the apiary
apiaryNew <- addColonyToColonies(apiary, colony3)
getId(apiary)
getId(apiaryNew)
# ... and alternative
apiaryNew <- c(apiary, colony3)
getId(apiaryNew)

}
