% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{addDrones}
\alias{addDrones}
\title{Add (raise) drones in the colony}
\usage{
addDrones(colony, nInd, new = FALSE)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{nInd}{integer, number of drones to add}

\item{new}{logical, should the drones be added a fresh (ignoring currently
present drones)}
}
\value{
\code{\link{Colony-class}} with drones added
}
\description{
Level 2 function that adds (raises) the specified number of
  drones in the colony by crossing the current queen and the fathers. If
  there are already some drones present, new and present drones are combined.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
colony
addDrones(colony, nInd = 20)

}
