% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{addVirginQueens}
\alias{addVirginQueens}
\title{Add (raise) virgin queens in the colony}
\usage{
addVirginQueens(colony, nInd = NULL, year = NULL, simParamBee = NULL)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{nInd}{numeric or function, number of virgin queens to add; if
\code{NULL} then \code{simParamBee$nVirginQueens} is used}

\item{year}{numeric, year of birth for virgin queens}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colony-class}} with virgin queens added
}
\description{
Level 2 function that adds (raises) the specified number of
  virgin queens in the colony by crossing the current queen and the fathers.
}
\details{
\code{addVirginQueens} replaces any currently present virgin queens.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
colony
addVirginQueens(colony, nInd = 10)

# Using a default in SP$nVirginQueens
# (just to have some virgin queens - change this to your needs!)
addVirginQueens(colony)
addVirginQueens(colony)

SP$nVirginQueens <- 15
addVirginQueens(colony)
addVirginQueens(colony)

nVirginQueensFun <- function(colony) { rpois(n = 1, lambda = 15) }
addVirginQueens(colony, nInd = nVirginQueensFun)
addVirginQueens(colony, nInd = nVirginQueensFun)

SP$nVirginQueens <- nVirginQueensFun
addVirginQueens(colony)
addVirginQueens(colony)

}
