% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{addWorkers}
\alias{addWorkers}
\title{Add (raise) workers in the colony}
\usage{
addWorkers(colony, nInd = NULL, new = FALSE, simParamBee = NULL)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{nInd}{numeric or function, number of workers; if \code{NULL} then
\code{simParamBee$nWorkers} is used}

\item{new}{logical, should the workers be added a fresh (ignoring currently
present workers in the colony)}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colony-class}} with workers added
}
\description{
Level 2 function that adds (raises) the specified number of
  workers in the colony by crossing the current queen and the fathers. If
  there are already some workers present, new and present workers are
  combined.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
colony
addWorkers(colony, nInd = 20)

# Using a default in SP$nWorkers
# (just to have some workers - change this to your needs!)
addWorkers(colony)
addWorkers(colony)

SP$nWorkers <- 15
addWorkers(colony)
addWorkers(colony)

nWorkersFun <- function(colony) { rpois(n = 1, lambda = 15) }
addWorkers(colony, nInd = nWorkersFun)
addWorkers(colony, nInd = nWorkersFun)

SP$nWorkers <- nWorkersFun
addWorkers(colony)
addWorkers(colony)

}
