% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{areVirginQueensPresent}
\alias{areVirginQueensPresent}
\title{Are virgin queen(s) present}
\usage{
areVirginQueensPresent(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
logical, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns virgin queen(s) presence status.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createDrones(x = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)
areVirginQueensPresent(colony1)
areVirginQueensPresent(colony2)
areVirginQueensPresent(apiary)

colony1 <- addVirginQueens(colony1)
colony2 <- buildUpColony(colony2)
apiary <- c(colony1, colony2)

areVirginQueensPresent(colony1)
areVirginQueensPresent(colony2)
areVirginQueensPresent(apiary)

colony1r <- removeQueen(colony1)
areVirginQueensPresent(colony1r)

colony2s <- supersedeColony(colony2)
areVirginQueensPresent(colony2s)

tmp <- swarmColony(colony1)
areVirginQueensPresent(tmp$swarm)
areVirginQueensPresent(tmp$remnant)

tmp <- splitColony(colony1)
areVirginQueensPresent(tmp$split)
areVirginQueensPresent(tmp$remnant)

}
