% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{assignColonyToColonies}
\alias{assignColonyToColonies}
\title{Assign (replace) a colony to colonies}
\usage{
assignColonyToColonies(colonies, colony, pos)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{colony}{\code{\link{Colony-class}}, colony that will be added}

\item{pos}{numeric or character, index or ID of the old colony}
}
\value{
\code{\link{Colonies-class}} with a replaced colony
}
\description{
Level 3 function that assigns (replaces) a colony among a set of
  colonies, for example, to replace an old colony. By defining a position, it
  will insert the colony to that position and with this replace the old
  colony at that position.
}
\examples{
founderGenomes <- quickHaplo(nInd = 6, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

founderDrones <- createFounderDrones(pop = basePop[1:3], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[4], fathers = founderDrones[1:10])
colony2 <- createColony(queen = basePop[5], fathers = founderDrones[11:20])
colony3 <- createColony(queen = basePop[6], fathers = founderDrones[21:30])
apiary <- c(colony1, colony2)

getId(apiary)
apiary <- assignColonyToColonies(apiary, colony3, pos = 1)
getId(apiary)

}
