% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{asVirginQueen}
\alias{asVirginQueen}
\title{Converts individuals into virgin queens}
\usage{
asVirginQueen(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}}
}
\value{
\code{\link{Pop-class}} with sex set to \code{"F"} and
  \code{@misc[[*]]$caste} set to \code{"V"}
}
\description{
Level 1 function that takes individuals from a population and
  makes them virgin queens. These individuals are diploid.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

isVirginQueen(basePop)

basePop <- asVirginQueen(basePop)
isVirginQueen(basePop)

}
