% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{calcBeeGRMIbd}
\alias{calcBeeGRMIbd}
\title{Calculate Genomic Relatedness Matrix (GRM) for honey bees from
  Identical By Descent genomic data}
\usage{
calcBeeGRMIbd(x)
}
\arguments{
\item{x}{\code{\link{matrix}} of haplotypes representing allele indicators
of the founders as 1, 2, ... Individuals are in rows and sites are in
columns; no missing values are allowed (this is not checked!)}

\item{id}{character vector indicating to which individual the ????}
}
\value{
matrix of genomic relatedness coefficients
}
\description{
Level 0 function that returns Genomic Relatedness Matrix (GRM)
  for honey bees from Identical By Descent genomic data (tracked alleles
  since the founders)
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(isTrackRec = TRUE)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony <- addWorkers(colony, nInd = 5)
colony <- addDrones(colony, nInd = 5)
colony <- addVirginQueens(colony, nInd = 2)

genoQ <- getQueensIbdHaplo(colony)
genoF <- getFathersIbdHaplo(colony)
genoW <- getWorkersIbdHaplo(colony)
genoD <- getDronesIbdHaplo(colony)
genoV <- getVirginQueensIbdHaplo(colony)
SP$pedigree

geno <- rbind(genoQ, genoF, genoW, genoD, genoV)

sex <- c(getQueen(colony)@sex,
         getFathers(colony)@sex,
         getWorkers(colony)@sex,
         getDrones(colony)@sex,
         getVirginQueens(colony)@sex,
         "F",
         "M")

GRM <- calcBeeGRMIbd(x = geno)

# library("Matrix"); image(as(GRM, "Matrix"))

x <- diag(GRM)
hist(x)
summary(x)

x <- GRM[lower.tri(x = GRM, diag = FALSE)]
hist(x)
summary(x)

q <- rownames(genoQ)
f <- rownames(genoF)
w <- rownames(genoW)
d <- rownames(genoD)
v <- rownames(genoV)

# Queen vs others
GRM[q, q]
GRM[f, q]
GRM[w, q]
GRM[d, q]
GRM[v, q]

# Fathers vs others
GRM[f, f]
GRM[w, f]
GRM[d, f]
GRM[v, f]

# Workers vs others
GRM[w, w]
GRM[d, w]
GRM[v, w]

}
