% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{calcBeeGRMIbs}
\alias{calcBeeGRMIbs}
\title{Calculate Genomic Relatedness Matrix (GRM) for honey bees from
  Identical By State genomic data}
\usage{
calcBeeGRMIbs(x, sex)
}
\arguments{
\item{x}{\code{\link{matrix}} of genotypes represented as allele dosage coded
as 0, 1, or 2 in females (queens or workers) and as 0 or 1 in males
(drones); individuals are in rows and sites are in columns; no missing
values are allowed (this is not checked!)}

\item{sex}{character vector denoting sex for individuals with genotypes in
\code{x} - \code{"F"} for female and \code{"M"} for male}
}
\value{
matrix of genomic relatedness coefficients
}
\description{
Level 0 function that returns Genomic Relatedness Matrix (GRM)
  for honey bees from Identical By State genomic data (bi-allelic SNP
  represented as allele dosages) following the method of Druet and Legarra
  (2020)
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony1 <- addWorkers(colony1, nInd = 10)
colony1 <- addDrones(colony1, nInd = 2)
colony1 <- addVirginQueens(colony1, nInd = 2)

genoQ <- getQueensSegSiteGeno(colony1)
genoF <- getFathersSegSiteGeno(colony1)
genoW <- getWorkersSegSiteGeno(colony1)
genoD <- getDronesSegSiteGeno(colony1)
genoV <- getVirginQueensSegSiteGeno(colony1)
genoM <- apply(X = genoW, MARGIN = 2, FUN = mean)
geno <- rbind(genoQ, genoF, genoW, genoD, genoV, genoM)
rownames(geno)[length(rownames(geno))] <- "M"
sex <- c(
  "F",
  rep("M", times = nrow(genoF)),
  rep("F", times = nrow(genoW)),
  rep("M", times = nrow(genoD)),
  rep("F", times = nrow(genoV)),
  "F"
)

GRM <- calcBeeGRMIbs(x = geno, sex = sex)

if (require("Matrix")) {
  image(as(GRM, "Matrix"),
    xlab = "Individual", ylab = "Individual"
  )
}

x <- diag(GRM)
hist(x)
summary(x)

x <- GRM[lower.tri(x = GRM, diag = FALSE)]
hist(x)
summary(x)

q <- rownames(genoQ)
f <- rownames(genoF)
w <- rownames(genoW)
d <- rownames(genoD)
v <- rownames(genoV)
m <- "M"

# Queen vs others
GRM[q, f]
GRM[q, w]
GRM[q, d]
GRM[q, v]
GRM[q, m]

# Fathers vs others
GRM[f, f]
GRM[f, w]
GRM[f, d]
GRM[f, v]
GRM[f, m]

# Workers vs others
GRM[w, w]
GRM[w, d]
GRM[w, v]
GRM[w, m]
}
\references{
Druet and Legarra (2020) Theoretical and empirical comparisons of
  expected and realized relationships for the X-chromosome. Genetics
  Selection Evolution, 52:50 \url{https://doi.org/10.1186/s12711-020-00570-6}
}
