% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SimParamBee.R
\name{calcColonyPhenoFromCaste}
\alias{calcColonyPhenoFromCaste}
\title{Sample colony phenotype based on caste phenotypes}
\usage{
calcColonyPhenoFromCaste(
  colony,
  queenTrait = 1,
  queenFUN = function(x) x,
  workersTrait = 2,
  workersFUN = sum,
  dronesTrait = NULL,
  dronesFUN = NULL,
  combineFUN = function(q, w, d) q + w,
  checkProduction = TRUE,
  ...
)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{queenTrait}{numeric, trait that represents queen's effect on the colony
phenotype (defined in \code{\link{SimParamBee}} - see examples); if
\code{NULL} then this effect is 0}

\item{queenFUN}{function, function that will be applied to the queen phenotype}

\item{workersTrait}{numeric, trait that represents workers' effect on the colony
phenotype (defined in \code{\link{SimParamBee}} - see examples); if
\code{NULL} then this effect is 0}

\item{workersFUN}{function, function that will be applied to the worker phenotypes}

\item{dronesTrait}{numeric, trait that represents drones' effect on the colony
phenotype (defined in \code{\link{SimParamBee}} - see examples); if
\code{NULL} then this effect is 0}

\item{dronesFUN}{function, function that will be applied to the drone phenotypes}

\item{combineFUN, }{function that will combine the queen, worker, and drone effects
- the function expects that this function is defined as \code{function(q, w, d)}
where \code{q} represents queen's effect, \code{q} represents workers' effect,
\code{d} represents drones' effect}

\item{checkProduction}{logical, does the phenotype depend on the production
status of colony; if yes and production is not \code{TRUE}, the result is
a 0}

\item{...}{arguments passed to \code{queenFUN}, \code{workersFUN}, or \code{combineFUN}}
}
\value{
numeric matrix with a single value
}
\description{
Sample colony phenotype based on caste phenotypes -
  used when \code{FUN = NULL} in \code{\link{setColonyPheno}} but
  \code{SimParamBee$colonyPheno} is set.

  This is just an example - quite flexible one, though;) You can provide your
  own functions that satisfy your needs!
}
\seealso{
\code{\link{SimParamBee}} field \code{colonyPheno}, and functions
  \code{\link{getEvents}} and \code{\link{setColonyPheno}} (and its example!)
}
