% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{collapse}
\alias{collapse}
\title{Collapse}
\usage{
collapse(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}
}
\value{
\code{\link{Colony-class}} or \code{\link{MultiColony-class}} with the collapse
  event set to \code{TRUE}
}
\description{
Level 2 function that collapses a Colony or MultiColony object
  by setting the collapse event slot to \code{TRUE}. The production status
  slot is also changed (to \code{FALSE}).
}
\examples{
founderGenomes <- quickHaplo(nInd = 10, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)
drones <- createDrones(basePop[1], n = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = 10)

# Create Colony and MultiColony class
colony <- createColony(x = basePop[1])
colony <- cross(colony, fathers = fatherGroups[[1]])
apiary <- createMultiColony(x = basePop[2:10], n = 9)
apiary <- cross(apiary, fathers = fatherGroups[2:10])

# Collapse
hasCollapsed(colony)
colony <- collapse(colony)
hasCollapsed(colony)

hasCollapsed(apiary)
tmp <- pullColonies(apiary, n = 2)
tmp
apiaryLost <- collapse(tmp$pulledColonies)
hasCollapsed(apiaryLost)
apiaryLeft <- tmp$remainingColonies
hasCollapsed(apiaryLeft)

}
