% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{collapseColony}
\alias{collapseColony}
\title{Collapse colony}
\usage{
collapseColony(colony)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}
}
\value{
\code{\link{Colony-class}} with the collapse event set to \code{TRUE}
}
\description{
Level 2 function that collapses colony by setting the collapse
  event slot to \code{TRUE}. The production status slot is also changed (to
  \code{FALSE}).
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
(colony <- buildUpColony(colony, nWorkers = 100))
collapseColony(colony)

}
