% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Colonies.R
\docType{class}
\name{Colonies-class}
\alias{Colonies-class}
\title{Honeybee colonies}
\arguments{
\item{x}{\code{\link{Colonies-class}}}

\item{i}{numeric or character, index or name to select a colony}
}
\value{
\code{\link{Colonies-class}}
}
\description{
An object holding a collection of honeybee colonies. It behaves
  like a list.
}
\section{Slots}{

\describe{
\item{\code{colonies}}{list, a collection of \code{\link{Colony-class}} objects}
}}

\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 15)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony3 <- createColony(queen = basePop[4], fathers = drones[11:15])
apiary <- c(colony1, colony2, colony3)
apiary
getId(apiary)
apiary[1]
getId(apiary[1])
getId(apiary["2"])
getId(apiary[2])

getId(apiary)
getId(apiary[c(1, 3)])
getId(apiary[c("2", "4")])

apiary[[1]]
apiary[["2"]]
apiary[[3]]
apiary[["4"]]

getId(c(apiary[c(1, 3)], apiary[2]))

}
\seealso{
\code{\link{createColonies}}
}
