% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Colony.R
\docType{class}
\name{PopOrNULL-class}
\alias{PopOrNULL-class}
\title{Honeybee colony}
\value{
\code{\link{Colony-class}}
}
\description{
An object holding honeybee colony
}
\section{Slots}{

\describe{
\item{\code{id}}{character, ID of the colony, which is equal to the ID of the queen}

\item{\code{location}}{numeric, location of the colony (x, y)}

\item{\code{queen}}{\code{\link{Pop-class}}, the queen of the colony (we use
its misc slot for queen's age and drones (fathers) she mated with)}

\item{\code{virgin_queens}}{\code{\link{Pop-class}}, virgin queens of the
colony}

\item{\code{drones}}{\code{\link{Pop-class}}, drones of the colony}

\item{\code{workers}}{\code{\link{Pop-class}}, workers of the colony}

\item{\code{nHomDrones}}{integer, the number of homozygous drones in the colony}

\item{\code{pheno}}{matrix, phenotypes of the colony TODO: revise}

\item{\code{split}}{logical, has colony split}

\item{\code{swarm}}{logical, has colony swarmed}

\item{\code{supersedure}}{logical, has colony superseded}

\item{\code{collapse}}{logical, has colony collapsed TODO: revise}

\item{\code{production}}{logical, is colony productive}

\item{\code{last_event}}{character, the last event of the colony TODO: we probably don't need this}

\item{\code{misc}}{list, available for storing extra information about the colony}
}}

\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])

colony1
colony2
is(colony1)
is(colony2)

apiary <- c(colony1, colony2)
is(apiary)

}
\seealso{
\code{\link{createColony}}
}
