% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{combine}
\alias{combine}
\title{Combine two colony objects}
\usage{
combine(strong, weak)
}
\arguments{
\item{strong}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{weak}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
a combined \code{\link{Colony-class}} or \code{\link{Colonies-class}}
}
\description{
Level 2 function that combines two colony objects into one or
  two colonies objects of the same length to one. For example, to combine a
  weak and a strong colony (or colonies). Workers and drones of the weak
  colony are added to the strong. User has to remove the weak colony (or
  colonies) from the workspace.
}
\examples{
founderGenomes <- quickHaplo(nInd = 10, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 70)
col1 <- createColony(x = basePop[2])
col1 <- crossColony(col1, drones = drones[1:10], nFathers = 5)
col2 <- createColony(x = basePop[3])
col2 <- crossColony(col2, drones = drones[11:20], nFathers = 5)
col1 <- buildUpColony(colony = col1, nWorkers = 100, nDrones = 10)
col1
col2 <- buildUpColony(colony = col2, nWorkers =  20, nDrones =  2)
col2
col1 <- combine(strong = col1, weak = col2)
col1
rm(col2)

drones <- createDrones(x = basePop[1], nInd = 70)
col1 <- createColony(x = basePop[4])
col1 <- crossColony(col1, drones = drones[1:10], nFathers = 5)
col2 <- createColony(x = basePop[5])
col2 <- crossColony(col2, drones = drones[11:20], nFathers = 5)
col3 <- createColony(x = basePop[6])
col3 <- crossColony(col3, drones = drones[21:30], nFathers = 5)
col4 <- createColony(x = basePop[7])
col4 <- crossColony(col4, drones = drones[31:40], nFathers = 5)
col1 <- buildUpColony(colony = col1, nWorkers = 100, nDrones = 10)
col2 <- buildUpColony(colony = col2, nWorkers = 20, nDrones = 2)
col3 <- buildUpColony(colony = col3, nWorkers = 100, nDrones = 10)
col4 <- buildUpColony(colony = col4, nWorkers = 20, nDrones = 2)
colsStrong <- c(col1, col3)
colsStrong[[1]]
colsStrong[[2]]
colsWeak <- c(col2, col4)
cols <- combine(strong = colsStrong, weak = colsWeak)
cols[[1]]
cols[[2]]
rm(colsWeak)
}
