% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{combineColony}
\alias{combineColony}
\title{Combine two colony objects into one}
\usage{
combineColony(strong, weak)
}
\arguments{
\item{strong}{\code{\link{Colony-class}}}

\item{weak}{\code{\link{Colony-class}}}
}
\value{
a colony \code{\link{Colony-class}} that combines workers and drones
  of both colony.
}
\description{
Level 2 function that combines two colony objects into one.
  For example, to combine a weak and a strong colony. Workers and drones
  of the weak colony are added to the strong. User has to remove the weak
.  colony.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 30)
col1 <- createColony(queen = basePop[2], fathers = drones[1:15] )
col2 <- createColony(queen = basePop[3], fathers = drones[16:30])
(col1 <- buildUpColony(col1, nWorkers = 100, nDrones = 10))
(col2 <- buildUpColony(col2, nWorkers =  20, nDrones =  2))

(col2 <- combineColony(strong = col1, weak = col2))
rm(col1)
}
