% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{computeQueensPHomBrood}
\alias{computeQueensPHomBrood}
\title{Theoretical percentage of homozygous brood of a queen}
\usage{
computeQueensPHomBrood(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}
}
\value{
numeric, theoretical homozygosity named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns the theoretical or
expected percentage of homozygous brood per queen. The percentage
is computed based on the csd allele of the queen and the drones
the queen mated with (the fathers).
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 10)
colony2 <- addDrones(colony2, nInd = 20)

# Mated queen
computeQueensPHomBrood(colony1@queen)

# Colony
computeQueensPHomBrood(colony1)

# Colonies
apiary <- c(colony1, colony2)
computeQueensPHomBrood(apiary)
}
