% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{computeTheoreticalpHomBrood}
\alias{computeTheoreticalpHomBrood}
\title{Empirical percentage of homozygous brood of a queen}
\usage{
computeTheoreticalpHomBrood(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}
}
\value{
numeric, empirical homozygosity named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns the empirical or
expected percentage of homozygous brood per queen. The percentage
is computed based on the csd allele of the queen and the drones
the queen mated with (the fathers.)
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 10)
colony2 <- addDrones(colony2, nInd = 20)

# Mated queen
computeTheoreticalpHomBrood(colony1@queen)

# Colony
computeTheoreticalpHomBrood(colony1)

# Colonies
apiary <- c(colony1, colony2)
computeTheoreticalpHomBrood(apiary)
}
