% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{createColonies}
\alias{createColonies}
\title{Create colonies}
\usage{
createColonies(
  x = NULL,
  n = NULL,
  yearOfBirth = NULL,
  location = NULL,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, virgin queens or queens for the colonies
(selected at random if there are more than \code{n} in \code{Pop});
\code{\link{isVirginQueen}} and \code{\link{isQueen}} test will be run on
individuals}

\item{n}{integer, number of colonies to create (if only \code{nCol} is
given then \code{\link{Colonies-class}} is created with \code{nCol} empty
(\code{NULL}) individual colony) - this is mostly useful for programming)}

\item{yearOfBirth}{numeric, year of birth of the queens}

\item{location}{numeric, location of the colonies as \code{c(x, y)}}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colonies-class}}
}
\description{
Level 3 function that creates a set of colonies. Usually to
  start a simulation.
}
\details{
When both \code{pop} and \code{nCol} are \code{NULL}, then an empty
  \code{NULL} \code{\link{Colonies-class}} is created with 0 colonies.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

# Create 2 empty (NULL) colonies
apiary <- createColonies(n = 2)
apiary
apiary[[1]]
apiary[[2]]

# Create 2 mated colonies
apiary <- createColonies(pop = basePop, n = 2)
apiary
apiary[[1]]
apiary[[2]]

# Create 2 unmated/virgin colonies
apiary <- createColonies(x = basePop[1:2], n = 2)
drones <- createDrones(x = basePop[3], n = 50)
apiary <- crossColonies(apiary, drones = drones, nFathers = 15)
apiary
apiary[[1]]
apiary[[2]]
}
