% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{createColony}
\alias{createColony}
\title{Create a new Colony}
\usage{
createColony(x, location = NULL)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, one queen or virgin queen(s)}

\item{location}{numeric, location of the colony as \code{c(x, y)}}
}
\value{
new \code{\link{Colony-class}}
}
\description{
Level 2 function that creates a new \code{\link{Colony-class}}
  to initiate simulations.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 50)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones, nFathers = 15)
colony1

colony2 <- createColony(x = basePop[3:4])
colony2
}
