% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createFounderDrones}
\alias{createFounderDrones}
\title{Creates drones from a founding (base) population}
\usage{
createFounderDrones(pop, nDronesPerQueen = 100)
}
\arguments{
\item{pop}{\code{\link{Pop-class}}}

\item{nDronesPerQueen}{integer, number of drones to create per founding
individual (a substitute for a queen).}
}
\value{
\code{\link{Pop-class}} with drones
}
\description{
Level 1 function that creates the specified number of drones
  from a founding (base) population. Drones are haploid and created from the
  diploid genome of individuals in the population. Individuals' ID is stored
  as the father and mother of drones.

Level 1 function that creates drones from a founding (base)
  population by doubling recombined genomes of each founding individual -
  mimicking a queen laying drones. Such drones are usually used as fathers
  when creating colonies.
}
\details{
Note that this function creates \code{nDronesPerQueen} for each
  individual in the \code{pop}, which will amount to
  \code{nInd(pop) * nDronesPerQueen} drones - this can be slow if either or
  both of \code{c(nInd(pop), nDronesPerQueen)} is large; tune the numbers
  to your needs.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
# TODO: expand exapmles by showing how to use SP

}
