% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{createMultipleVirginColonies}
\alias{createMultipleVirginColonies}
\title{Create a list object of class "colonies" containing only unmated virgin queens}
\usage{
\method{createMultipleVirginColonies}(colonies, ID)
}
\arguments{
\item{founderPop}{The initial founder population}

\item{nColonies}{Number of colonies the use wants to create}
}
\value{
A AlphaSimRBee Colonies object
}
\description{
The function is intended for creating initial colonies from
'FOUNDERPOP'. The user can create a list containing their desired number of colonies (nColonies).
All colonies created are populated with un-mated virgin queens, no other members of the caste are
present and must be populated in further steps.
}
\examples{
#' #Create founder haplotypes
founderPop <- quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP <- SimParam$new(founderPop)

#Create population
base <- newPop(founderPop, simParam=SP)

#Create 10 virgin queen colonies
 apiary1 <- createMultipleVirginColonies(founderPop = base, nColonies = 10)

}
