% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{crossColony}
\alias{crossColony}
\title{Crosses a colony with a virgin queen to a group of fathers pulled from the DCA.}
\usage{
\method{crossColony}(colony, fathers, nWorkers, nDrones)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object with a non-mated virgin queen}

\item{fathers}{Pop Class. Father group pulled from the DCA.}

\item{nWorkers}{Integer.Number of workers to create}

\item{nDrones}{Integer. Number of drones to create}
}
\value{
Single AlphaSim population object of a mated colony
}
\description{
Crosses a colony with a virgin queen to a group of fathers pulled from the DCA
\creates workers, drones and a new virgin queen and write them to the corresponding
\slots of the colony object.
#IF the colony is queenless - select a queen from the virgin queen - if not, mate the current queen!!!
}
\examples{
inst/examples/examples_crossColony.R
}
\seealso{
\code{\link[??????]{createColony}}
}
