% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{downsizeColony}
\alias{downsizeColony}
\title{Reduce number of workers and remove all drones and virgin queens from hive}
\usage{
downsizeColony(colony, p = NULL, use = "rand", new = FALSE, simParamBee = NULL)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{p}{numeric, proportion of workers to be removed from the colony; if
\code{NULL} then \code{\link{SimParamBee}$downsizeP} is used}

\item{use}{character, all the options provided by \code{\link{selectInd}};
it guides the selection of workers that will be removed}

\item{new}{logical, should we remove all current workers and add a targeted
proportion anew (say, create winter workers)}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colony-class}} with workers reduced and drones/virgin queens removed
}
\description{
Level 2 function that downsizes colony by removing a proportion
  of workers, all drones and all virgin queens. Usually in the autumn, such
  an event occurs in preparation for the winter months.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)
drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones, nFathers = 5)
colony <- buildUpColony(colony)
colony <- addVirginQueens(x = colony, nInd = 10)
colony
colony <- downsizeColony(colony = colony, new = TRUE, use = "rand")
colony
}
