% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{editGenome_OneSite_Diploid_Pop}
\alias{editGenome_OneSite_Diploid_Pop}
\title{Edit genome at a single diploid site in a population}
\usage{
editGenome_OneSite_Diploid_Pop(pop, chr, segSite, alleles, simParamBee = NULL)
}
\arguments{
\item{pop}{an object of \code{\link{Pop-class}}}

\item{chr}{integer, the chromosomes to edit.}

\item{segSite}{integer, the position of the segregating site to edit}

\item{alleles}{list, each element of the list contains a vector with twe two
desired alleles for each individual at the segregating sites. The length of the list
should match the number of individuals in the population}

\item{simParamBee}{an object of \code{\link{SimParamBee}}}
}
\value{
Returns an object of \code{\link{Pop-class}}
}
\description{
Edits a single selected diploid locus in an entire population
of individuals to the desired diploid genotype. The gv slot is recalculated to
reflect the any changes due to editing, but other slots remain the same.
}
\examples{
#Create founder haplotypes
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 5)
SP <- SimParamBee$new(founderGenomes, csdChr = NULL)
pop <- createVirginQueens(founderGenomes)

#Change individual 1 to homozygous for the 1 allele
#at locus 1, chromosome 1
pop2 <- SIMplyBee:::editGenome_OneSite_Diploid_Pop(pop,
                                                   chr = 1, segSite = 1,
                                                   alleles = list(c(1, 1), c(0, 1)),
                                                   simParam = SP)
getSegSiteHaplo(pop)
getSegSiteHaplo(pop2)
}
