% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteId}
\alias{getCasteId}
\title{Get IDs of individuals of a caste, or ID of all members of colony}
\usage{
getCasteId(x, caste = "all")
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}

\item{caste}{character, "queen", "fathers", "virginQueens", "workers",
"drones", or "all"}
}
\value{
When x is \code{\link{Pop-class}}, character of individual ID. When x
  is \code{\link{Colony-class}}, list with character vectors (list is named
  with caste). When x is \code{\link{Colonies-class}}, list of lists with
  character vectors (list is named with colony id).
}
\description{
Level 0 function that returns the ID individuals of a caste. To
  get the individuals, use \code{\link{getCaste}}.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])

colony1 <- addWorkers(colony1, nInd = 10)
colony1 <- addVirginQueens(colony1, nInd = 4)
colony1 <- addDrones(colony1, nInd = 2)
colony2 <- addWorkers(colony2, nInd = 20)

apiary = c(colony1, colony2)

getCasteId(x = drones)
getCasteId(x = colony1)
getCasteId(x = colony1, caste = "workers")
getCasteId(x = apiary)
getCasteId(x = apiary, caste = "virginQueens")

# Create a data.frame with id, colony, and caste information
(tmpC <- caste(colony1))
(tmpI <- getCasteId(colony1))
tmp <- data.frame(caste = unlist(tmpC), id = unlist(tmpI))
head(tmp);tail(tmp)

(tmpC <- caste(apiary))
(tmpI <- getCasteId(apiary))
(tmp <- data.frame(caste = unlist(tmpC), id = unlist(tmpI)))
tmp$colony <- sapply(X = strsplit(x = rownames(tmp), split = ".",
                                  fixed = TRUE),
                     FUN = function(z) z[[1]])
head(tmp);tail(tmp)

}
\seealso{
\code{\link{getCaste}}

\code{\link{getCaste}} and \code{\link{caste}}
}
