% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteQtlHaplo}
\alias{getCasteQtlHaplo}
\alias{getQueensQtlHaplo}
\alias{getFathersQtlHaplo}
\alias{getVirginQueensQtlHaplo}
\alias{getWorkersQtlHaplo}
\alias{getDronesQtlHaplo}
\title{Access QTL haplotypes of individuals in a caste}
\usage{
getCasteQtlHaplo(
  x,
  caste,
  nInd = NULL,
  trait = 1,
  haplo = "all",
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getQueensQtlHaplo(x, trait = 1, haplo = "all", chr = NULL, simParamBee = NULL)

getFathersQtlHaplo(
  x,
  nInd = NULL,
  trait = 1,
  haplo = "all",
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getVirginQueensQtlHaplo(
  x,
  nInd = NULL,
  trait = 1,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)

getWorkersQtlHaplo(
  x,
  nInd = NULL,
  trait = 1,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)

getDronesQtlHaplo(
  x,
  nInd = NULL,
  trait = 1,
  haplo = "all",
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{caste}{character, "queen", "fathers", "virginQueens", "workers", or
"drones"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample}

\item{trait}{numeric, indicates which trait's QTL haplotypes to retrieve}

\item{haplo}{character, either "all" for all haplotypes or an integer for a
single set of haplotypes, use a value of 1 for female haplotypes and a
value of 2 for male haplotypes}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{dronesHaploid}{logical, return haploid result for drones?}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
matrix with haplotypes when \code{x} is \code{\link{Colony-class}}
  and list of matrices with haplotypes when \code{x} is
  \code{\link{Colonies-class}}, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns QTL haplotypes of individuals in a
  caste.
}
\section{Functions}{
\itemize{
\item \code{getQueensQtlHaplo}: Access QTL haplotype data of the queen

\item \code{getFathersQtlHaplo}: Access QTL haplotype data of fathers

\item \code{getVirginQueensQtlHaplo}: Access QTL haplotype data of virgin queens

\item \code{getWorkersQtlHaplo}: Access QTL haplotype of workers

\item \code{getDronesQtlHaplo}: Access QTL haplotype data of drones
}}

\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$addTraitAD(nQtlPerChr = 10, meanDD = 0.2, varDD = 0.1)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 2)
colony2 <- addDrones(colony2, nInd = 4)
colony1 <- addVirginQueens(colony1, nInd = 2)

getCasteQtlHaplo(colony1, caste = "queen")
getQueensQtlHaplo(colony1)

getCasteQtlHaplo(colony1, caste = "fathers")
getCasteQtlHaplo(colony1, caste = "fathers", nInd = 2)
getCasteQtlHaplo(colony1, caste = "fathers", nInd = 2)
getFathersQtlHaplo(colony1)
getFathersQtlHaplo(colony1, nInd = 2)

getCasteQtlHaplo(colony1, caste = "virginQueens")
getVirginQueensQtlHaplo(colony1)

getCasteQtlHaplo(colony1, caste = "workers")
getWorkersQtlHaplo(colony1)

getCasteQtlHaplo(colony1, caste = "drones")
getDronesQtlHaplo(colony1)

apiary <- c(colony1, colony2)
getCasteQtlHaplo(apiary, caste = "queen")
getQueensQtlHaplo(apiary)

getCasteQtlHaplo(apiary, caste = "fathers")
getCasteQtlHaplo(apiary, caste = "fathers", nInd = 2)
getCasteQtlHaplo(apiary, caste = "fathers", nInd = 2)
getFathersQtlHaplo(apiary)
getFathersQtlHaplo(apiary, nInd = 2)

getCasteQtlHaplo(apiary, caste = "virginQueens")
getVirginQueensQtlHaplo(apiary)

getCasteQtlHaplo(apiary, caste = "workers")
getWorkersQtlHaplo(apiary)

getCasteQtlHaplo(apiary, caste = "drones")
getDronesQtlHaplo(apiary)

}
\seealso{
\code{\link{getQtlHaplo}} and \code{\link{pullQtlHaplo}}
}
