% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteSnpHaplo}
\alias{getCasteSnpHaplo}
\alias{getQueenSnpHaplo}
\alias{getFathersSnpHaplo}
\alias{getVirginQueensSnpHaplo}
\alias{getWorkersSnpHaplo}
\alias{getDronesSnpHaplo}
\title{Access SNP array haplotypes of individuals in a caste}
\usage{
getCasteSnpHaplo(
  x,
  caste,
  nInd = NULL,
  snpChip = 1,
  haplo = "all",
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getQueenSnpHaplo(x, snpChip = 1, haplo = "all", chr = NULL, simParamBee = NULL)

getFathersSnpHaplo(
  x,
  nInd = NULL,
  snpChip = 1,
  haplo = "all",
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getVirginQueensSnpHaplo(
  x,
  nInd = NULL,
  snpChip = 1,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)

getWorkersSnpHaplo(
  x,
  nInd = NULL,
  snpChip = 1,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)

getDronesSnpHaplo(
  x,
  nInd = NULL,
  snpChip = 1,
  haplo = "all",
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}

\item{caste}{character, "queen", "fathers", "workers", "drones", or
"virginQueens"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample}

\item{snpChip}{numeric, indicates which SNP array haplotypes to retrieve}

\item{haplo}{character, either "all" for all haplotypes or an integer for a
single set of haplotypes, use a value of 1 for female haplotypes and a
value of 2 for male haplotypes}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{dronesHaploid}{logical, return haploid result for drones?}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
matrix with haplotypes when \code{x} is \code{\link{Colony-class}}
  and list of matrices with haplotypes when \code{x} is
  \code{\link{MultiColony-class}}, named by colony id when \code{x} is
  \code{\link{MultiColony-class}}
}
\description{
Level 0 function that returns SNP array haplotypes of
  individuals in a caste.
}
\section{Functions}{
\itemize{
\item \code{getQueenSnpHaplo()}: Access SNP array haplotype data of the queen

\item \code{getFathersSnpHaplo()}: Access SNP array haplotype data of fathers

\item \code{getVirginQueensSnpHaplo()}: Access SNP array haplotype data of virgin queens

\item \code{getWorkersSnpHaplo()}: Access SNP array haplotype of workers

\item \code{getDronesSnpHaplo()}: Access SNP array haplotype data of drones

}}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(TRUE)
SP$setTrackPed(isTrackPed = TRUE)
SP$addTraitA(10)
SP$addSnpChip(5)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(x = colony)
colony <- addVirginQueens(x = colony, nInd = 5)

apiary <- createMultiColony(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])
apiary <- buildUp(x = apiary)
apiary <- addVirginQueens(x = apiary, nInd = 5)

getCasteSnpHaplo(colony, caste = "queen")
getQueenSnpHaplo(colony)

getCasteSnpHaplo(colony, caste = "fathers")
getCasteSnpHaplo(colony, caste = "fathers", nInd = 2)
getCasteSnpHaplo(colony, caste = "fathers", nInd = 2)
getFathersSnpHaplo(colony)
getFathersSnpHaplo(colony, nInd = 2)

getCasteSnpHaplo(colony, caste = "virginQueens")
getVirginQueensSnpHaplo(colony)

getCasteSnpHaplo(colony, caste = "workers")
getWorkersSnpHaplo(colony)

getCasteSnpHaplo(colony, caste = "drones")
getDronesSnpHaplo(colony)

getCasteSnpHaplo(apiary, caste = "queen")
getQueenSnpHaplo(apiary)

getCasteSnpHaplo(apiary, caste = "fathers")
getCasteSnpHaplo(apiary, caste = "fathers", nInd = 2)
getCasteSnpHaplo(apiary, caste = "fathers", nInd = 2)
getFathersSnpHaplo(apiary)
getFathersSnpHaplo(apiary, nInd = 2)

getCasteSnpHaplo(apiary, caste = "virginQueens")
getVirginQueensSnpHaplo(apiary)

getCasteSnpHaplo(apiary, caste = "workers")
getWorkersSnpHaplo(apiary)

getCasteSnpHaplo(apiary, caste = "drones")
getDronesSnpHaplo(apiary)
}
\seealso{
\code{\link{getSnpHaplo}} and \code{\link{pullSnpHaplo}}
}
