% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getColonyDd}
\alias{getColonyDd}
\title{Access dominance deviations of individuals in colony}
\usage{
getColonyDd(
  x,
  caste = c("queen", "fathers", "workers", "drones", "virginQueens"),
  nInd = NULL,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{caste}{character, a combination of "queen", "fathers", "workers",
"drones", or "virginQueens"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample; can be a list to
access different number of different caste - when this is the case
\code{nInd} takes precedence over \code{caste} (see examples)}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
list of vector of dominance deviations when \code{x} is
  \code{\link{Colony-class}} (list nodes named by caste) and list of a list
  of vectors of dominance deviations when \code{x} is
  \code{\link{Colonies-class}}, outer list is named by colony id when
  \code{x} is \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns dominance deviations of
  individuals in colony.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$addTraitAD(nQtlPerChr = 10, meanDD = 0.2, varDD = 0.1)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(x = colony)
colony <- addVirginQueens(x = colony, nInd = 5)

apiary <- createColonies(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])
apiary <- buildUp(x = apiary)
apiary <- addVirginQueens(x = apiary, nInd = 5)

getColonyDd(colony)
getColonyDd(colony, caste = c("queen", "fathers"))
getColonyDd(colony, nInd = 1)
getColonyDd(colony, nInd = list("queen" = 1, "fathers" = 2, "virginQueens" = 1))

getColonyDd(apiary)
getColonyDd(apiary, caste = c("queen", "fathers"))
getColonyDd(apiary, nInd = 1)
getColonyDd(apiary, nInd = list("queen" = 1, "fathers" = 2, "virginQueens" = 1))
}
\seealso{
\code{\link{dd}}
}
