% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getColonyGv}
\alias{getColonyGv}
\title{Access genetic values of individuals in colony}
\usage{
getColonyGv(
  x,
  caste = c("queen", "fathers", "virginQueens", "workers", "drones"),
  nInd = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or Colonies}

\item{caste}{character, a combination of "queen", "fathers", "virginQueens",
"workers", or "drones"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample; can be a list to
access different number of different caste - when this is the case
\code{nInd} takes precedence over \code{caste} (see examples)}
}
\value{
list of vector of genetic values when \code{x} is
  \code{\link{Colony-class}} (list nodes named by caste) and list of a list
  of vectors of genetic values when \code{x} is \code{\link{Colonies-class}},
  outer list is named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns genetic values of individuals in
  colony.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$addTraitAD(nQtlPerChr = 10, meanDD = 0.2, varDD = 0.1)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 2)

getColonyGv(colony1)
getColonyGv(colony1, caste = c("queen", "fathers"))
getColonyGv(colony1, nInd = 1)
getColonyGv(colony1, nInd = list("queen" = 1, "fathers" = 2, "virginQueens" = 1))

getColonyGv(colony2)

apiary <- c(colony1, colony2)
getColonyGv(apiary)
getColonyGv(apiary, caste = c("queen", "fathers"))
getColonyGv(apiary, nInd = 1)
getColonyGv(apiary, nInd = list("queen" = 1, "fathers" = 2, "virginQueens" = 1))
}
\seealso{
\code{\link{gv}}
}
