% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getColonySegSiteGeno}
\alias{getColonySegSiteGeno}
\title{Access genotypes for all segregating sites of individuals in
  colony}
\usage{
getColonySegSiteGeno(
  x,
  caste = c("queen", "fathers", "workers", "drones", "virginQueens"),
  nInd = NULL,
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}

\item{caste}{character, a combination of "queen", "fathers", "workers",
"drones", or "virginQueens"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample; can be a list to
access different number of different caste - when this is the case
\code{nInd} takes precedence over \code{caste} (see examples)}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{dronesHaploid}{logical, return haploid result for drones?}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
list of matrices with genotypes when \code{x} is
  \code{\link{Colony-class}} (list nodes named by caste) and list of a list
  of matrices with genotypes when \code{x} is \code{\link{MultiColony-class}},
  outer list is named by colony id when \code{x} is
  \code{\link{MultiColony-class}}
}
\description{
Level 0 function that returns genotypes for all segregating
  sites of individuals in colony.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(x = colony, nWorkers = 6, nDrones = 3)
colony <- addVirginQueens(x = colony, nInd = 5)

apiary <- createMultiColony(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])
apiary <- buildUp(x = apiary, nWorkers = 6, nDrones = 3)
apiary <- addVirginQueens(x = apiary, nInd = 5)

getColonySegSiteGeno(colony)
getColonySegSiteGeno(colony, caste = c("queen", "fathers"))
getColonySegSiteGeno(colony, nInd = 1)
getColonySegSiteGeno(colony, nInd = list("queen" = 1, "fathers" = 2, "virginQueens" = 1))

getColonySegSiteGeno(apiary)
getColonySegSiteGeno(apiary, caste = c("queen", "fathers"))
getColonySegSiteGeno(apiary, nInd = 1)
getColonySegSiteGeno(apiary, nInd = list("queen" = 1, "fathers" = 2, "virginQueens" = 1))
}
\seealso{
\code{\link{getCasteSegSiteHaplo}} and \code{\link{getSegSiteHaplo}}
}
