% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getColonySnpHaplo}
\alias{getColonySnpHaplo}
\title{Access SNP array haplotypes of individuals in colony}
\usage{
getColonySnpHaplo(
  x,
  caste = c("queen", "fathers", "workers", "drones", "virginQueens"),
  nInd = NULL,
  snpChip = 1,
  haplo = "all",
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}

\item{caste}{character, a combination of "queen", "fathers", "workers",
"drones", or "virginQueens"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample; can be a list to
access different number of different caste - when this is the case
\code{nInd} takes precedence over \code{caste} (see examples)}

\item{snpChip}{numeric, indicates which SNP array haplotypes to retrieve}

\item{haplo}{character, either "all" for all haplotypes or an integer for a
single set of haplotypes, use a value of 1 for female haplotypes and a
value of 2 for male haplotypes}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{dronesHaploid}{logical, return haploid result for drones?}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
list of matrices with haplotypes when \code{x} is
  \code{\link{Colony-class}} (list nodes named by caste) and list of a list
  of matrices with haplotypes when \code{x} is \code{\link{MultiColony-class}},
  outer list is named by colony id when \code{x} is
  \code{\link{MultiColony-class}}
}
\description{
Level 0 function that returns SNP array haplotypes of
  individuals in colony.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(TRUE)
SP$setTrackPed(isTrackPed = TRUE)
SP$addTraitA(10)
SP$addSnpChip(5)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(x = colony)
colony <- addVirginQueens(x = colony, nInd = 5)

apiary <- createMultiColony(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])
apiary <- buildUp(x = apiary)
apiary <- addVirginQueens(x = apiary, nInd = 5)

getColonySnpHaplo(colony)
getColonySnpHaplo(colony, caste = c("queen", "fathers"))
getColonySnpHaplo(colony, nInd = 1)
getColonySnpHaplo(colony, nInd = list("queen" = 1, "fathers" = 2, "virginQueens" = 1))

getColonySnpHaplo(apiary)
getColonySnpHaplo(apiary, caste = c("queen", "fathers"))
getColonySnpHaplo(apiary, nInd = 1)
getColonySnpHaplo(apiary, nInd = list("queen" = 1, "fathers" = 2, "virginQueens" = 1))
}
\seealso{
\code{\link{getCasteSnpHaplo}} and \code{\link{getSnpHaplo}}
}
