% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{getDrones}
\alias{getDrones}
\title{Access drones}
\usage{
getDrones(x, nInd = NULL)
}
\arguments{
\item{x}{Colony or Colonies}

\item{nInd}{numeric, number of drones to access, if NULL all drones
are accessed, otherwise a random sample}
}
\value{
When \code{x} is Colony then return is Pop, population object with drones
When \code{x} is Colonies then return is a list of Pop, population objects with drones
}
\description{
Access drones
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybee
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addDrones(colony1, nInd = 10)
colony2 <- addDrones(colony2, nInd = 20)
getDrones(colony1)
getDrones(colony1)@id
getDrones(colony1, nInd = 2)@id
getDrones(colony1, nInd = 2)@id

apiary <- c(colony1, colony2)
getDrones(apiary)
getDrones(apiary)[[1]]@id
getDrones(apiary)[[2]]@id

getDrones(apiary, nInd = 10)
getDrones(apiary, nInd = 10)[[1]]@id
getDrones(apiary, nInd = 10)[[2]]@id

}
