% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getEvents}
\alias{getEvents}
\title{Report which colony events have occurred}
\usage{
getEvents(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
matrix of logicals, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns a matrix of logicals reporting the
  status of the colony events. The events are: split, swarm, supersedure,
  collapse, and production. These events impact colony status, strength, and
  could also impact downstream phenotypes.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])

colony1 <- addWorkers(colony1, nInd = 10)
colony1 <- addVirginQueens(colony1, nInd = 4)
colony1 <- addDrones(colony1, nInd = 2)
getEvents(colony1)

colony1 <- supersedeColony(colony1)
getEvents(colony1)

colony2 <- addWorkers(colony2, nInd = 20)
apiary <- c(colony1, colony2)
getEvents(apiary)

}
