% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getnHomDrones}
\alias{getnHomDrones}
\title{Extract the number of homozygous drones in the colony}
\usage{
getnHomDrones(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
integer, named by colony id when \code{x} is \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns the number of homozygous
drones stored in the nHomDrones slot of the colony. Takes either Colony or Colonies
class as the input.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony
isProductive(colony)
colony <- buildUpColony(colony, nWorkers = 100)
colony
isProductive(colony)

colony1 <- createColony(queen = basePop[1], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[2], fathers = drones[6:10])
apiary <- c(colony1, colony2)
apiary <- buildUpColonies(apiary, nWorkers = 100)
getNHomDrones(apiary)

}
