% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getPoolGeno}
\alias{getPoolGeno}
\title{Get a pool genotype from true genotypes}
\usage{
getPoolGeno(x, type = NULL, sex = NULL)
}
\arguments{
\item{x}{matrix, true genotypes with individuals in rows and sites in columns}

\item{type}{character, "mean" for average genotype or "count" for the counts
of reference and alternative alleles}

\item{sex}{character, vector of "F" and "M" to denote the sex of individuals
in \code{x}}
}
\value{
a numeric vector with average allele dosage when \code{type = "mean"}
  and a two-row matrix with the counts of reference (1st row) and
  alternative (2nd row) alleles
}
\description{
Level 0 function that returns a pool genotype from true
  genotypes to mimic pool genotyping of colony members.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones[1:5], nFathers = 5)
colony <- addWorkers(colony, nInd = 5)
colony <- addDrones(colony, nInd = 5)
colony <- addVirginQueens(colony, nInd = 2)

genoQ <- getQueensSegSiteGeno(colony)
genoF <- getFathersSegSiteGeno(colony)
genoW <- getWorkersSegSiteGeno(colony)
genoD <- getDronesSegSiteGeno(colony)
genoV <- getVirginQueensSegSiteGeno(colony)

# Pool of drones
sexD <- rep(x = "M", times = nrow(genoD))
getPoolGeno(x = genoD, type = "count", sex = sexD)
(poolD <- getPoolGeno(x = genoD, type = "mean", sex = sexD))
# ... compare to queen's genotype
genoQ
plot(y = poolD, x = genoQ[1, ], ylim = c(0, 2), xlim = c(0, 2),
     ylab = "Average allele dosage in drones",
     xlab = "Allele dosage in the queen")

# Pool of workers
getPoolGeno(x = genoW, type = "count")
(poolW <- getPoolGeno(x = genoW, type = "mean"))
# ... compare to fathers' and queen's avearage genotype
sexF <- rep(x = "M", times = nrow(genoF))
sexQ <- rep(x = "F", times = nrow(genoF))
sexFQ <- c(sexF, sexQ)
genoFQ <- rbind(genoF, genoQ[rep(x = 1, times = nrow(genoF)), ])
(poolFQ <- getPoolGeno(x = genoFQ, type = "mean", sex = sexFQ))
plot(y = poolW, x = poolFQ, ylim = c(0, 2), xlim = c(0, 2),
     ylab = "Average allele dosage in workers",
     xlab = "Allele dosage in the queen and fathers")
}
