% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getSegSiteGeno}
\alias{getSegSiteGeno}
\title{Access genotypes for all segregating sites}
\usage{
getSegSiteGeno(pop, chr = NULL, simParam = NULL)
}
\arguments{
\item{pop}{\code{\link{Pop-class}}}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{simParam}{\code{\link{SimParam}}, global simulation parameters}
}
\value{
Matrix of genotypes
}
\description{
Level 0 function that returns genotypes for all segregating
  sites.
}
\seealso{
\code{\link{pullSegSiteHaplo}}
}
