% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getSnpGeno}
\alias{getSnpGeno}
\title{Access SNP array genotypes}
\usage{
getSnpGeno(pop, snpChip = 1, chr = NULL, simParam = NULL)
}
\arguments{
\item{pop}{\code{\link{Pop-class}}}

\item{snpChip}{numeric, indicates which SNP array genotype to retrieve}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{simParam}{\code{\link{SimParam}}, global simulation parameters}
}
\value{
Matrix of genotypes
}
\description{
Level 0 function that returns SNP array genotypes.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(TRUE)
SP$setTrackPed(isTrackPed = TRUE)
SP$addTraitA(10)
SP$addSnpChip(5)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony  class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(x = colony)

# Use getSnpGeno on a Population
getSnpGeno(getWorkers(colony))
getSnpGeno(getQueen(colony))
getSnpGeno(getDrones(colony))
}
\seealso{
\code{\link{pullSnpHaplo}}
}
