% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{getWorkers}
\alias{getWorkers}
\title{Access workers}
\usage{
getWorkers(x, nInd = NULL)
}
\arguments{
\item{x}{Colony or Colonies}

\item{nInd}{numeric, number of workers to access, if NULL all workers
are accessed, otherwise a random sample}
}
\value{
When \code{x} is Colony then return is Pop, population object with workers
When \code{x} is Colonies then return is a list of Pop, population objects with workers
}
\description{
Access workers
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybee
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
getWorkers(colony1)
getWorkers(colony1)@id
getWorkers(colony1, nInd = 2)@id
getWorkers(colony1, nInd = 2)@id

apiary <- c(colony1, colony2)
getWorkers(apiary)
getWorkers(apiary)[[1]]@id
getWorkers(apiary)[[2]]@id

getWorkers(apiary, nInd = 10)
getWorkers(apiary, nInd = 10)[[1]]@id
getWorkers(apiary, nInd = 10)[[2]]@id

}
