% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{hasSuperseded}
\alias{hasSuperseded}
\title{Test to see if colony/colonies have superseded in the current period}
\usage{
\method{hasSuperseded}(x)
}
\arguments{
\item{x}{Undefined argument. Can be a "Colony" class or "Colonies" class}
}
\value{

}
\description{
Test to see if colony/colonies have superseded in the current period
}
\examples{

}
