% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{hasSwarmed}
\alias{hasSwarmed}
\title{Test if colony has swarmed}
\usage{
hasSwarmed(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
logical, named by colony id when \code{x} is \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns colony swarmed status. This will
  obviously have major impact on the colony and its downstream events.
}
\examples{
founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 20)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones[1:5], nFathers = 5)
colony
hasSwarmed(colony)
colony <- buildUpColony(colony, nWorkers = 100)
colony
hasSwarmed(colony)
tmp <- swarmColony(colony)
tmp
hasSwarmed(tmp$swarm)
hasSwarmed(tmp$remnant)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[3])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[4])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
apiary <- c(colony1, colony2)
apiary <- buildUpColonies(apiary, nWorkers = 100)
tmp <- swarmColonies(apiary)
tmp
hasSwarmed(tmp$swarms)
hasSwarmed(tmp$remnants)
}
