% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isCsdHeterozygous}
\alias{isCsdHeterozygous}
\title{Test if individuals are heterozygous at the csd locus}
\usage{
isCsdHeterozygous(pop, simParamBee = NULL)
}
\arguments{
\item{pop}{\code{\link{Pop-class}}}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
logical
}
\description{
Level 0 function that returns if individuals of a population are
  heterozygous at the csd locus. See \code{\link{SimParamBee}} for more
  information about the csd locus.
}
\details{
We could expand \code{isCsdHeterozygous} to work also with
  \code{\link{Colony-class}} and \code{\link{Colonies-class}} if needed
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 3, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 5)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones, nFathers = 5)
colony1 <- addWorkers(colony1, nInd = 10)
colony1 <- addVirginQueens(colony1, nInd = 10)

isCsdHeterozygous(getQueen(colony1))

isCsdHeterozygous(getWorkers(colony1))

isCsdHeterozygous(getVirginQueens(colony1))
}
