% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isDrone}
\alias{isDrone}
\title{Is individual a drone}
\usage{
isDrone(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}}
}
\value{
logical
}
\description{
Level 0 function that tests if individuals are drones
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony <- addWorkers(colony)
colony <- addDrones(colony)
colony <- addVirginQueens(colony)

isDrone(getQueen(colony))
isDrone(getWorkers(colony, nInd = 2))
isDrone(getDrones(colony, nInd = 2))
isDrone(getVirginQueens(colony, nInd = 2))

bees <- c(getQueen(colony),
          getWorkers(colony, nInd = 2),
          getDrones(colony, nInd = 2),
          getVirginQueens(colony, nInd = 2))
isDrone(bees)

}
