% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isGenoHeterozygous}
\alias{isGenoHeterozygous}
\title{Test if a multilocus genotype is heterozygous}
\usage{
isGenoHeterozygous(x)
}
\arguments{
\item{x}{integer or matrix, output from \code{\link{getCsdGeno}}}
}
\value{
logical
}
\description{
Level 0 function that returns heterozygote status for a
  multilocus genotype.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 3, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 5)
colony1 <- createColony(queen = basePop[2], fathers = drones)
colony1 <- addWorkers(colony1, nInd = 10)
colony1 <- addVirginQueens(colony1, nInd = 10)

(tmp <- getCsdGeno(getQueen(colony1)))
SIMplyBee:::isGenoHeterozygous(tmp)

(tmp <- getCsdGeno(getVirginQueens(colony1)))
SIMplyBee:::isGenoHeterozygous(tmp)

(tmp <- getCsdGeno(getWorkers(colony1)))
SIMplyBee:::isGenoHeterozygous(tmp)

}
