% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isQueen}
\alias{isQueen}
\title{Is individual a queen}
\usage{
isQueen(x, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Pop-class}}}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
logical
}
\description{
Level 0 function that tests if individuals are queens
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones, nFathers = 5)
colony <- addWorkers(colony)
colony <- addDrones(colony)
colony <- addVirginQueens(colony)

isQueen(getQueen(colony))
isQueen(getWorkers(colony, nInd = 2))
isQueen(getDrones(colony, nInd = 2))
isQueen(getVirginQueens(colony, nInd = 2))

bees <- c(
  getQueen(colony),
  getWorkers(colony, nInd = 2),
  getDrones(colony, nInd = 2),
  getVirginQueens(colony, nInd = 2)
)
isQueen(bees)
}
