% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isQueenMated}
\alias{isQueenMated}
\title{Is the queen mated?}
\usage{
isQueenMated(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}
}
\value{
logical, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns queen's mating status.
}
\examples{
founderGenomes <- quickHaplo(nInd = 6, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 20)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
colony3 <- createColony(x = basePop[4])
apiary <- c(colony1, colony2, colony3)
isQueenMated(getQueen(colony1))
isQueenMated(colony1)
isQueenMated(apiary)

isQueenMated(removeQueen(colony1))

isQueenMated(supersedeColony(colony2))

isQueenMated(c(getQueen(colony1), getQueen(colony2), getVirginQueens(colony3)))

isQueenMated(crossVirginQueen(basePop[6], drones[11:20], nFather = 5))
}
