% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isQueenPresent}
\alias{isQueenPresent}
\title{Is the queen present}
\usage{
isQueenPresent(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
logical, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns queen's present status (is she
  present/alive or not).
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)
isQueenPresent(colony1)
isQueenPresent(colony2)
isQueenPresent(apiary)

colony1 <- removeQueen(colony1)
isQueenPresent(colony1)

colony2 <- supersedeColony(colony2)
isQueenPresent(colony2)

}
