% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isVirginQueen}
\alias{isVirginQueen}
\title{Is individual a virgin queen}
\usage{
isVirginQueen(x, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Pop-class}}}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
logical
}
\description{
Level 0 function that tests if individuals are virgin queens
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
droneGroups <- pullDroneGroupsFromDCA(drones, n = 10, nDrones = nFathersPoisson)

# Create a Colony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, drones = droneGroups[[1]])
colony <- buildUp(x = colony, nWorkers = 120, nDrones = 20)
colony <- addVirginQueens(x = colony, nInd = 4)

# Test isVirginQueen on colony class
isVirginQueen(getQueen(colony))
isVirginQueen(getFathers(colony, nInd = 2))
isVirginQueen(getVirginQueens(colony, nInd = 2))

bees <- c(
  getQueen(colony),
  getFathers(colony, nInd = 2),
  getVirginQueens(colony, nInd = 2)
)
isVirginQueen(bees)
}
