% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isVirginQueensPresent}
\alias{isVirginQueensPresent}
\title{Are virgin queen(s) present}
\usage{
isVirginQueensPresent(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
logical, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns virgin queen(s) presence status.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- addVirginQueens(x = colony, nInd = 4)
isVirginQueensPresent(colony)

apiary <- createColonies(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2,3)])
apiary <- buildUp(x = apiary, nWorkers = 100, nDrones = 10)
isVirginQueensPresent(apiary)

colony <- removeQueen(colony)
isVirginQueensPresent(colony)

tmp <- swarm(x = apiary)
isVirginQueensPresent(tmp$swarm)
isVirginQueensPresent(tmp$remnant)
}
