% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nColonies}
\alias{nColonies}
\title{Number of colonies in a MultiColony object}
\usage{
nColonies(multicolony)
}
\arguments{
\item{multicolony}{\code{\link{MultiColony-class}}}
}
\description{
Level 0 function that returns the number of colonies in a
  MultiColony object.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)
apiary <- createMultiColony(basePop[2:3], n = 2)
nColonies(apiary)
nColonies(createMultiColony(n = 10))
}
