% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nCsdAlleles}
\alias{nCsdAlleles}
\title{Report the number of distinct csd alleles}
\usage{
nCsdAlleles(x, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
integer representing the number of distinct csd alleles when \code{x}
  is \code{\link{Pop-class}} (or ), list of integer
  when \code{x} is \code{\link{Colony-class}} (list nodes named by caste) and
  list of a list of integer when \code{x} is \code{\link{Colonies-class}},
  outer list is named by colony id when \code{x} is
  \code{\link{Colonies-class}}; the integer rep
}
\description{
Level 0 function that returns the number of distinct csd alleles
  in input. See \code{\link{SimParamBee}} for more information about the csd
  locus.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 3, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony1 <- addDrones(colony1, nInd = 10)
colony1 <- addVirginQueens(colony1, nInd = 3)
colony2 <- addWorkers(colony2, nInd = 20)
apiary <- c(colony1, colony2)

nCsdAlleles(getQueen(colony1))
nCsdAlleles(colony1)
nCsdAlleles(apiary)

nCsdAlleles(colony1, collapse = TRUE)
nCsdAlleles(apiary, collapse = TRUE)

}
