% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nDrones}
\alias{nDrones}
\title{Number of drones in a colony}
\usage{
nDrones(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}
}
\value{
integer, named by colony id when \code{x} is \code{\link{MultiColony-class}}
}
\description{
Returns the number of drones in a colony
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
droneGroups <- pullDroneGroupsFromDCA(drones, n = 10, nDrones = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, drones = droneGroups[[1]])
colony <- addDrones(x = colony, nInd = 50)

apiary <- createMultiColony(basePop[3:7], n = 5)
apiary <- cross(x = apiary, drones = droneGroups[2:6])
tmp <- addDrones(x = apiary[1:3], nInd = 100)
tmp2 <- addDrones(x = apiary[4:5], nInd = 200)
apiary <- c(tmp, tmp2)

nDrones(colony)
nDrones(apiary)
}
