% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SimParamBee.R
\name{nFathersPoisson}
\alias{nFathersPoisson}
\alias{nFathersTruncPoisson}
\title{Functions to sample the number of fathers}
\usage{
nFathersPoisson(n = 1, lambda = 15)

nFathersTruncPoisson(n = 1, lambda = 15)
}
\arguments{
\item{n}{integer, number of samples}

\item{lambda}{numeric, average number of fathers}
}
\value{
numeric, number of fathers
}
\description{
Functions to sample the number of fathers when
  \code{nFathers = NULL} in various functions
}
\details{
\code{nFathersPoisson()} samples from a Poisson distribution, which
  can return a vale 0 (that would lead to a failed queen mating), while
  \code{nFathersTruncPoisson} samples from a truncated Poisson distribution
  (truncated at zero) to avoid failed matings
}
\section{Functions}{
\itemize{
\item \code{nFathersTruncPoisson}: Functions to sample the number of fathers
}}

\examples{
nFathersPoisson()
hist(nFathersPoisson(n = 100))

nFathersTruncPoisson()
hist(nFathersTruncPoisson(n = 100))
}
