% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nWorkers}
\alias{nWorkers}
\title{Number of workers in a colony}
\usage{
nWorkers(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
integer, named by colony id when \code{x} is \code{\link{Colonies-class}}
}
\description{
Returns the number of workers in a colony
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
colony1 <- addWorkers(colony1, nInd = 100)
colony2 <- addWorkers(colony2, nInd = 200)
nWorkers(colony1)
nWorkers(colony2)

apiary <- c(colony1, colony2)
nWorkers(apiary)
}
