% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SimParamBee.R
\name{pDownsizeUnif}
\alias{pDownsizeUnif}
\alias{pDownsizeColonyStrength}
\title{Sample the proportion of removed workers in downsizing}
\usage{
pDownsizeUnif(colony, n = 1, min = 0.8, max = 0.9)

pDownsizeColonyStrength(colony, n = 1, nWorkersFull = 100, scale = 1)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{n}{integer, number of samples}

\item{min}{numeric, lower limit for \code{pDownsizeUnif}}

\item{max}{numeric, upper limit for \code{pDownsizeUnif}}

\item{nWorkersFull}{numeric, average number of workers in a full/strong
colony for \code{pSplitColonyStrength} (actual number can go beyond this
value)}

\item{scale}{numeric, scaling of numbers in \code{pDownsizeColonyStrength}
to avoid to narrow range when colonies have a large number of bees (in that
case change \code{nWorkersFull} too!)}
}
\value{
numeric, proportion of removed workers
}
\description{
Sample the proportion of removed workers in downsizing when
  \code{p = NULL} (see \code{\link{SimParamBee}$pDownsize}) - this is just an
  example and you will likely want to define your own sampling functions!
}
\details{
\code{pDownsizeUnif} samples from a uniform distribution between
  values 0.8 and 0.9 irrespective of colony strength.
  \code{pDownsizeColonyStrength} samples from a beta distribution with mean
  \code{a / (a + b)}, where \code{a = nWorkers} and \code{b = 3 * nWorkers +
  nWorkersFull}. This beta sampling mimics larger winter colonies for strong
  colonies and small winter colonies for weak colonies - see examples - this
  is just an example, not based on actual data!

  The \code{nWorkersFull} default value used in this function is geared
  towards a situation where we simulate ~100 workers per colony (down-scaled
  simulation for efficiency). If you simulate more workers, you should change
  the default accordingly.
}
\section{Functions}{
\itemize{
\item \code{pDownsizeColonyStrength}: Sample the proportion of removed workers in downsizing
based on colony strength
}}

\examples{
pDownsizeUnif()
pDownsizeUnif()
p <- pDownsizeUnif(n = 1000)
hist(p, breaks = seq(from = 0, to = 1, by = 0.01), xlim = c(0, 1))

# Example for pDownsizeColonyStrength()
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)
drones <- createDrones(x = basePop[1], nInd = 5)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones, nFathers = 5)
colony <- addWorkers(colony, nInd = 10)
nWorkers(colony) # weak colony
pDownsizeColonyStrength(colony)
pDownsizeColonyStrength(colony)
colony <- addWorkers(colony, nInd = 100)
nWorkers(colony) # strong colony
pDownsizeColonyStrength(colony)
pDownsizeColonyStrength(colony)

# Logic behind pDownsizeColonyStrength()
nWorkersFull <- 100
nWorkers <- 0:200
pRemove <- 1 - rbeta(n = length(nWorkers),
                     shape1 = nWorkers,
                     shape2 = 3 * nWorkers + nWorkersFull)
plot(pRemove ~ nWorkers, ylim = c(0, 1)); abline(v = nWorkersFull)
pNew <- 1 - pRemove
plot(pNew ~ nWorkers, ylim = c(0, 1)); abline(v = nWorkersFull)
}
\seealso{
\code{\link{SimParamBee}} field \code{pDownsize}
}
