% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{pullDroneGroupsFromDCA}
\alias{pullDroneGroupsFromDCA}
\title{Pulls drone groups from a Drone Congregation Area (DCA)}
\usage{
pullDroneGroupsFromDCA(DCA, nGroup, avgGroupSize)
}
\arguments{
\item{DCA}{Pop, population of drones}

\item{avgGroupSize}{Numeric, average number of drones per group}

\item{nGroups}{Integer, number of drone groups to be created}
}
\value{
Two AlphaSim population objects of the colony and the group of pulled individuals.
}
\description{
Pulls drone groups from a Drone Congregation Area (DCA) to use them later in
mating. Number of drones per group is sampled from a Poisson distribution with
average group size. Pulled drones are removed from the DCA.
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybee
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addDrones(colony1, nInd = 10)
colony2 <- addDrones(colony2, nInd = 20)
apiary <- c(colony1, colony2)
DCA <- createDCA(apiary)
pullDroneGroupsFromDCA(DCA, nGroup = 4, avgGroupSize = 5)

}
