% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{pullInd}
\alias{pullInd}
\title{Pull individuals from a population}
\usage{
pullInd(pop, nInd = NULL, use = "rand")
}
\arguments{
\item{pop}{\code{\link{Pop-class}}}

\item{nInd}{numeric, number of individuals to pull, if \code{NULL} pull all
individuals}

\item{use}{character, all options provided by \code{\link{selectInd}}}
}
\value{
list with a node \code{pulled} holding \code{\link{Pop-class}} of
  pulled individuals and a node \code{remainder)} holding \code{\link{Pop-class}}
  of remaining individuals
}
\description{
Level 1 function that pulls individuals from a population and
  update the population (these individuals don't stay in a population).
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

pullInd(basePop, nInd = 2)
pullInd(basePop, nInd = 3)
pullInd(basePop)

}
