% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{pullQueen}
\alias{pullQueen}
\title{Pull queen from a colony}
\usage{
pullQueen(x)
}
\arguments{
\item{x}{Colony or Colonies}
}
\value{
When \code{x} is Colony then return is a list with a population object (Pop) with the pulled queen and remaining colony
When \code{x} is Colonies then return is a list with a list of population object (Pop) with the pulled queens and remaining colonies
}
\description{
Pull queen from a colony
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybee
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony2 <- addDrones(colony2, nInd = 5)
pullQueen(colony1)

apiary <- c(colony1, colony2)
pullQueen(apiary)
nQueen(apiary)
nQueen(pullQueen(apiary)$colonies)

}
