% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeCastePop}
\alias{removeCastePop}
\alias{removeQueen}
\alias{removeVirginQueens}
\alias{removeWorkers}
\alias{removeDrones}
\title{Remove a proportion of caste population individuals from a colony}
\usage{
removeCastePop(x, caste = NULL, p = 1, use = "rand")

removeQueen(x)

removeVirginQueens(x, p = 1, use = "rand")

removeWorkers(x, p = 1, use = "rand")

removeDrones(x, p = 1, use = "rand")
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}

\item{caste}{character, "queen", "workers", "drones", or "virginQueens"}

\item{p}{numeric, proportion to be removed}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of virgins queens that will stay when \code{p < 1}}
}
\value{
\code{\link{Colony-class}} or \code{\link{MultiColony-class}} without virgin queens
}
\description{
Level 2 function that removes a proportion of virgin queens of
  a Colony or MultiColony object
}
\section{Functions}{
\itemize{
\item \code{removeQueen()}: Remove queen from a colony

\item \code{removeVirginQueens()}: Remove virgin queens from a colony

\item \code{removeWorkers()}: Remove workers from a colony

\item \code{removeDrones()}: Remove workers from a colony

}}
\examples{
founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 100)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 5, nFathers = nFathersPoisson)

# Create and cross Colony and MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
apiary <- createMultiColony(basePop[4:5], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[3:4])

# Add virgin queens
colony <- addVirginQueens(colony, nInd = 10)
apiary <- addVirginQueens(apiary, nInd = 10)

# Remove virgin queens
nVirginQueens(colony)
colony <- removeCastePop(colony, caste = "virginQueens", p = 0.5)
# or shortcut: removeVirginQueens(colony, p = 0.5)
nVirginQueens(colony)
colony <- removeCastePop(colony, caste = "virginQueens")
# or shortcut: colony <- removeVirginQueens(colony)
nVirginQueens(colony)
nWorkers(colony)
nDrones(colony)
colony <- removeCastePop(colony, caste = "workers", p = 0.3)
# or shortcut: colony <- removeWorkers(colony, p = 0.3)
colony <- removeCastePop(colony, caste = "drones", p = 0.3)
# or shortcut: colony <- removeDrones(colony, p = 0.3)

nVirginQueens(apiary)
nWorkers(apiary)
nDrones(apiary)
apiary <- removeCastePop(apiary, caste = "virginQueens", p = 0.3)
# or shortcut: removeVirginQueens(apiary, p = 0.3)
nVirginQueens(apiary)
apiary <- removeCastePop(apiary, caste = "workers", p = 0.3)
# or shortcut: removeWorkers(apiary, p = 0.3)
nWorkers(apiary)
#' apiary <- removeCastePop(apiary, caste = "drones", p = 0.3)
# or shortcut: removeDrones(apiary, p = 0.3)
nDrones(apiary)
}
