% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{removeColonies}
\alias{removeColonies}
\title{Remove the colonies from the colony list based on IDs.}
\usage{
\method{removeColonies}(colonies, ID)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{ID}{IDs of "colony" class objects listed in the "colonies" object}
}
\value{
A list of remaining colonies. Updated AlphaSimRBee Colonies object
}
\description{
Remove the colonies from the list of all colonies based
on colony IDs and return a list of remaining colonies.
}
\examples{
# Create founder haplotypes

founderPop <- quickHaplo(nInd=300, nChr=1, segSites=10)

# Set simulation parameters

SP <- SimParam$new(founderPop)

# Create population

pop <- newPop(founderPop, simParam=SP)

# Create colonies

founderDrones <- createFounderDrones(pop[13:300], nDronesPerQueen = 17)
colony1 <- createColony(queen = pop[1], fathers = founderDrones[1:17])
colony2 <- createColony(queen = pop[2], fathers = founderDrones[18:37])
colony3 <- createColony(queen = pop[3], fathers = founderDrones[37:51])

# Put the colonies together to the apiary

apiary <- c(colony1, colony2, colony3)

# Remove colonies

apiary <- removeColonies(apiary, ID = c(1,2))

}
