% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeDrones}
\alias{removeDrones}
\title{Remove a proportion of drones}
\usage{
removeDrones(colony, p = 1, use = "rand")
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{p}{numeric, proportion to be removed}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of drones that will stay when \code{p < 1}}
}
\value{
\code{\link{Colony-class}} without drones
}
\description{
Level 2 function that removes a proportion of drones, for
  example, at the end of summer.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
colony
colony <- addDrones(colony, nInd = 10)
colony
getDrones(colony)@id

colony <- removeDrones(colony, p = 0.5)
colony
getDrones(colony)@id

colony <- removeDrones(colony, p = 1.0)
colony
getDrones(colony)
}
