% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeWorkers}
\alias{removeWorkers}
\title{Remove selected percentage of workers}
\usage{
\method{removeWorkers}(colony, p)
}
\arguments{
\item{colony}{Colony class. AlphaSimR Colony object from the \code{createColony(...)} call}

\item{p}{Numeric. 0<=p>=1 .}
}
\value{
Updated AlphaSimRBee Colony object
}
\description{
To decrease the number of workers (for example, in winter)
}
\examples{
inst/examples/examples_removeWorkers.R
}
