% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{replaceDrones}
\alias{replaceDrones}
\title{Replace a proportion of drones with new ones}
\usage{
replaceDrones(x, p = 1, use = "rand")
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{p}{numeric, proportion of drones to be replaced with new ones}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of drones that stay when \code{p < 1}}
}
\value{
\code{\link{Colony-class}} or \code{\link{Colonies-class}} with
  replaced drones
}
\description{
Level 2 function that replaces a proportion of drones with new
  drones from the colony. Useful after events like season change, swarming,
  supersedure, etc. due to the short life span of the drones.
}
\examples{
founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 100)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 5, nFathers = nFathersPoisson)

# Create and cross Colony and MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
apiary <- createColonies(basePop[4:5], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[3:4])

# Add drones
colony <- addDrones(colony, nInd = 20)
apiary <- addDrones(apiary, nInd = 20)

# Replace drones
getDrones(colony)@id
colony <- replaceDrones(colony, p = 0.5)
getDrones(colony)@id
colony <- replaceDrones(colony, p = 1.5)
getDrones(colony)@id

lapply(getDrones(apiary), FUN = function(x) x@id)
apiary <- replaceDrones(apiary, p = 0.5)
lapply(getDrones(apiary), FUN = function(x) x@id)
apiary <- replaceDrones(apiary, p = 1.5)
lapply(getDrones(apiary), FUN = function(x) x@id)
}
