% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{replaceDronesColonies}
\alias{replaceDronesColonies}
\title{Replace drones in all the colonies of a Colonies object}
\usage{
replaceDronesColonies(colonies)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}
}
\value{
An updated AlphaSimRBee Colonies object
}
\description{
The function replaces a given proportion of drones in all the colonies
of a Colonies object with new drones from the same queen. The default percentage
is set to 1, hence replacing all drones. If some colonies do not have drones, the
function does not add them.
}
\examples{
#Create founder haplotypes
founderPop <- quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP <- SimParam$new(founderPop)

#Create population
base <- newPop(founderPop, simParam=SP)

#Create 10 virgin queen colonies
 apiary1 <- createMultipleMatedColonies(founderPop = base, nColonies = 10, nAvgFathers = 15)

#Build up colonies by adding 1000 workers and 100 drones to each colony in the "colonies" list
 apiary1 <- buildUpColonies(apiary1, nWorkers = 1000, nDrones = 100)

#Replace the drones
  apiary1 <- replaceDrones(apiary1)

}
