% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{replaceVirginQueens}
\alias{replaceVirginQueens}
\title{Replace a proportion of virgin queens with new ones}
\usage{
replaceVirginQueens(
  colony,
  p = 1,
  use = "rand",
  year = NULL,
  simParamBee = NULL
)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{p}{numeric, proportion of virgin queens to be replaced with new ones}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of virgin queens that stay when \code{p < 1}}

\item{year}{numeric, year of birth for virgin queens}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colony-class}} with replaced virgin queens
}
\description{
Level 2 function that replaces a proportion of virgin queens with new
  virgin queens from the colony. Useful after events like season change, swarming,
  supersedure, etc. due to the short life span of the virgin queens.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
colony
colony <- addVirginQueens(colony, nInd = 10)
colony
getVirginQueens(colony)@id
colony <- replaceVirginQueens(colony, p = 0.5)
getVirginQueens(colony)@id

colony <- replaceVirginQueens(colony, p = 1.0)
getVirginQueens(colony)@id

colony <- replaceVirginQueens(colony, p = 1.5)
getVirginQueens(colony)@id

}
