% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{replaceWorkersColonies}
\alias{replaceWorkersColonies}
\title{Replace a proportion of workers with new ones for all given colonies}
\usage{
replaceWorkersColonies(colonies, p = 1, use = "rand", simParamBee = NULL)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{p}{numeric, proportion of workers to be replaced with new ones}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of workers that stay when \code{p < 1}}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colonies-class}} with replaced workers
}
\description{
Level 3 function that does the same as
  \code{\link{replaceWorkers}} but for all given colonies.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)
apiary <- buildUpColonies(apiary, nWorkers = 10, nDrones = 10)
apiary
apiary[[1]]
getWorkers(apiary[[1]])@id
apiary[[2]]
getWorkers(apiary[[2]])@id
lapply(X = getWorkers(apiary), FUN = function(z) z@id)

apiary <- replaceWorkersColonies(apiary)
lapply(X = getWorkers(apiary), FUN = function(z) z@id)

}
