% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{reQueenColony}
\alias{reQueenColony}
\title{Re-queen a colony}
\usage{
reQueenColony(colony, queen)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{queen}{\code{\link{Pop-class}} with one individual that will be the
queen of the colony; if she is not mated, she will be added as a virgin
queen that will have to be mated later.}
}
\value{
\code{\link{Colony-class}} with new queen
}
\description{
Level 2 function that re-queens a colony adds a mated or a
  virgin queen, removes the previous queen, and changes colony id to the new
  mated queen.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony

virginQueen <- basePop[3]
reQueenColony(colony, queen = virginQueen)

matedQueen <- crossVirginQueen(pop = basePop[3], fathers = drones[6:10])
reQueenColony(colony, queen = matedQueen)

}
