% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{setColonyPheno}
\alias{setColonyPheno}
\title{Set colony phenotype}
\usage{
setColonyPheno(x, FUN = NULL, reset = TRUE, simParamBee = NULL, ...)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}

\item{FUN}{function, any function that can be run on \code{colony} and
returns colony phenotypes; if \code{NULL} then
\code{\link{SimParamBee}$colonyPheno} is used - if even this is \code{NULL},
then colony phenotype is not set, but phenotypes of colony individuals are
(see \code{reset} though!)}

\item{reset}{logical, should previous phenotype values for individuals in the
castes be used or sampled anew?}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}

\item{...}{arguments passed to \code{\link{setPheno}}}
}
\value{
\code{\link{Colony-class}} or \code{\link{MultiColony-class}} with phenotypes
}
\description{
Level 2 function that sets phenotypes for all colony individuals
  (queen, workers, drones, and virgin queens) and for the colony, or each
  colony in a MultiColony object.
}
\details{
When this function is called on a colony, phenotypes for all colony
  individuals and possibly the whole colony are set (or reset if phenotypes
  already exist).
}
\examples{
founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)

# Define two traits that collectively affect colony honey yield:
# 1) queen's effect on colony honey yield, say via pheromone secretion phenotype
# 2) workers' effect on colony honey yield, say via foraging phenotype
# The traits will have a negative genetic correlation of -0.5 and heritability
# of 0.25 (on an individual level)
meanP <- c(20, 0)
nWorkers <- 10
varA <- c(1, 1 / nWorkers)
corA <- matrix(data = c(
  1.0, -0.5,
  -0.5, 1.0
), nrow = 2, byrow = TRUE)
varE <- c(3, 3 / nWorkers)
varA / (varA + varE)
SP$addTraitA(nQtlPerChr = 100, mean = meanP, var = varA, corA = corA)

basePop <- createVirginQueens(founderGenomes)
drones <- createDrones(x = basePop[1], nInd = 100)
droneGroups <- pullDroneGroupsFromDCA(drones, n = 5, nDrones = 14)

# Create and cross Colony and MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, drones = droneGroups[[1]])
colony <- buildUp(colony, nWorkers = nWorkers, nDrones = 3)
apiary <- createMultiColony(basePop[3:5], n = 2)
apiary <- cross(apiary, drones = droneGroups[c(2, 3)])
apiary <- buildUp(apiary, nWorkers = nWorkers, nDrones = 3)

# Set phenotypes for all colony individuals
colony <- setColonyPheno(colony, varE = varE)

apiary <- setColonyPheno(apiary, varE = varE)

# Queen's phenotype for both traits
getQueenPheno(colony)
getQueenPheno(apiary)

# Workers' phenotype for both traits
getWorkersPheno(colony)
getWorkersPheno(apiary)

# For the whole colony
getColonyPheno(colony)
getColonyPheno(apiary)

# Set phenotypes for all colony individuals AND Colony
colony <- setColonyPheno(colony, FUN = calcColonyPhenoFromCaste, varE = varE)
getColonyPheno(colony)$colony

# Set phenotypes for all colony individuals AND MultiColony
apiary <- setColonyPheno(apiary, FUN = calcColonyPhenoFromCaste, varE = varE)
sapply(X = getColonyPheno(apiary), FUN = function(x) x$colony)

# Colony phenotype - store the colony function into the SP object
SP$colonyPheno <- calcColonyPhenoFromCaste
getColonyPheno(setColonyPheno(colony, varE = varE))$colony
getColonyPheno(setColonyPheno(colony, varE = varE))$colony
sapply(X = getColonyPheno(setColonyPheno(apiary, varE = varE)), FUN = function(x) x$colony)
sapply(X = getColonyPheno(setColonyPheno(apiary, varE = varE)), FUN = function(x) x$colony)
# phenotype will vary between function calls by default (see reset)
}
\seealso{
\code{\link{calcColonyPhenoFromCaste}} as an example for \code{FUN}
  and \code{\link{setPheno}} for the basic population phenotyping
}
