% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{setMisc}
\alias{setMisc}
\title{Set miscelaneous information in a population}
\usage{
setMisc(x, slot, value)
}
\arguments{
\item{x}{\code{\link{Pop-class}}}

\item{slot}{character, name of the node to set within the \code{x@misc} slot}

\item{value, }{value to be saved into \code{x@misc[[*]][[slot]]}; length of
\code{value} should be equal to \code{nInd(x)}; if its length is 1, then
it is repeated using \code{rep}}
}
\value{
\code{\link{Pop-class}} with \code{x@misc[[*]][[slot]]} set
}
\description{
Level 1 function that sets the queen's year of birth.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)
basePop@misc

basePop <- setMisc(basePop, slot = "info", value = c("A", "B", "C"))
basePop@misc

basePop <- setMisc(basePop, slot = "info", value = c("B", "C", "A"))
basePop@misc

basePop <- setMisc(basePop, slot = "info2", value = "A")
basePop@misc

}
