% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{setPhenoColonies}
\alias{setPhenoColonies}
\title{TODO}
\usage{
setPhenoColonies(colonies, FUN = NULL, ...)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{FUN}{TODO}

\item{...}{TODO}
}
\value{
\code{\link{Colonies-class}} with phenotypes
}
\description{
Level 3 function that TODO
}
\examples{
TODO

}
