% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{setQueensYOB}
\alias{setQueensYOB}
\title{Set the queen's year of birth}
\usage{
\method{setQueenYOB}(colony)
}
\arguments{
\item{x}{Undefined argument. Can be a "Pop" class or "Colony" class}

\item{year}{Integer, the year of the birth of the queen}
}
\value{
AlphaSimRBee Colony object
}
\description{
Set the year of birth of the queen in the \code{colony@queen@misc$yearOfBirth} slot
}
\examples{
#Create founder haplotypes
founderPop <- quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP <- SimParam$new(founderPop)

#Create population
pop <- newPop(founderPop, simParam=SP)

#Creates colony
colony1 <- createColony(queen = base[1], fathers = base[2:15])
setQueenAge(colony, 1)

}
