% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{simulateHoneyBeeGenomes}
\alias{simulateHoneyBeeGenomes}
\title{Simulate the Honey bee genome, including the csd locus}
\usage{
\method{simulateHoneyBeeGenomes}(nInd, nChr, nSegSites, Ne, nBp. genLen,
                                        mutRate, histNe, histGen, split, csdChr,
                                        csdPos, nCsdHaplos, nThreads)
}
\arguments{
\item{nInd}{number of individuals to simulate}

\item{nChr}{number of chromosomes to simulate}

\item{nSegSites}{number of segregating sites to keep per chromosome}

\item{Ne}{effective population size}

\item{nBp}{base pair length of chromosome}

\item{genLen}{genetic length of chromosome in Morgans}

\item{mutRate}{per base pair mutation rate}

\item{histNe}{effective population size in previous generations}

\item{histGen}{number of generations ago for effective population sizes given in histNe}

\item{split}{an optional historic population split in terms of generations ago}

\item{csdChr}{which number chromosome the csd locus is located on}

\item{csdPos}{vector. position of the csd locus on the chromosome}

\item{nCsdHaplos}{number of possible csd haplotypes}

\item{nThreads}{if OpenMP is available, this will allow for simulating chromosomes in parallel.
If the value is NULL, the number of threads is automatically detected}
}
\value{

}
\description{

}
\examples{
tmp <- simulateHoneyBeeGenomes(nInd = 1000, ...)
founderGenomes <- tmp$founderGenomes
SP <- tmp$SP
csd <- tmp$csd
rm(tmp)

}
