% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{splitColonies}
\alias{splitColonies}
\title{splitColonies}
\usage{
\method{splitColonies}(colonies, crossVirginQueen, fathers, pWorkers, pDrones)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{pDrones}{}
}
\value{
Two AlphaSim population objects of the split colonies and the remaining colonies
}
\description{
: Two AlphaSim population objects of the split colonies and the remaining colonies are created.
Spit the colony into two new colonies to prevent swarming (in managed populations)
- one colony is with the old queen and a part of the workers and drones (this is the remaining colony)
- the split colony is taken to a new location with part of the workers.
 A new mated queen can be introduced to the split colony.
 If no new queen is introduced, a virgin queen must be present to mate with fathers from DCA and continue colony
}
\examples{
#Create founder haplotypes
founderPop <- quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP <- SimParam$new(founderPop)

#Create population
base <- newPop(founderPop, simParam=SP)

#Create 10 virgin queen colonies
 apiary1 <- createMultipleMatedColonies(founderPop = base, nColonies = 10, nAvgFathers = 15)

#Build up colonies by adding 1000 workers and 100 drones to each colony in the "colonies" list
 apiary1 <- buildUpColonies(apiary1, nWorkers = 1000, nDrones = 100)

 TODO FINISH EXAMPLE

}
