% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{splitColony}
\alias{splitColony}
\title{Split colony in two colonies}
\usage{
splitColony(colony, p = NULL, year = NULL, simParamBee = NULL)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{p}{numeric, proportion of workers that will go to the split colony; if
\code{NULL} then \code{\link{SimParamBee}$splitP} is used}

\item{year}{numeric, year of birth for virgin queens}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
list with two \code{\link{Colony-class}}, the \code{split} and the
  \code{remnant} (see the description what each colony holds!); both colonies
  have the split even slot set do \code{TRUE}
}
\description{
Level 2 function that splits colony into two new colonies to
  prevent swarming (in managed situation). The remnant colony retains the
  queen and a proportion of the workers and all drones. The split colony gets
  the other part of the workers, which raise virgin queens, of which only one
  prevails. Location of the split is the same as for the remnant.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones, nFathers = 5)
(colony <- buildUpColony(colony, nWorkers = 100))

tmp <- splitColony(colony)
tmp$split
tmp$remnant
}
