% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{supersedeColony}
\alias{supersedeColony}
\title{Supersede colony}
\usage{
supersedeColony(colony)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}
}
\value{
\code{\link{Colony-class}} with the supersede event set to
  \code{TRUE}
}
\description{
Level 2 function that supersedes colony - an event where the
  queen dies and one of virgin queens prevails. The workers and drones stay
  unchanged.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony <- createColony(queen = basePop[2], fathers = drones)
(colony <- buildUpColony(colony, nWorkers = 100))

supersedeColony(colony)

}
