% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{swarm}
\alias{swarm}
\title{Swarm}
\usage{
swarm(x, p = NULL, year = NULL, nVirginQueens = NULL, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{p}{numeric, proportion of workers that will leave with the swarm colony;
if \code{NULL} then \code{\link{SimParamBee}$swarmP} is used}

\item{year}{numeric, year of birth for virgin queens}

\item{nVirginQueens}{integer, the number of virgin queens to be created in the
colony; of these one is randomly selected as the new virgin queen of the
remnant colony. If \code{NULL}, the value from \code{simParamBee$nVirginQueens}
is used}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
list with two \code{\link{Colony-class}} or \code{\link{Colonies-class}},
the \code{swarm} and the \code{remnant} (see the description what each colony holds!); both
outputs have the swarm event set to \code{TRUE}
}
\description{
Level 2 function that swarms colony - an event where the queen
  leaves with a proportion of workers to create a new colony (the swarm). The
  remnant colony retains the other proportion of workers and all drones, and
  the workers raise virgin queens, of which only one prevails. Location of
  the swarm is the same as for the remnant (for now).
}
\examples{
founderGenomes <- quickHaplo(nInd = 10, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)
drones = createDrones(basePop[1], n = 1000)
fatherGroups = pullDroneGroupsFromDCA(drones, n = 10, nFathers = 10)

# Create Colony and MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
(colony <- buildUp(colony, nWorkers = 100))
apiary <- createColonies(basePop[3:8], n = 6)
apiary <- cross(apiary, fathers = fatherGroups[2:7])
apiary <- buildUp(apiary, nWorkers = 100)

# Swarm a colony
tmp <- swarm(colony)
tmp$swarm
tmp$remnant

# Swarm all colonies in the apiary with p = 0.6 (60\% of workers leave)
tmp <- swarm(apiary, p = 0.6)
tmp$swarm[[1]]
tmp$remnant[[1]]

# Sample colonies from the apiary that will swarm (sample with probability of 0.2)
tmp <- pullColonies(apiary, p = 0.2)
# Swarm only the pulled colonies
(swarm(tmp$pulledColonies, p = 0.6))

}
