% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{swarmColony}
\alias{swarmColony}
\title{Swarm colony}
\usage{
swarmColony(colony, p = NULL, year = NULL, simParamBee = NULL)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{p}{numeric, proportion of workers that will leave with the swarm colony}

\item{year}{numeric, year of birth for virgin queens}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
list with two \code{\link{Colony-class}}, the \code{swarm} and the
  \code{remnant} (see the description what each colony holds!); both colonies
  have the swarm event set to \code{TRUE}
}
\description{
Level 2 function that swarms colony - an event where the queen
  leaves with a proportion of workers to create a new colony (the swarm). The
  remnant colony retains the other proportion of workers and all drones, and
  the workers raise virgin queens, of which only one prevails. Location of
  the swarm is the same as for the remnant (for now).
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones, nFathers = 5)
(colony <- buildUpColony(colony, nWorkers = 100))

tmp <- swarmColony(colony)
tmp$swarm
tmp$remnant
}
