% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{addDrones}
\alias{addDrones}
\title{}
\usage{
\method{addDrones}(colony, nInd)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object}

\item{nInd}{Integer, number of drones to add.}
}
\value{
Updated AlphaSimRBee Colony object
}
\description{
Create drones and store them in the \code{colony@drones} slot. If there is
already some drones present in the hive, the function will not overwrite them but instead
combine the newly created and the existing drones The function returns the updated colony.
}
\examples{
Create founder haplotypes
founderPop <- quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP <- SimParam$new(founderPop)

#Create population
pop <- newPop(founderPop, simParam=SP)

#Creates colony with queen and fathers
colony1 <- createColony(queen = base[1], fathers = base[2:15])

#Insert workers and drones into the colony
colony1 <- addWorkers(colony1, nInd = 2000)
colony1 <- addDrones(colony, nInd = 100)

}
