% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{addVirginQueens}
\alias{addVirginQueens}
\title{Add (raise) virgin queens in the colony}
\usage{
addVirginQueens(x, nInd = NULL, new = FALSE, year = NULL, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{nInd}{numeric or function, number of virgin queens to add; if
\code{NULL} then \code{simParamBee$nVirginQueens} is used}

\item{new}{logical, should the virgin queens be added a fresh (ignoring
currently present virgin queens in the colony)}

\item{year}{numeric, year of birth for virgin queens}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colony-class}} or \code{\link{Colonies-class}} with
  virgin queens added
}
\description{
Level 2 function that adds (raises) the specified number of
  virgin queens in the colony by crossing the current queen and the fathers.
}
\details{
\code{addVirginQueens} replaces any currently present virgin queens.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)
addVirginQueens(colony1, nInd = 20)
nVirginQueens(addVirginQueens(apiary, nInd = 20))

# Using a default in SP$nVirginQueens
# (just to have some workers - change this to your needs!)
addVirginQueens(colony1)
nVirginQueens(addVirginQueens(apiary))

SP$nVirginQueens <- 15
addVirginQueens(colony1)
nVirginQueens(addVirginQueens(apiary))

nVirginQueensFun <- function(colony) { rpois(n = 1, lambda = 15) }
addVirginQueens(colony1, nInd = nVirginQueensFun)
nVirginQueens(addVirginQueens(apiary, nInd = nVirginQueensFun))

SP$nVirginQueens <- nVirginQueensFun
addVirginQueens(colony1)
nVirginQueens(addVirginQueens(apiary))

}
