% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{beeCross}
\alias{beeCross}
\title{Cross a queen and drones}
\usage{
beeCross(queen, drones, nProgeny = 1, simParamBee = NULL)
}
\arguments{
\item{queen}{\code{\link{Pop-class}}, with a single diploid individual}

\item{drones}{\code{\link{Pop-class}}, with one or more haploid individual(s)}

\item{nProgeny}{integer, number of progeny to create per cross}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters;
passed to \code{\link{reduceGenome}} in generating queen gametes and
\code{\link{mergeGenome}} in merging queen and drone gametes}
}
\value{
\code{\link{Pop-class}} with diploid individuals
}
\description{
Level 1 function that crosses a queen and drones. Drones are
  haploid, while the queen is diploid, so we first generate gametes (with
  recombination) from her and merge them with drone genomes (=gametes), where
  we randomly re-sample drones to get the desired number of progeny. This is
  an utility function, and you most likely want to use
  \code{\link{crossColony}} function.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

queen <- basePop[1]
drones <- createDrones(x = basePop[2], nDronesPerQueen = 5)
beeCross(queen, drones, nProgeny = )

}
