% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{buildUpColonies}
\alias{buildUpColonies}
\title{Build up colony by adding (raising) workers and drones for all given
  colonies}
\usage{
buildUpColonies(
  colonies,
  nWorkers = NULL,
  nDrones = NULL,
  new = FALSE,
  exact = FALSE,
  resetEvents = FALSE,
  simParamBee = NULL
)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{nWorkers}{numeric or function, number of worker; if \code{NULL} then
\code{\link{SimParamBee}$nWorkers} is used (unless \code{new = TRUE},
currently present workers are taken into account and only the missing
difference is added)}

\item{nDrones}{numeric or function, number of drones; if \code{NULL} then
\code{\link{SimParamBee}$nDrones} is used (unless \code{new = TRUE},
currently present drones are taken into account so only the missing
difference is added)}

\item{new}{logical, should the workers and drones be added a fresh (ignoring
currently present workers and drones)}

\item{exact}{logical, if the csd locus is turned on and exact is \code{TRUE},
create the exactly specified number of viable workers (they are
heterozygous at the csd locus)}

\item{resetEvents}{logical, call \code{\link{resetEvents}} as part of the
build up}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colonies-class}} with workers and drones added
}
\description{
Level 3 function that does the same as
  \code{\link{buildUpColony}} but for all given colonies.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
apiary <- c(colony1, colony2)
isProductive(apiary)
apiary[[1]]
apiary[[2]]

# Using defaults in SP
# (just to have some bees - change this to your needs!)
(apiary <- buildUpColonies(apiary))
isProductive(apiary)
apiary[[1]]
apiary[[2]]

apiary <- buildUpColonies(apiary, nWorkers = 100)
apiary[[1]] # we are already at the target
apiary <- buildUpColonies(apiary, nWorkers = 150)
apiary[[1]] # increasing the target
apiary <- buildUpColonies(apiary, nWorkers = 100)
apiary[[1]] # we are already at the target
apiary <- buildUpColonies(apiary, nWorkers = 100, new = TRUE)
apiary[[1]] # adding completely new workers & drones

# Using functions
apiary <- c(colony1, colony2)
tmp <- buildUpColonies(apiary, nWorkers = nWorkersPoisson, nDrones = nDronesPoisson)
tmp[[1]]
tmp[[2]]
tmp <- buildUpColonies(apiary, nWorkers = nWorkersPoisson, nDrones = nDronesPoisson)
tmp[[1]]
tmp[[2]]
# nWorkers and nDrones will vary between function calls when a function is used

# Using functions in simParamBee
SP$nWorkers <- nWorkersPoisson
SP$nDrones <- nDronesPoisson
apiary <- c(colony1, colony2)
tmp <- buildUpColonies(apiary)
tmp[[1]]
tmp[[2]]
tmp <- buildUpColonies(apiary)
tmp[[1]]
tmp[[2]]
# nWorkers and nDrones will vary between function calls when a function is used
}
