% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{buildUpColony}
\alias{buildUpColony}
\title{buildUpColony}
\usage{
\method{buildUpColony}(colony, nWorkers, nDrones)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object}

\item{nWorkers}{Desired number of workers wanted in the colony}

\item{nDrones}{Desired number of drones wanted in the colony}
}
\value{
Updated AlphaSimRBee Colony object
}
\description{
Workers and drones are added to a colony/colonies to build it/them
up to the number of desired workers and drones (nWorkers and nDrones).
A user would want to build up a colony in the Period 1 and if events such as split or swarming
occur.
}
\examples{
#Create founder haplotypes
founderPop <- quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP <- SimParam$new(founderPop)

#Create population
base <- newPop(founderPop, simParam=SP)

#Create 10 mated colonies from the base population
apiary1 <- createMultipleMatedColonies(base, nColonies = 10, nAvgFathers = 15)

#Build-up the colonies
colonyFullSize <- 500
apiary1 <- buildUpColonies(apiary1, nWorkers = colonyFullSize, nDrones = colonyFullSize * 0.1)

}
