% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{calcBeeGRMIbd}
\alias{calcBeeGRMIbd}
\title{Calculate Genomic Relatedness Matrix (GRM) for honey bees from
  Identical By Descent genomic data}
\usage{
calcBeeGRMIbd(x)
}
\arguments{
\item{x}{\code{\link{matrix}} of haplotypes/genomes with allele indicators
for the founders coded as 1, 2, ... Haplotypes/genome are in rows and sites
are in columns; no missing values are allowed (this is not checked!). Row
names are essential (formated as ind_genome as returned by AlphaSimR IBD
functions) to infer the individual and their ploidy (see examples)!}
}
\value{
a list with a matrix of gametic relatedness coefficients (genome) and
  a matrix of individual relatedness coefficients (indiv)
}
\description{
Level 0 function that returns Genomic Relatedness Matrix (GRM)
  for honey bees from Identical By Descent genomic data (tracked alleles
  since the founders) - see references on the background theory.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(isTrackRec = TRUE)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony <- addWorkers(colony, nInd = 5)
colony <- addDrones(colony, nInd = 5)
colony <- addVirginQueens(colony, nInd = 2)

genoQ <- getQueensIbdHaplo(colony)
genoF <- getFathersIbdHaplo(colony)
genoW <- getWorkersIbdHaplo(colony)
genoD <- getDronesIbdHaplo(colony)
genoV <- getVirginQueensIbdHaplo(colony)
SP$pedigree

geno <- rbind(genoQ, genoF, genoW, genoD, genoV)

GRMs <- calcBeeGRMIbd(x = geno)

# library("Matrix"); image(as(GRMs$genome, "Matrix"))
# library("Matrix"); image(as(GRMs$indiv, "Matrix"))

GRMs

x <- diag(GRMs$genome)
hist(x)
summary(x)

x <- GRMs$genome[lower.tri(x = GRMs$genome, diag = FALSE)]
hist(x)
summary(x)

x <- diag(GRMs$indiv)
hist(x)
summary(x)

x <- GRMs$GRMindiv[lower.tri(x = GRMs$GRMIndiv, diag = FALSE)]
hist(x)
summary(x)

qI <- getQueen(colony)@id
fI <- sort(getFathers(colony)@id)
wI <- sort(getWorkers(colony)@id)
dI <- sort(getDrones(colony)@id)
vI <- sort(getVirginQueens(colony)@id)

qG <- c(t(outer(X = qI, Y = 1:2, FUN = paste, sep = "_")))
fG <- paste(fI, 1, sep = "_")
wG <- c(t(outer(X = wI, Y = 1:2, FUN = paste, sep = "_")))
dG <- paste(dI, 1, sep = "_")
vG <- c(t(outer(X = vI, Y = 1:2, FUN = paste, sep = "_")))

# Queen vs others
GRMs$genome[qG, qG]
GRMs$indiv[qI, qI]

GRMs$genome[fG, qG]
GRMs$indiv[fI, qI]

GRMs$genome[wG, qG]
GRMs$indiv[wI, qI]

GRMs$genome[dG, qG]
GRMs$indiv[dI, qI]

GRMs$genome[vG, qG]
GRMs$indiv[vI, qI]

# Fathers vs others
GRMs$genome[fG, fG]
GRMs$indiv[fI, fI]

GRMs$genome[wG, fG]
GRMs$indiv[wI, fI]

GRMs$genome[dG, fG]
GRMs$indiv[dI, fI]

GRMs$genome[vG, fG]
GRMs$indiv[vI, fI]

# Workers vs others
GRMs$genome[wG, wG]
GRMs$indiv[wI, wI]

GRMs$genome[dG, wG]
GRMs$indiv[dI, wI]

GRMs$genome[vG, wG]
GRMs$indiv[vI, wI]

}
\references{

}
