% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{calcBeeGRMIbd}
\alias{calcBeeGRMIbd}
\title{Calculate Genomic Relatedness Matrix (GRM) for honeybees from
  Identical By Descent genomic data}
\usage{
calcBeeGRMIbd(x)
}
\arguments{
\item{x}{\code{\link{matrix}} of haplotypes/genomes with allele indicators
for the founders coded as 1, 2, ... Haplotypes/genome are in rows and sites
are in columns; no missing values are allowed (this is not checked!). Row
names are essential (formated as ind_genome as returned by AlphaSimR IBD
functions) to infer the individual and their ploidy (see examples)!}
}
\value{
a list with a matrix of gametic relatedness coefficients (genome) and
  a matrix of individual relatedness coefficients (indiv)
}
\description{
Level 0 function that returns Genomic Relatedness Matrix (GRM)
  for honeybees from Identical By Descent genomic data (tracked alleles
  since the founders) - see references on the background theory.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(TRUE)
SP$setTrackPed(isTrackPed = TRUE)
SP$addTraitA(10)
SP$addSnpChip(5)

basePop <- createVirginQueens(founderGenomes)
drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)
apiary <- createMultiColony(basePop[2:3], n = 2)
apiary <- cross(x = apiary, fathers = fatherGroups[c(2, 3)])
apiary <- buildUp(x = apiary)
apiary <- addVirginQueens(x = apiary, nInd = 5)

haploQ <- getQueenIbdHaplo(apiary[[1]])
haploF <- getFathersIbdHaplo(apiary[[1]])
haploW <- getWorkersIbdHaplo(apiary[[1]])
haploD <- getDronesIbdHaplo(apiary[[1]])
haploV <- getVirginQueensIbdHaplo(apiary[[1]])
SP$pedigree

haplo <- rbind(haploQ, haploF, haploW, haploD, haploV)

GRMs <- calcBeeGRMIbd(x = haplo)

# library("Matrix"); image(as(GRMs$genome, "Matrix"))
# library("Matrix"); image(as(GRMs$indiv, "Matrix"))

GRMs

x <- diag(GRMs$genome)
hist(x)
summary(x)

x <- GRMs$genome[lower.tri(x = GRMs$genome, diag = FALSE)]
hist(x)
summary(x)

x <- diag(GRMs$indiv)
hist(x)
summary(x)

x <- GRMs$indiv[lower.tri(x = GRMs$indiv, diag = FALSE)]
hist(x)
summary(x)

qI <- getQueen(apiary[[1]])@id
fI <- sort(getFathers(apiary[[1]])@id)
wI <- sort(getWorkers(apiary[[1]])@id)
dI <- sort(getDrones(apiary[[1]])@id)
vI <- sort(getVirginQueens(apiary[[1]])@id)

qG <- c(t(outer(X = qI, Y = 1:2, FUN = paste, sep = "_")))
fG <- paste(fI, 1, sep = "_")
wG <- c(t(outer(X = wI, Y = 1:2, FUN = paste, sep = "_")))
dG <- paste(dI, 1, sep = "_")
vG <- c(t(outer(X = vI, Y = 1:2, FUN = paste, sep = "_")))

# Queen vs others
GRMs$genome[qG, qG]
GRMs$indiv[qI, qI]

GRMs$genome[fG, qG]
GRMs$indiv[fI, qI]

GRMs$genome[wG, qG]
GRMs$indiv[wI, qI]

GRMs$genome[dG, qG]
GRMs$indiv[dI, qI]

GRMs$genome[vG, qG]
GRMs$indiv[vI, qI]

# Fathers vs others
GRMs$genome[fG, fG]
GRMs$indiv[fI, fI]

GRMs$genome[wG, fG]
GRMs$indiv[wI, fI]

GRMs$genome[dG, fG]
GRMs$indiv[dI, fI]

GRMs$genome[vG, fG]
GRMs$indiv[vI, fI]

# Workers vs others
GRMs$genome[wG, wG]
GRMs$indiv[wI, wI]

GRMs$genome[dG, wG]
GRMs$indiv[dI, wI]

GRMs$genome[vG, wG]
GRMs$indiv[vI, wI]
}
\references{
Grossman and Eisen (1989) Inbreeding, coancestry, and covariance between
  relatives for X-chromosomal loci. The Journal of Heredity,
  \url{https://doi.org/10.1093/oxfordjournals.jhered.a110812}

Fernando and Grossman (1989) Covariance between relatives for X-chromosomal
  loci in a population in disequilibrium. Theoretical and Applied Genetics,
  \url{https://doi.org/10.1007/bf00305821}

Fernando and Grossman (1990) Genetic evaluation with autosomal
  and X-chromosomal inheritance. Theoretical and Applied Genetics,
  \url{https://doi.org/10.1007/bf00224018}

Van Arendonk, Tier, and Kinghorn (1994) Use of multiple genetic markers in
  prediction of breeding values. Genetics,
  \url{https://doi.org/10.1093/genetics/137.1.319}

Hill and Weir (2011) Variation in actual relationship as a consequence of
  Mendelian sampling and linkage. Genetics Research,
  \url{https://doi.org/10.1017/s0016672310000480}
}
