% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{calcBeeGRMIbs}
\alias{calcBeeGRMIbs}
\alias{calcBeeAlleleFreq}
\title{Calculate Genomic Relatedness Matrix (GRM) for honeybees from
  Identical By State genomic data}
\usage{
calcBeeGRMIbs(x, sex, alleleFreq = NULL)

calcBeeAlleleFreq(x, sex)
}
\arguments{
\item{x}{\code{\link{matrix}} of genotypes represented as allele dosage coded
as 0, 1, or 2 in females (queens or workers) and as 0 or 1 in males
(fathers or drones); individuals are in rows and sites are in columns; no
missing values are allowed (this is not checked - you will get NAs!)}

\item{sex}{character vector denoting sex for individuals with genotypes in
\code{x} - \code{"F"} for female and \code{"M"} for male}

\item{alleleFreq}{numeric, vector of allele frequencies for the sites in
\code{x}; if \code{NULL}, then \code{\link{calcBeeAlleleFreq}} is used}
}
\value{
matrix of genomic relatedness coefficients
}
\description{
Level 0 function that returns Genomic Relatedness Matrix (GRM)
  for honeybees from Identical By State genomic data (bi-allelic SNP
  represented as allele dosages) following the method for the sex X
  chromosome (Druet and Legarra, 2020)
}
\section{Functions}{
\itemize{
\item \code{calcBeeAlleleFreq}: Calculate allele frequencies from honeybee genotypes
}}

\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(TRUE)
SP$setTrackPed(isTrackPed = TRUE)
SP$addTraitA(10)
SP$addSnpChip(5)

basePop <- createVirginQueens(founderGenomes)
drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)
apiary <- createMultiColony(basePop[2:3], n = 2)
apiary <- cross(x = apiary, fathers = fatherGroups[c(2, 3)])
apiary <- buildUp(x = apiary)
apiary <- addVirginQueens(x = apiary, nInd = 5)

genoQ <- getQueensSegSiteGeno(apiary[[1]])
genoF <- getFathersSegSiteGeno(apiary[[1]])
genoW <- getWorkersSegSiteGeno(apiary[[1]])
genoD <- getDronesSegSiteGeno(apiary[[1]])
genoV <- getVirginQueensSegSiteGeno(apiary[[1]])
genoMeanW <- apply(X = genoW, MARGIN = 2, FUN = mean)
genoMeanD <- apply(X = genoD, MARGIN = 2, FUN = mean)

geno <- rbind(genoQ, genoF, genoW, genoD, genoV, genoMeanW, genoMeanD)
n <- length(rownames(geno))
rownames(geno)[c(n - 1, n)] <- c("mw", "md")

sex <- getCasteSex(x = apiary[[1]])
sex <- c(
  sex$queen, sex$fathers, sex$workers, sex$drones, sex$virginQueens,
  "F", "M"
)

GRM <- calcBeeGRMIbs(x = geno, sex = sex)

# library("Matrix"); image(as(GRM, "Matrix"))

x <- diag(GRM)
hist(x)
summary(x)

x <- GRM[lower.tri(x = GRM, diag = FALSE)]
hist(x)
summary(x)

q <- rownames(genoQ)
f <- rownames(genoF)
w <- rownames(genoW)
d <- rownames(genoD)
v <- rownames(genoV)
mw <- "mw"
md <- "md"

# Queen vs others
GRM[q, f]
GRM[q, w]
GRM[q, d]
GRM[q, v]
GRM[q, mw]
GRM[q, md]

# Fathers vs others
GRM[f, f]
GRM[f, w]
GRM[f, d]
GRM[f, v]
GRM[f, mw]
GRM[f, md]

# Workers vs others
GRM[w, w]
GRM[w, d]
GRM[w, v]
GRM[w, mw]
GRM[w, md]

# Calculating allele frequencies ourselves (say, to "shift" base population)
aF <- calcBeeAlleleFreq(x = geno, sex = sex)
hist(aF)
GRM2 <- calcBeeGRMIbs(x = geno, sex = sex, alleleFreq = aF)
stopifnot(identical(GRM2, GRM))
}
\references{
Druet and Legarra (2020) Theoretical and empirical comparisons of
  expected and realized relationships for the X-chromosome. Genetics
  Selection Evolution, 52:50 \url{https://doi.org/10.1186/s12711-020-00570-6}
}
