% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{calcQueensPHomBrood}
\alias{calcQueensPHomBrood}
\title{Calculate expected percentage of csd homozygous brood of a queen}
\usage{
calcQueensPHomBrood(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}
}
\value{
numeric, expected csd homozygosity named by colony id when \code{x}
  is \code{\link{Colonies-class}}
}
\description{
Level 0 function that calculates the expected percentage of csd
  homozygous brood of a queen. The percentage is computed based on the csd
  alleles of the queen and the drones the queen mated with (the fathers).
}
\details{
TODO: describe queen's and colony's pHomBrood / nHomBrood
        https://github.com/HighlanderLab/SIMplyBee/issues/80
        https://github.com/HighlanderLab/SIMplyBee/issues/104
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 10)
colony2 <- addDrones(colony2, nInd = 20)

# Mated queen
calcQueensPHomBrood(colony1@queen)

# Colony
calcQueensPHomBrood(colony1)

# Colonies
apiary <- c(colony1, colony2)
calcQueensPHomBrood(apiary)

founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 20)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 10)
colony2 <- addDrones(colony2, nInd = 20)
colony1 <- addVirginQueens(colony1, nInd = 1)

# Virgin queen
calcQueensPHomBrood(basePop[2])

# Mated queen
matedQueen <- crossVirginQueen(pop = basePop[4], drones = drones[11:15], nFathers = 5)
pHomBrood(matedQueen)

# Queen of the colony
pHomBrood(getQueen(colony1))

# Colony
pHomBrood(colony1)

# Colonies
apiary <- c(colony1, colony2)
pHomBrood(apiary)

# Now inbreed virgin queen with her brothers to get csd homozygous brood
colony3 <- createColony(x = getVirginQueens(colony1))
colony3 <- crossColony(colony3, drones = getDrones(colony1), nFathers = 10)
colony3 <- addWorkers(colony3, nInd = 100)
calcQueensPHomBrood(colony3)
}
