% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{caste}
\alias{caste}
\title{Report caste of an individual}
\usage{
caste(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}}
}
\value{
character with entries \code{"queen"}, \code{"workers"},
  \code{"drones"}, and \code{"virginQueens"}; if you get \code{NA} note that
  this is not supposed to happen
}
\description{
Level 0 function that reports caste of an individual
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony <- addWorkers(colony)
colony <- addDrones(colony)
colony <- addVirginQueens(colony)

caste(getQueen(colony))
caste(getWorkers(colony, nInd = 2))
caste(getDrones(colony, nInd = 2))
caste(getVirginQueens(colony, nInd = 2))

bees <- c(getQueen(colony),
          getWorkers(colony, nInd = 2),
          getDrones(colony, nInd = 2),
          getVirginQueens(colony, nInd = 2))
caste(bees)

}
