% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{collapseColonies}
\alias{collapseColonies}
\title{Collapse colony for all given colonies}
\usage{
collapseColonies(colonies)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}
}
\value{
\code{\link{Colonies-class}} with the collapse event set to
  \code{TRUE}
}
\description{
Level 3 function that does the same as
  \code{\link{collapseColony}} but for all given colonies.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

founderDrones <- createDrones(x = basePop[1], nInd = 30)
colony1 <- createColony(queen = basePop[2], fathers = founderDrones[ 1:10])
colony2 <- createColony(queen = basePop[3], fathers = founderDrones[11:20])
colony3 <- createColony(queen = basePop[4], fathers = founderDrones[21:30])
apiary <- c(colony1, colony2, colony3)
apiary
hasCollapsed(apiary)

tmp <- pullColonies(apiary, ID = c("2", "3"))
tmp
apiaryLost <- collapseColonies(tmp$pulledColonies)
hasCollapsed(apiaryLost)
apiaryLeft <- tmp$remainingColonies
hasCollapsed(apiaryLeft)

}
