% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{collapseColonies}
\alias{collapseColonies}
\title{Collapse colony for all given colonies}
\usage{
collapseColonies(colonies)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}
}
\value{
\code{\link{Colonies-class}} with the collapse event set to
  \code{TRUE}
}
\description{
Level 3 function that does the same as
  \code{\link{collapseColony}} but for all given colonies.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

founderDrones <- createDrones(x = basePop[1], nInd = 30)
drones <- createDrones(x = basePop[1:4], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[5:10], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[11:20], nFathers = 5)
colony3 <- createColony(x = basePop[4])
colony3 <- crossColony(colony3, drones = drones[21:30], nFathers = 5)
apiary <- c(colony1, colony2, colony3)
apiary
hasCollapsed(apiary)

tmp <- pullColonies(apiary, ID = c("2", "3"))
tmp
apiaryLost <- collapseColonies(tmp$pulledColonies)
hasCollapsed(apiaryLost)
apiaryLeft <- tmp$remainingColonies
hasCollapsed(apiaryLeft)
}
