% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{collapseColony}
\alias{collapseColony}
\title{Replicates colony collapse}
\usage{
\method{collapseColony}(colony)
}
\arguments{
\item{colony}{Colony class. AlphaSimRBee Colony object from the \code{createColony(...)} call}
}
\value{
Single AlphaSim population object of collapsed colony
}
\description{
Replicates the collapse of a colony. This can be due to winter losses, disease or other factors.
}
\examples{
inst/examples/examples_collapseColony.R
}
\seealso{
\code{\link[??????]{collapseColony}}
}
