% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Colonies.R
\docType{class}
\name{Colonies-class}
\alias{Colonies-class}
\alias{show,Colonies-method}
\alias{[,Colonies,integerOrNumericOrLogical-method}
\alias{[,Colonies,character-method}
\alias{[[,Colonies,integerOrNumericOrLogical-method}
\alias{[[,Colonies,character-method}
\alias{c,Colonies-method}
\alias{c,ColoniesOrNULL-method}
\alias{isColonies}
\title{Honeybee colonies}
\usage{
\S4method{show}{Colonies}(object)

\S4method{[}{Colonies,integerOrNumericOrLogical}(x, i, j, drop)

\S4method{[}{Colonies,character}(x, i, j, drop)

\S4method{[[}{Colonies,integerOrNumericOrLogical}(x, i)

\S4method{[[}{Colonies,character}(x, i)

\S4method{c}{Colonies}(x, ...)

\S4method{c}{ColoniesOrNULL}(x, ...)

isColonies(x)
}
\arguments{
\item{object}{\code{\link{Colonies-class}}}

\item{x}{\code{\link{Colonies-class}}}

\item{i}{integer, numeric, logical, or character, index or ID to select
a colony (see examples)}

\item{j}{not used}

\item{drop}{not used}

\item{...}{\code{NULL}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}
}
\value{
\code{\link{Colonies-class}} or \code{\link{Colony-class}}
}
\description{
An object holding a collection of honeybee colonies. It behaves
  like a list.
}
\section{Functions}{
\itemize{
\item \code{show,Colonies-method}: Show colonies object

\item \code{[,Colonies,integerOrNumericOrLogical-method}: Extract a colony (one or more!) by integer/numeric/logical index (return \code{\link{Colonies-class}})

\item \code{[,Colonies,character-method}: Extract a colony (one or more!) by character ID (return \code{\link{Colonies-class}})

\item \code{[[,Colonies,integerOrNumericOrLogical-method}: Extract a colony (just one!) by integer/numeric index (return \code{\link{Colony-class}})

\item \code{[[,Colonies,character-method}: Extract a colony (just one!) by character ID (return \code{\link{Colony-class}})

\item \code{c,Colonies-method}: Combine multiple colony and colonies objects

\item \code{c,ColoniesOrNULL-method}: Combine multiple colony and colonies objects

\item \code{isColonies}: Test if x is a Colonies class object
}}

\section{Slots}{

\describe{
\item{\code{colonies}}{list, a collection of \code{\link{Colony-class}} objects}
}}

\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 15)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony3 <- createColony(queen = basePop[4], fathers = drones[11:15])
apiary <- c(colony1, colony2, colony3)
apiary
show(apiary)
is(apiary)
isColonies(apiary)

getId(apiary)
apiary[1]
getId(apiary[1])
getId(apiary["2"])
getId(apiary[2])

getId(apiary)
getId(apiary[c(1, 3)])
getId(apiary[c("2", "4")])
getId(apiary[c(TRUE, FALSE, TRUE)])

apiary[[1]]
apiary[["2"]]
apiary[[3]]
apiary[["4"]]

getId(c(apiary[c(1, 3)], apiary[2]))
getId(c(apiary[2], apiary[c(1, 3)]))

getId(c(apiary[2], apiary[0]))
getId(c(apiary[0], apiary[2]))

getId(c(apiary[2], NULL))
getId(c(NULL, apiary[2]))

}
\seealso{
\code{\link{createColonies}}
}
