% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Colonies-class}
\alias{Colonies-class}
\alias{[,Colonies-method}
\alias{[[,Colonies-method}
\alias{c,Colonies-method}
\title{Colonies
#'}
\usage{
\S4method{[}{Colonies}(x, i)

\S4method{[[}{Colonies}(x, i)

\S4method{c}{Colonies}(x, ...)
}
\arguments{
\item{x}{a 'Colonies' object}

\item{i}{index of populations or colonies}
}
\description{
The Colonies represents the list of the colonies.
It is designed to behave like a list of colonies.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: Extract Colonies by index

\item \code{[[}: Extract Colony by index

\item \code{c}: Extract Colony by index
}}

\section{Slots}{

\describe{
\item{\code{colonies}}{of \code{\link{Colony-class}} and/or
\code{Colonies-class}}
}}

