% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{PopOrNULL-class}
\alias{PopOrNULL-class}
\title{Creates a honeybee colony}
\usage{
\method{createColony}(id, location, queen, drones, workers, virgin_queen, fathers, pheno, last_event)
}
\value{
Returns a \code{Colony} object
}
\description{
List. Creates a honeybee colony as a list with the following elements:
  \id location, queen, drones, workers, virgin_queens, pheno, fathers, and last event
  \All elements of the list, expect for \code{last_even}, are assumed NULL if not specified
  \otherwise. \code{last_event} is set to "new_colony".
}
\section{Slots}{

\describe{
\item{\code{id}}{ID of the colony.}

\item{\code{location}}{Numeric, location of the colony (x, y).}

\item{\code{queen}}{AlphaSimR population object to become the queen of the colony (we use its misc slot for queen's age).}

\item{\code{drones}}{AlphaSimR population object to become the drones of the colony.}

\item{\code{workers}}{AlphaSimR population object to become the workers of the colony.}

\item{\code{virgin_queens}}{AlphaSimR individual or population object to become the virgin queen(s) of the colony.}

\item{\code{pheno}}{A matrix of the phenotypes of the colony}

\item{\code{swarm}}{Logical, whether the colony has swarmed}

\item{\code{split}}{Logical, whether the colony has split}

\item{\code{supersedure}}{Logical, whether the colony has superseded}

\item{\code{collapse}}{Logical, whether the colony has collapsed}

\item{\code{production}}{Logical, whether the colony produces hive products}

\item{\code{last_event}}{Character, the last event of the colony #TODO: WE probably don't need this}

\item{\code{misc}}{A list, normally empty and exists solely as an open slot available for uses to store extra information about individuals.}
}}

