% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{createColonies}
\alias{createColonies}
\title{Create colonies}
\usage{
createColonies(x = NULL, n = NULL, location = NULL)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, virgin queens or queens for the colonies
(selected at random if there are more than \code{n} in \code{Pop})}

\item{n}{integer, number of colonies to create (if only \code{nCol} is
given then \code{\link{Colonies-class}} is created with \code{nCol} empty
(\code{NULL}) individual colony) - this is mostly useful for programming)}

\item{location}{list, location of the colonies as \code{c(x, y)}}
}
\value{
\code{\link{Colonies-class}}
}
\description{
Level 3 function that creates a set of colonies. Usually to
  start a simulation.
}
\details{
When both \code{x} and \code{nCol} are \code{NULL}, then an empty
  \code{NULL} \code{\link{Colonies-class}} is created with 0 colonies.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

# Create 2 empty (NULL) colonies
apiary <- createColonies(n = 2)
apiary
apiary[[1]]
apiary[[2]]

# Create 2 virgin colonies
apiary <- createColonies(x = basePop, n = 2)
apiary
apiary[[1]]
apiary[[2]]

# Create mated colonies by crossing
apiary <- createColonies(x = basePop[1:2], n = 2)
drones <- createDrones(x = basePop[3], n = 30)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 2, nFathers = 15)
apiary <- cross(apiary, fathers = fatherGroups)
apiary
apiary[[1]]
apiary[[2]]
}
