% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{createColony}
\alias{createColony}
\title{Create a new Colony}
\usage{
createColony(
  location = NULL,
  queen = NULL,
  yearOfBirth = NULL,
  fathers = NULL,
  virgin_queens = NULL,
  simParamBee = NULL
)
}
\arguments{
\item{location}{numeric, location of the colony as \code{c(x, y)}}

\item{queen}{\code{\link{Pop-class}} with one individual that will be the
queen of the colony; error is thrown if more than one individual is given}

\item{yearOfBirth}{numeric, year of birth of the queen}

\item{fathers}{\code{\link{Pop-class}} with drones that the queen will mate
with as part of this function (if she is already mated, a warning is given
and the fathers argument is ignored)}

\item{virgin_queens}{\code{\link{Pop-class}} with one or more individuals of
which one will become the queen of the colony in the future
TODO: think and explain what happens if we provide both a queen and virgin
queens (possibly of different origin)!}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
new \code{\link{Colony-class}}
}
\description{
Level 2 function that creates a new \code{\link{Colony-class}}
  to initiate simulations.
}
\details{
When a mated queen is given or a queen alongside with fathers, the
  function also generates one virgin queen.
  TODO: discuss this - see below
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones)
colony1

colony2 <- createColony(virgin_queens = basePop[3])
colony2

}
