% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createDrones}
\alias{createDrones}
\title{Creates drones from the colony}
\usage{
createDrones(colony, nInd)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{nInd}{integer, the number of drones to create}
}
\value{
\code{\link{Pop-class}} with drones
}
\description{
Level 1 function that creates the specified number of drones
  from the colony. Currently this is done by creating doubled-haploids from a
  queen (generating recombinant gametes and doubling them).
  TODO The drones will eventually be made properly haploid!
       Follow https://github.com/HighlanderLab/SIMplyBee/issues/24
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
createWorkers(colony, nInd = 10)
createDrones(colony, nInd = 2)

}
