% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createVirginQueens}
\alias{createVirginQueens}
\title{Creates virgin queens from the colony}
\usage{
createVirginQueens(colony, nInd = NULL, simParamBee = NULL)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{nInd}{numeric or function, number of virgin queens; if \code{NULL} then
\code{simParamBee$nVirginQueens} is used}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
list with two nodes named \code{virgin_queens} (a
  \code{\link{Pop-class}}) and \code{pHomBrood} (a numeric)
}
\description{
Level 1 function that creates the specified number of virgin
  queens from the colony by mating the colony queen and the fathers. If csd
  locus is defined, it takes it into account and any csd homozygotes are
  removed and counted towards homozygous brood.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
createVirginQueens(colony, nInd = 10)

# Using a default in SP$nVirginQueens
# (just to have some virgin queens - change this to your needs!)
createVirginQueens(colony)
createVirginQueens(colony)

SP$nVirginQueens <- 15
createVirginQueens(colony)
createVirginQueens(colony)

nVirginQueensFun <- function(colony) { rpois(n = 1, lambda = 15) }
createVirginQueens(colony, nInd = nVirginQueensFun)
createVirginQueens(colony, nInd = nVirginQueensFun)

SP$nVirginQueens <- nVirginQueensFun
createVirginQueens(colony)
createVirginQueens(colony)

}
