% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createVirginQueens}
\alias{createVirginQueens}
\title{Creates virgin queens from the colony}
\usage{
createVirginQueens(x, nInd = NULL, year = NULL, simParamBee = NULL)
}
\arguments{
\item{x}{\code{link{MapPop-class}} or \code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{nInd}{numeric or function, number of virgin queens; if \code{NULL} then
\code{simParamBee$nVirginQueens} is used}

\item{year}{numeric, year of birth for virgin queens}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
when \code{x} is \code{link{MapPop-class}} returns
  \code{virginQueens} (a \code{\link{Pop-class}});
  when \code{x} is \code{\link{Colony-class}} returns
  \code{virginQueens} (a \code{\link{Pop-class}});
  when \code{x} is \code{\link{Colonies-class}}
  return is a named list of \code{virginQueens} (a \code{\link{Pop-class}});
  named by colony ID
}
\description{
Level 1 function that creates the specified number of virgin
  queens from the colony by mating the colony queen and the fathers. If csd
  locus is defined, it takes it into account and any csd homozygotes are
  removed and counted towards homozygous brood.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackPed(isTrackPed = TRUE)
SP$setTrackRec(isTrackRec = TRUE)
# Create virgin queens from the founder population (Map-Pop object)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)

(virginQueens <- createVirginQueens(colony1, nInd = 10))
colony1@queen@id
virginQueens@id
virginQueens@sex
virginQueens@misc
virginQueens@mother
virginQueens@father
SP$pedigree
SP$recHist
SP$recHist[[23]][[1]][1]
SP$recHist[[23]][[1]][2]
SP$caste

createVirginQueens(apiary, nInd = 10)

# Using a default in SP$nVirginQueens
# (just to have some virgin queens - change this to your needs!)
createVirginQueens(colony1)
createVirginQueens(apiary)

# Specify own number
SP$nVirginQueens <- 15
createVirginQueens(colony1)
createVirginQueens(apiary)

# Specify a function that will give a number
nVirginQueensFun <- function(colony) {
  rpois(n = 1, lambda = 15)
}
createVirginQueens(colony1, nInd = nVirginQueensFun)
createVirginQueens(apiary, nInd = nVirginQueensFun)

# Store a function or a value in the SP object
SP$nVirginQueens <- nVirginQueensFun
createVirginQueens(colony1)
createVirginQueens(apiary)
}
