% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createWorkers}
\alias{createWorkers}
\alias{createVirginQueens}
\title{Creates workers from the colony}
\usage{
createWorkers(colony, nInd, simParamBee = NULL)

createVirginQueens(colony, nInd, simParamBee = NULL)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{nInd}{integer, number of workers to create}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
list with two nodes named \code{workers} (a \code{\link{Pop-class}})
  and \code{nHomDrones} (an integer)
}
\description{
Level 1 function that creates the specified number of workers
  from the colony by mating the colony queen and the fathers. If csd locus is
  defined, it takes it into account and any csd homozygotes are removed and
  counted as homozygous drones.
}
\section{Functions}{
\itemize{
\item \code{createVirginQueens}: Create virgin queens
}}

\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
createWorkers(colony, nInd = 10)

}
