% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{crossColony}
\alias{crossColony}
\title{Cross (mate) a virgin queen of a colony to a group of drones}
\usage{
crossColony(colony, fathers, simParamBee = NULL)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{fathers}{\code{\link{Pop-class}}, drones}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colony-class}} with a mated queen
}
\description{
Level 2 function that crosses (mates) a virgin queen of colony
  to a group of drones. When there are multiple virgin queens in the colony,
  one is selected at random, mated, and promoted to the queen of the colony.
  Other virgin queens are destroyed. Mated drones (fathers) are stored for
  producing progeny at a later stage.
  TODO: is this correct (about destroy)?
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createDrones(x = basePop[1], nDronesPerQueen = 5)

colony <- createColony(virginQueen = basePop[2])
colony
colony <- crossColony(colony, fathers = drones)
colony

}
\seealso{
\code{\link{Colony-class}} on how we store the fathers along the
  queen.
}
