% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{crossVirginQueen}
\alias{crossVirginQueen}
\title{Cross (mate) a virgin queen to a group drones}
\usage{
crossVirginQueen(pop, drones, nFathers = NULL, simParamBee = NULL)
}
\arguments{
\item{pop}{\code{\link{Pop-class}}, one or more virgin queens to be mated;
\code{\link{isVirginQueen}} test will be run on these individuals}

\item{drones}{\code{\link{Pop-class}}, a group of drones that could be mated
with virgin queen(s); if there is more than one virgin queen, then the
\code{fathers} are partitioned into multiple groups of \code{nFathers} size
 using \code{\link{pullDroneGroupsFromDCA}}}

\item{nFathers}{numeric, average number of drones (fathers) used in mating
the virgin queen(s); if \code{NULL} then \code{simParamBee$nFathers} is used}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Pop-class}} with mated queen(s). The misc slot of the queens
contains additional information about the number of workers, drones and homozygous
brood produced and the theoretical percentage of homozygous brood.
}
\description{
Level 1 function that crosses (mates) a virgin queen to a group
  of drones. This function does not create any progeny, it only stores the
  mated drones (fathers) so we can later create progeny as needed.
}
\examples{
founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 20)

virginQueen1 <- basePop[2]
(matedQueen1 <- crossVirginQueen(
  pop = virginQueen1,
  drones = drones[1:5],
  nFathers = 5
))
isQueenMated(virginQueen1)
isQueenMated(matedQueen1)
nFathers(matedQueen1)
getFathers(matedQueen1)@id

virginQueen2 <- basePop[3]
(matedQueen2 <- crossVirginQueen(
  pop = virginQueen2,
  drones = drones[6:10],
  nFathers = 5
))
isQueenMated(virginQueen2)
isQueenMated(matedQueen2)
nFathers(matedQueen2)
getFathers(matedQueen2)@id

matedQueens <- crossVirginQueen(
  pop = c(basePop[4], basePop[5]),
  drones = drones[11:15], nFathers = 2
)
matedQueens
isQueenMated(matedQueens)
nFathers(matedQueens)
getFathers(matedQueens)

# Inbred mated queen (mated with her own sons)
matedQueen3 <- crossVirginQueen(
  pop = basePop[1],
  drones = drones[16:20],
  nFathers = 5
)
# Check the theoretical homozygosity
matedQueen3@misc[[1]]$pHomBrood
}
\seealso{
\code{\link{Colony-class}} on how we store the fathers along the
  queen.
}
