% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{downsizeColonies}
\alias{downsizeColonies}
\title{Reduce number of workers and remove all drones and virgin queens from
  colonies}
\usage{
downsizeColonies(
  colonies,
  p = NULL,
  use = "rand",
  new = FALSE,
  simParamBee = NULL
)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{p}{numeric, proportion of workers to be removed from each colony; if
\code{NULL} then \code{\link{SimParamBee}$downsizeP} is used}

\item{use}{character, all the options provided by \code{\link{selectInd}};
it guides the selection of workers that will be removed}

\item{new}{logical, should we remove all current workers and add a targeted
proportion anew (say, create winter workers)}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colonies-class}} with workers reduced and drones/virgin queens removed
}
\description{
Level 3 function that downsizes colonies by removing a proportion
  of workers, all drones and all virgin queens. Usually in the autumn, such
  an event occurs in preparation for the winter months.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

apiary <- createColonies(x = basePop[1:2], n = 2)
drones <- createDrones(x = basePop[3:4], nInd = 20)
apiary <- crossColonies(apiary, drones = drones, nFathers = 5)
apiary <- buildUpColonies(apiary)
apiary <- addVirginQueens(x = apiary, nInd = 20)
apiary[[1]]
apiary[[2]]

apiary <- downsizeColonies(colonies = apiary, use = "rand", new = TRUE)
apiary[[1]]
apiary[[2]]
}
