% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{downsizeColonies}
\alias{downsizeColonies}
\title{Reduce number of workers and remove all drones and virgin queens from colonies}
\usage{
downsizeColonies(colonies, p = 0.85, use = "rand")
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{p}{numeric, percentage of workers to remove from the colonies}

\item{use}{character, all the options provided by \code{\link{selectInd}};
it guides the selection of workers that will be removed}
}
\value{
\code{\link{Colonies-class}} with workers reduced and drones/virgin queens removed
}
\description{
Level 3 function that downsizes colonies by removing a percentage of
  workers, all drones and all virgin queens. Usually in the autumn, such an event occurs
  in preparation for the winter months.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

apiary <- createColonies(pop = basePop, n = 2)
apiary <- buildUpColonies(apiary)
apiary <- addVirginQueens(x = apiary, nInd = 20)
apiary[[1]]
apiary[[2]]

apiary <- downsizeColonies(colonies = apiary)
apiary[[1]]
apiary[[2]]

# TODO: FIND REFERENCES for defaults
#   https://github.com/HighlanderLab/SIMplyBee/issues/197
}
