% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{downsizeColony}
\alias{downsizeColony}
\title{Reduce number of workers and remove all drones and virgin queens from hive}
\usage{
downsizeColony(colony, p = 0.85, use = "rand")
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{p}{numeric, percentage of workers to remove from the colony}

\item{use}{character, all the options provided by \code{\link{selectInd}};
it guides the selection of workers that will be removed}
}
\value{
\code{\link{Colony-class}} with workers reduced and drones/virgin queens removed
}
\description{
Level 2 function that downsizes colony by removing a percentage of
  workers, all drones and all virgin queens.  Usually in the autumn, such an event occurs
  in preparation for the winter months.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)
colony <- buildUpColony(colony)
colony <- addVirginQueens(x = colony, nInd = 10)
colony
colony <- downsizeColony(colony = colony)
colony

# TODO: FIND REFERENCES for defaults
#   https://github.com/HighlanderLab/SIMplyBee/issues/197
}
