% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{editCsdLocus}
\alias{editCsdLocus}
\title{Edit the csd locus}
\usage{
editCsdLocus(pop, alleles = NULL, simParamBee = NULL)
}
\arguments{
\item{pop}{an object of \code{\link{Pop-class}}}

\item{alleles}{list or \code{NULL}. If not null, the user has to provide a list of
matrices or data frames with two rows and n columns, where n is the length of the csd
as specific in \code{\link{SimParamBee}}, each representing a csd haplotype for each
individual. The two haplotypes must not be the same, as the function does not allow
to edit the csd to a homozygous state. If the parameter is \code{NULL},
the function samples a heterozygous csd genotype for each individual from
all possible csd alleles.}

\item{simParamBee}{an object of \code{\link{SimParamBee}}}
}
\value{
Returns an object of \code{\link{Pop-class}}
}
\description{
Edits the csd locus in an entire population of individuals. The user
can provide a list of csd alleles for each individual or, alternatively,
the function samples a heterozygous genotype for each individual
from all possible csd alleles. The gv slot is recalculated to
reflect the any changes due to editing, but other slots remain the same.
}
\examples{
founderGenomes <- quickHaplo(nInd = 100, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes, csdChr = 1, nCsdAlleles = 8)
basePop <- createVirginQueens(founderGenomes, editCsd = FALSE)
nrow(getCsdAlleles(basePop, unique = TRUE))
all(isCsdHeterozygous(basePop))

basePopEdited <- SIMplyBee:::editCsdLocus(basePop)
nrow(getCsdAlleles(basePopEdited, unique = TRUE))
all(isCsdHeterozygous(basePopEdited))
}
