% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCaste}
\alias{getCaste}
\title{Report caste of an individual}
\usage{
getCaste(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}
}
\value{
When x is \code{\link{Pop-class}}, character of caste status; if you
  get \code{NA} note that this is not supposed to happen. When x is
  \code{\link{Colony-class}}, list with character vectors (list is named with
  caste). When x is \code{\link{Colonies-class}}, list of lists with
  character vectors (list is named with colony id).
}
\description{
Level 0 function that reports caste of an individual
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[1:5])
colony1 <- addWorkers(colony1)
colony2 <- addWorkers(colony2)
colony1 <- addDrones(colony1)
colony2 <- addDrones(colony2)
colony1 <- addVirginQueens(colony1)
apiary <- c(colony1, colony2)

getCaste(getQueen(colony1))
getCaste(getFathers(colony1))
getCaste(getWorkers(colony1))
getCaste(getDrones(colony1))
getCaste(getVirginQueens(colony1))

bees <- c(getQueen(colony1),
          getFathers(colony1, nInd = 2),
          getWorkers(colony1, nInd = 2),
          getDrones(colony1, nInd = 2),
          getVirginQueens(colony1, nInd = 2))
getCaste(bees)

# Create a data.frame with id, colony, and caste information
(tmpC <- getCaste(colony1))
(tmpI <- getCasteId(colony1))
tmp <- data.frame(caste = unlist(tmpC), id = unlist(tmpI))
head(tmp);tail(tmp)

(tmpC <- getCaste(apiary))
(tmpI <- getCasteId(apiary))
(tmp <- data.frame(caste = unlist(tmpC), id = unlist(tmpI)))
tmp$colony <- sapply(X = strsplit(x = rownames(tmp), split = ".",
                                  fixed = TRUE),
                     FUN = function(z) z[[1]])
head(tmp);tail(tmp)

}
\seealso{
\code{\link{getCastePop}} and \code{\link{getCasteId}}
}
