% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteBv}
\alias{getCasteBv}
\title{Access breeding values of individuals in a caste}
\usage{
getCasteBv(x, caste, nInd = NULL, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{caste}{character, "queen", "fathers", "virgin_queens", "workers", or
"drones"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
vector of breeding values when \code{x} is \code{\link{Colony-class}}
  and list of vectors of breeding values when \code{x} is
  \code{\link{Colonies-class}}, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns breeding values of individuals in
  a caste.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$addTraitA(nQtlPerChr = 10)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 2)
colony2 <- addDrones(colony2, nInd = 4)

getCasteBv(colony1, caste = "queen")
getQueensBv(colony1)

getCasteBv(colony1, caste = "fathers")
getCasteBv(colony1, caste = "fathers", nInd = 2)
getCasteBv(colony1, caste = "fathers", nInd = 2)
getFathersBv(colony1)
getFathersBv(colony1, nInd = 2)

getCasteBv(colony1, caste = "virgin_queens")
getVirginQueensBv(colony1)

getCasteBv(colony1, caste = "workers")
getWorkersBv(colony1)

getCasteBv(colony1, caste = "drones")
getDronesBv(colony1)

apiary <- c(colony1, colony2)
getCasteBv(apiary, caste = "queen")
getQueensBv(apiary)

getCasteBv(apiary, caste = "fathers")
getCasteBv(apiary, caste = "fathers", nInd = 2)
getCasteBv(apiary, caste = "fathers", nInd = 2)
getFathersBv(apiary)
getFathersBv(apiary, nInd = 2)

getCasteBv(apiary, caste = "virgin_queens")
getVirginQueensBv(apiary)

getCasteBv(apiary, caste = "workers")
getWorkersBv(apiary)

getCasteBv(apiary, caste = "drones")
getDronesBv(apiary)

}
\seealso{
\code{\link{bv}}
}
