% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteDd}
\alias{getCasteDd}
\alias{getQueensDd}
\alias{getFathersDd}
\alias{getVirginQueensDd}
\alias{getWorkersDd}
\alias{getDronesDd}
\title{Access dominance deviations of individuals in a caste}
\usage{
getCasteDd(x, caste, nInd = NULL, simParamBee = NULL)

getQueensDd(x, simParamBee = NULL)

getFathersDd(x, nInd = NULL, simParamBee = NULL)

getVirginQueensDd(x, nInd = NULL, simParamBee = NULL)

getWorkersDd(x, nInd = NULL, simParamBee = NULL)

getDronesDd(x, nInd = NULL, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{caste}{character, "queen", "fathers", "virgin_queens", "workers", or
"drones"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
vector of dominance deviations when \code{x} is
  \code{\link{Colony-class}} and list of vectors of dominance deviations when
  \code{x} is \code{\link{Colonies-class}}, named by colony id when \code{x}
  is \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns dominance deviations of
  individuals in a caste.
}
\section{Functions}{
\itemize{
\item \code{getQueensDd}: Access dominance deviation of the queen

\item \code{getFathersDd}: Access dominance deviations of fathers

\item \code{getVirginQueensDd}: Access dominance deviations of virgin queens

\item \code{getWorkersDd}: Access dominance deviations of workers

\item \code{getDronesDd}: Access dominance deviations of drones
}}

\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$addTraitA(nQtlPerChr = 10)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 2)
colony2 <- addDrones(colony2, nInd = 4)

getCasteDd(colony1, caste = "queen")
getQueensDd(colony1)

getCasteDd(colony1, caste = "fathers")
getCasteDd(colony1, caste = "fathers", nInd = 2)
getCasteDd(colony1, caste = "fathers", nInd = 2)
getFathersDd(colony1)
getFathersDd(colony1, nInd = 2)

getCasteDd(colony1, caste = "virgin_queens")
getVirginQueensDd(colony1)

getCasteDd(colony1, caste = "workers")
getWorkersDd(colony1)

getCasteDd(colony1, caste = "drones")
getDronesDd(colony1)

apiary <- c(colony1, colony2)
getCasteDd(apiary, caste = "queen")
getQueensDd(apiary)

getCasteDd(apiary, caste = "fathers")
getCasteDd(apiary, caste = "fathers", nInd = 2)
getCasteDd(apiary, caste = "fathers", nInd = 2)
getFathersDd(apiary)
getFathersDd(apiary, nInd = 2)

getCasteDd(apiary, caste = "virgin_queens")
getVirginQueensDd(apiary)

getCasteDd(apiary, caste = "workers")
getWorkersDd(apiary)

getCasteDd(apiary, caste = "drones")
getDronesDd(apiary)

}
\seealso{
\code{\link{dd}}
}
