% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteId}
\alias{getCasteId}
\title{Get IDs of individuals of a caste, or ID of all members of colony}
\usage{
getCasteId(x, caste = "all", simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}

\item{caste}{character, "queen", "fathers", "virginQueens", "workers",
"drones", or "all"}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
when \code{x} is \code{\link{Pop-class}} for \code{caste != "all"}
 or list for \code{caste == "all"} with ID nodes named by caste;
   when \code{x} is \code{\link{Colony-class}} return is a named list of
  \code{\link{Pop-class}} for \code{caste != "all"}
  or named list for \code{caste == "all"} indluding caste members IDs;
   when \code{x} is \code{\link{Colonies-class}} return is a named list of
  \code{\link{Pop-class}} for \code{caste != "all"} or named list of lists of
  \code{\link{Pop-class}} for \code{caste == "all"} indluding caste members IDs
}
\description{
Level 0 function that returns the ID individuals of a caste. To
  get the individuals, use \code{\link{getCastePop}}. To get individuals'
  caste, use \code{\link{getCaste}}.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])

colony1 <- addWorkers(colony1, nInd = 10)
colony1 <- addVirginQueens(colony1, nInd = 4)
colony1 <- addDrones(colony1, nInd = 2)
colony2 <- addWorkers(colony2, nInd = 20)

apiary <- c(colony1, colony2)

getCasteId(x = drones)
getCasteId(x = colony1)
getCasteId(x = colony1, caste = "workers")
getCasteId(x = apiary)
getCasteId(x = apiary, caste = "virginQueens")

# Create a data.frame with id, colony, and caste information
(tmpC <- getCaste(colony1))
(tmpI <- getCasteId(colony1))
tmp <- data.frame(caste = unlist(tmpC), id = unlist(tmpI))
head(tmp)
tail(tmp)

(tmpC <- getCaste(apiary))
(tmpI <- getCasteId(apiary))
(tmp <- data.frame(caste = unlist(tmpC), id = unlist(tmpI)))
tmp$colony <- sapply(
  X = strsplit(
    x = rownames(tmp), split = ".",
    fixed = TRUE
  ),
  FUN = function(z) z[[1]]
)
head(tmp)
tail(tmp)
}
\seealso{
\code{\link{getCaste}}

\code{\link{getCastePop}} and \code{\link{getCaste}}
}
