% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteId}
\alias{getCasteId}
\title{Get IDs of individuals of a caste, or ID of all members of colony}
\usage{
getCasteId(x, caste = "all")
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}} class objects}

\item{caste}{character, "queen", "fathers", "virginQueens", "workers",
"drones", or "all"}
}
\value{
when \code{x} is \code{\link{Pop-class}} for \code{caste != "all"}
 or list for \code{caste == "all"} with ID nodes named by caste;
   when \code{x} is \code{\link{Colony-class}} return is a named list of
  \code{\link{Pop-class}} for \code{caste != "all"}
  or named list for \code{caste == "all"} indluding caste members IDs;
   when \code{x} is \code{\link{Colonies-class}} return is a named list of
  \code{\link{Pop-class}} for \code{caste != "all"} or named list of lists of
  \code{\link{Pop-class}} for \code{caste == "all"} indluding caste members IDs
}
\description{
Level 0 function that returns the ID individuals of a caste. To
  get the individuals, use \code{\link{getCaste}}.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])

colony1 <- addWorkers(colony1, nInd = 10)
colony1 <- addVirginQueens(colony1, nInd = 4)
colony1 <- addDrones(colony1, nInd = 2)
colony2 <- addWorkers(colony2, nInd = 20)

apiary1 = c(colony1, colony2)

getCasteId(x = drones)
getCasteId(x = colony1)
getCasteId(x = colony1, caste = "workers")
getCasteId(x = apiary1)
getCasteId(x = apiary1, caste = "virginQueens")

# Create a data.frame from the lists of ids
tmp <- getCasteId(x = apiary1)
df <- data.frame(unlist(tmp))
head(df)
colnames(df)[1] <- "id"
head(df)
tmp <- strsplit(row.names(df), split = ".", fixed = TRUE)
colony <- sapply(tmp, FUN = function(z) z[1])
caste <- sapply(tmp, FUN = function(z) z[2])
df$colony <- colony
df$caste <- gsub(x = caste, pattern = "[[:digit:]]", replacement = "")
head(df)

}
\seealso{
\code{\link{getCaste}}
}
