% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{getCastePop}
\alias{getCastePop}
\alias{getQueen}
\alias{getFathers}
\alias{getWorkers}
\alias{getDrones}
\alias{getVirginQueens}
\title{Access individuals of a caste}
\usage{
getCastePop(x, caste = "all", nInd = NULL, use = "order", removeFathers = TRUE)

getQueen(x)

getFathers(x, nInd = NULL, use = "rand")

getWorkers(x, nInd = NULL, use = "rand")

getDrones(x, nInd = NULL, use = "rand", removeFathers = TRUE)

getVirginQueens(x, nInd = NULL, use = "rand")
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}},
exceptionally \code{\link{Pop-class}} for calling \code{getFathers}
on a queen population}

\item{caste}{character, "queen", "fathers", "workers", "drones",
"virginQueens", or "all"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed; if there are less individuals than requested,
we return the ones available - this can return \code{NULL}.
If input is \code{\link{MultiColony-class}},
the input could also be a vector of the same length as the number of colonies. If
a single value is provided, the same value will be applied to all the colonies.}

\item{use}{character, all options provided by \code{\link{selectInd}} and
\code{"order"} that selects \code{1:nInd} individuals (meaning it always
returns at least one individual, even if \code{nInd = 0})}

\item{removeFathers}{logical, removes \code{drones} that have already mated;
set to \code{FALSE} if you would like to get drones for mating with multiple
virgin queens, say via insemination}
}
\value{
when \code{x} is \code{\link{Colony-class}} return is
  \code{\link{Pop-class}} for \code{caste != "all"} or list for \code{caste
  == "all"} with nodes named by caste; when \code{x} is
  \code{\link{MultiColony-class}} return is a named list of
  \code{\link{Pop-class}} for \code{caste != "all"} or named list of lists of
  \code{\link{Pop-class}} for \code{caste == "all"}. You can merge
  all the populations in the list with \code{\link{mergePops}} function.
}
\description{
Level 1 function that returns individuals of a caste. These
  individuals stay in the colony (compared to \code{\link{pullCastePop}}).
}
\section{Functions}{
\itemize{
\item \code{getQueen()}: Access the queen

\item \code{getFathers()}: Access fathers (drones the queen mated with)

\item \code{getWorkers()}: Access workers

\item \code{getDrones()}: Access drones

\item \code{getVirginQueens()}: Access virgin queens

}}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
droneGroups <- pullDroneGroupsFromDCA(drones, n = 10, nDrones = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, drones = droneGroups[[1]])
apiary <- createMultiColony(basePop[3:4], n = 2)
apiary <- cross(apiary, drones = droneGroups[c(2, 3)])

# Using defaults in SP$nWorkers & SP$nDrones
colony <- buildUp(x = colony, exact = TRUE)
apiary <- buildUp(x = apiary, exact = TRUE)

# Using a default in SP$nVirginQueens
colony <- addVirginQueens(x = colony)
apiary <- addVirginQueens(x = apiary)
colony
apiary

# Comparison of getCastePop() and getQueen()
getCastePop(colony, caste = "queen")
getQueen(colony)

# Comparison of getCastePop() and getFathers()
getCastePop(colony, caste = "fathers")
getCastePop(colony, caste = "fathers")@id
getCastePop(colony, caste = "fathers", nInd = 2)@id

getFathers(colony)
getFathers(colony)@id
getFathers(colony, nInd = 2)@id

getFathers(getQueen(colony))

# Comparison of getCastePop() and getVirginQueens()
getCastePop(colony, caste = "virginQueens")
getCastePop(colony, caste = "virginQueens")@id
getCastePop(colony, caste = "virginQueens", nInd = 2)@id

getVirginQueens(colony)
getVirginQueens(colony)@id
getVirginQueens(colony, nInd = 2)@id

# Comparison of getCastePop() and getWorkers()
getCastePop(colony, caste = "workers")
getCastePop(colony, caste = "workers")@id
getCastePop(colony, caste = "workers", nInd = 2)@id

getWorkers(colony)
getWorkers(colony)@id
getWorkers(colony, nInd = 2)@id

# Comparison of getCastePop() and getDrones()
getCastePop(colony, caste = "drones")
getCastePop(colony, caste = "drones")@id
getCastePop(colony, caste = "drones", nInd = 2)@id

getDrones(colony)
getDrones(colony)@id
getDrones(colony, nInd = 2)@id

# Comparisons on MultiColony class
getCastePop(apiary, caste = "queen")
getQueen(apiary)
getCastePop(apiary, caste = "queen")[[1]]@id
getCastePop(apiary, caste = "queen")[[2]]@id

getCastePop(apiary, caste = "fathers")
getFathers(apiary)
getFathers(apiary)[[1]]@id
getFathers(apiary)[[2]]@id
getFathers(apiary, nInd = 2)

getCastePop(apiary, caste = "virginQueens")
getVirginQueens(apiary)
getVirginQueens(apiary)[[1]]@id
getVirginQueens(apiary)[[2]]
getVirginQueens(apiary, nInd = 1)
getVirginQueens(apiary, nInd = 2)
mergePops(getVirginQueens(apiary))

getCastePop(apiary, caste = "workers")
# Get unequal number from colonies
getCastePop(apiary, caste = "workers", nInd = c(10, 20))
getWorkers(apiary)
getWorkers(apiary)[[1]]@id
getWorkers(apiary)[[2]]@id
getWorkers(apiary, nInd = 2)
mergePops(getWorkers(apiary))

getCastePop(apiary, caste = "drones")
# Get unequal number from colonies
getCastePop(apiary, caste = "drones", nInd = c(10, 50))
getDrones(apiary)
getDrones(apiary)[[1]]@id
getDrones(apiary)[[2]]
getDrones(apiary, nInd = 2)
mergePops(getDrones(apiary))

getCastePop(apiary)
getCastePop(apiary, caste = "queen")
getCastePop(apiary, caste = "drones")
}
\seealso{
\code{\link{getQueen}}, \code{\link{getFathers}},
  \code{\link{getVirginQueens}}, \code{\link{getWorkers}}, and
  \code{\link{getDrones}}

\code{\link{getCasteId}} and \code{\link{getCaste}}
}
