% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteQtlGeno}
\alias{getCasteQtlGeno}
\alias{getQueensQtlGeno}
\alias{getFathersQtlGeno}
\alias{getVirginQueensQtlGeno}
\alias{getWorkersQtlGeno}
\alias{getDronesQtlGeno}
\title{Access QTL genotypes of individuals in a caste}
\usage{
getCasteQtlGeno(
  x,
  caste,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getQueensQtlGeno(x, trait = 1, chr = NULL, simParamBee = NULL)

getFathersQtlGeno(
  x,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getVirginQueensQtlGeno(
  x,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  simParamBee = NULL
)

getWorkersQtlGeno(x, nInd = NULL, trait = 1, chr = NULL, simParamBee = NULL)

getDronesQtlGeno(
  x,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{caste}{character, "queen", "fathers", "virginQueens", "workers", or
"drones"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample}

\item{trait}{numeric, indicates which trait's QTL genotypes to retrieve}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{dronesHaploid}{logical, return haploid result for drones?}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
matrix with genotypes when \code{x} is \code{\link{Colony-class}} and
  list of matrices with genotypes when \code{x} is
  \code{\link{Colonies-class}}, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns QTL genotypes of individuals in a
  caste.
}
\section{Functions}{
\itemize{
\item \code{getQueensQtlGeno}: Access QTL genotype data of the queen

\item \code{getFathersQtlGeno}: Access QTL genotype data of fathers

\item \code{getVirginQueensQtlGeno}: Access QTL genotype data of virgin queens

\item \code{getWorkersQtlGeno}: Access QTL genotype data of workers

\item \code{getDronesQtlGeno}: Access QTL genotype data of drones
}}

\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$addTraitAD(nQtlPerChr = 10, meanDD = 0.2, varDD = 0.1)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 2)
colony2 <- addDrones(colony2, nInd = 4)
colony1 <- addVirginQueens(colony1, nInd = 2)

getCasteQtlGeno(colony1, caste = "queen")
getQueensQtlGeno(colony1)

getCasteQtlGeno(colony1, caste = "fathers")
getCasteQtlGeno(colony1, caste = "fathers", nInd = 2)
getCasteQtlGeno(colony1, caste = "fathers", nInd = 2)
getFathersQtlGeno(colony1)
getFathersQtlGeno(colony1, nInd = 2)

getCasteQtlGeno(colony1, caste = "virginQueens")
getVirginQueensQtlGeno(colony1)

getCasteQtlGeno(colony1, caste = "workers")
getWorkersQtlGeno(colony1)

getCasteQtlGeno(colony1, caste = "drones")
getDronesQtlGeno(colony1)

apiary <- c(colony1, colony2)
getCasteQtlGeno(apiary, caste = "queen")
getQueensQtlGeno(apiary)

getCasteQtlGeno(apiary, caste = "fathers")
getCasteQtlGeno(apiary, caste = "fathers", nInd = 2)
getCasteQtlGeno(apiary, caste = "fathers", nInd = 2)
getFathersQtlGeno(apiary)
getFathersQtlGeno(apiary, nInd = 2)

getCasteQtlGeno(apiary, caste = "virginQueens")
getVirginQueensQtlGeno(apiary)

getCasteQtlGeno(apiary, caste = "workers")
getWorkersQtlGeno(apiary)

getCasteQtlGeno(apiary, caste = "drones")
getDronesQtlGeno(apiary)
}
\seealso{
\code{\link{getQtlGeno}} and \code{\link{pullQtlGeno}}
}
