% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteQtlGeno}
\alias{getCasteQtlGeno}
\alias{getQueenQtlGeno}
\alias{getFathersQtlGeno}
\alias{getVirginQueensQtlGeno}
\alias{getWorkersQtlGeno}
\alias{getDronesQtlGeno}
\title{Access QTL genotypes of individuals in a caste}
\usage{
getCasteQtlGeno(
  x,
  caste,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  dronesHaploid = TRUE,
  collapse = FALSE,
  simParamBee = NULL
)

getQueenQtlGeno(x, trait = 1, chr = NULL, collapse = FALSE, simParamBee = NULL)

getFathersQtlGeno(
  x,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  dronesHaploid = TRUE,
  collapse = FALSE,
  simParamBee = NULL
)

getVirginQueensQtlGeno(
  x,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  collapse = FALSE,
  simParamBee = NULL
)

getWorkersQtlGeno(
  x,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  collapse = FALSE,
  simParamBee = NULL
)

getDronesQtlGeno(
  x,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  dronesHaploid = TRUE,
  collapse = FALSE,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}

\item{caste}{character, "queen", "fathers", "workers", "drones",
"virginQueens", or "all"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample}

\item{trait}{numeric, indicates which trait's QTL genotypes to retrieve}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{dronesHaploid}{logical, return haploid result for drones?}

\item{collapse}{logical, if the return value should be a single matrix
with genotypes of all the individuals}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
matrix with genotypes when \code{x} is \code{\link{Colony-class}} and
  list of matrices with genotypes when \code{x} is
  \code{\link{MultiColony-class}}, named by colony id when \code{x} is
  \code{\link{MultiColony-class}}
}
\description{
Level 0 function that returns QTL genotypes of individuals in a
  caste.
}
\section{Functions}{
\itemize{
\item \code{getQueenQtlGeno()}: Access QTL genotype data of the queen

\item \code{getFathersQtlGeno()}: Access QTL genotype data of fathers

\item \code{getVirginQueensQtlGeno()}: Access QTL genotype data of virgin queens

\item \code{getWorkersQtlGeno()}: Access QTL genotype data of workers

\item \code{getDronesQtlGeno()}: Access QTL genotype data of drones

}}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$addTraitA(nQtlPerChr = 10)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
droneGroups <- pullDroneGroupsFromDCA(drones, n = 10, nDrones = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, drones = droneGroups[[1]])
colony <- buildUp(x = colony, nWorkers = 6, nDrones = 3)
colony <- addVirginQueens(x = colony, nInd = 5)

apiary <- createMultiColony(basePop[3:4], n = 2)
apiary <- cross(apiary, drones = droneGroups[c(2, 3)])
apiary <- buildUp(x = apiary, nWorkers = 6, nDrones = 3)
apiary <- addVirginQueens(x = apiary, nInd = 5)

getCasteQtlGeno(colony, caste = "queen")
getQueenQtlGeno(colony)

getCasteQtlGeno(colony, caste = "fathers")
getCasteQtlGeno(colony, caste = "fathers", nInd = 2)
getCasteQtlGeno(colony, caste = "fathers", nInd = 2) # random sample!
getFathersQtlGeno(colony)
getFathersQtlGeno(colony, nInd = 2)

getCasteQtlGeno(colony, caste = "virginQueens")
getVirginQueensQtlGeno(colony)

getCasteQtlGeno(colony, caste = "workers")
getWorkersQtlGeno(colony)

getCasteQtlGeno(colony, caste = "drones")
getDronesQtlGeno(colony)

# Get genotypes for all individuals
getCasteQtlGeno(colony, caste = "all")
# Get all haplotypes in a single matrix
getCasteQtlGeno(colony, caste = "all", collapse = TRUE)

getCasteQtlGeno(apiary, caste = "queen")
getQueenQtlGeno(apiary)

getCasteQtlGeno(apiary, caste = "fathers")
getCasteQtlGeno(apiary, caste = "fathers", nInd = 2)
getCasteQtlGeno(apiary, caste = "fathers", nInd = 2) # random sample!
getFathersQtlGeno(apiary)
getFathersQtlGeno(apiary, nInd = 2)

getCasteQtlGeno(apiary, caste = "virginQueens")
getVirginQueensQtlGeno(apiary)

getCasteQtlGeno(apiary, caste = "workers")
getWorkersQtlGeno(apiary)

getCasteQtlGeno(apiary, caste = "drones")
getDronesQtlGeno(apiary)

# Get the genotypes of all individuals either by colony or in a single matrix
getCasteQtlGeno(apiary, caste = "all")
getCasteQtlGeno(apiary, caste = "all", collapse = TRUE)
}
\seealso{
\code{\link{getQtlGeno}} and \code{\link{pullQtlGeno}}
}
