% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteQtlGeno}
\alias{getCasteQtlGeno}
\alias{getQueensQtlGeno}
\alias{getFathersQtlGeno}
\alias{getVirginQueensQtlGeno}
\alias{getWorkersQtlGeno}
\alias{getDronesQtlGeno}
\title{Access QTL genotypes of individuals in a caste}
\usage{
getCasteQtlGeno(
  x,
  caste,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getQueensQtlGeno(x, trait = 1, chr = NULL, simParamBee = NULL)

getFathersQtlGeno(
  x,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getVirginQueensQtlGeno(
  x,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  simParamBee = NULL
)

getWorkersQtlGeno(x, nInd = NULL, trait = 1, chr = NULL, simParamBee = NULL)

getDronesQtlGeno(
  x,
  nInd = NULL,
  trait = 1,
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{caste}{character, "queen", "fathers", "workers", "drones", or
"virginQueens"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample}

\item{trait}{numeric, indicates which trait's QTL genotypes to retrieve}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{dronesHaploid}{logical, return haploid result for drones?}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
matrix with genotypes when \code{x} is \code{\link{Colony-class}} and
  list of matrices with genotypes when \code{x} is
  \code{\link{Colonies-class}}, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns QTL genotypes of individuals in a
  caste.
}
\section{Functions}{
\itemize{
\item \code{getQueensQtlGeno()}: Access QTL genotype data of the queen

\item \code{getFathersQtlGeno()}: Access QTL genotype data of fathers

\item \code{getVirginQueensQtlGeno()}: Access QTL genotype data of virgin queens

\item \code{getWorkersQtlGeno()}: Access QTL genotype data of workers

\item \code{getDronesQtlGeno()}: Access QTL genotype data of drones

}}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(TRUE)
SP$setTrackPed(isTrackPed = TRUE)
SP$addTraitA(10)
SP$addSnpChip(5)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(x = colony)
colony <- addVirginQueens(x = colony, nInd = 5)

apiary <- createColonies(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])
apiary <- buildUp(x = apiary)
apiary <- addVirginQueens(x = apiary, nInd = 5)

getCasteQtlGeno(colony, caste = "queen")
getQueensQtlGeno(colony)

getCasteQtlGeno(colony, caste = "fathers")
getCasteQtlGeno(colony, caste = "fathers", nInd = 2)
getCasteQtlGeno(colony, caste = "fathers", nInd = 2)
getFathersQtlGeno(colony)
getFathersQtlGeno(colony, nInd = 2)

getCasteQtlGeno(colony, caste = "virginQueens")
getVirginQueensQtlGeno(colony)

getCasteQtlGeno(colony, caste = "workers")
getWorkersQtlGeno(colony)

getCasteQtlGeno(colony, caste = "drones")
getDronesQtlGeno(colony)

getCasteQtlGeno(apiary, caste = "queen")
getQueensQtlGeno(apiary)

getCasteQtlGeno(apiary, caste = "fathers")
getCasteQtlGeno(apiary, caste = "fathers", nInd = 2)
getCasteQtlGeno(apiary, caste = "fathers", nInd = 2)
getFathersQtlGeno(apiary)
getFathersQtlGeno(apiary, nInd = 2)

getCasteQtlGeno(apiary, caste = "virginQueens")
getVirginQueensQtlGeno(apiary)

getCasteQtlGeno(apiary, caste = "workers")
getWorkersQtlGeno(apiary)

getCasteQtlGeno(apiary, caste = "drones")
getDronesQtlGeno(apiary)
}
\seealso{
\code{\link{getQtlGeno}} and \code{\link{pullQtlGeno}}
}
