% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteQtlHaplo}
\alias{getCasteQtlHaplo}
\alias{getQueenQtlHaplo}
\alias{getFathersQtlHaplo}
\alias{getVirginQueensQtlHaplo}
\alias{getWorkersQtlHaplo}
\alias{getDronesQtlHaplo}
\title{Access QTL haplotypes of individuals in a caste}
\usage{
getCasteQtlHaplo(
  x,
  caste,
  nInd = NULL,
  trait = 1,
  haplo = "all",
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getQueenQtlHaplo(x, trait = 1, haplo = "all", chr = NULL, simParamBee = NULL)

getFathersQtlHaplo(
  x,
  nInd = NULL,
  trait = 1,
  haplo = "all",
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getVirginQueensQtlHaplo(
  x,
  nInd = NULL,
  trait = 1,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)

getWorkersQtlHaplo(
  x,
  nInd = NULL,
  trait = 1,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)

getDronesQtlHaplo(
  x,
  nInd = NULL,
  trait = 1,
  haplo = "all",
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}

\item{caste}{character, "queen", "fathers", "workers", "drones", or
"virginQueens"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample}

\item{trait}{numeric, indicates which trait's QTL haplotypes to retrieve}

\item{haplo}{character, either "all" for all haplotypes or an integer for a
single set of haplotypes, use a value of 1 for female haplotypes and a
value of 2 for male haplotypes}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{dronesHaploid}{logical, return haploid result for drones?}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
matrix with haplotypes when \code{x} is \code{\link{Colony-class}}
  and list of matrices with haplotypes when \code{x} is
  \code{\link{MultiColony-class}}, named by colony id when \code{x} is
  \code{\link{MultiColony-class}}
}
\description{
Level 0 function that returns QTL haplotypes of individuals in a
  caste.
}
\section{Functions}{
\itemize{
\item \code{getQueenQtlHaplo()}: Access QTL haplotype data of the queen

\item \code{getFathersQtlHaplo()}: Access QTL haplotype data of fathers

\item \code{getVirginQueensQtlHaplo()}: Access QTL haplotype data of virgin queens

\item \code{getWorkersQtlHaplo()}: Access QTL haplotype of workers

\item \code{getDronesQtlHaplo()}: Access QTL haplotype data of drones

}}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$addTraitA(nQtlPerChr = 10)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(x = colony, nWorkers = 6, nDrones = 3)
colony <- addVirginQueens(x = colony, nInd = 5)

apiary <- createMultiColony(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])
apiary <- buildUp(x = apiary, nWorkers = 6, nDrones = 3)
apiary <- addVirginQueens(x = apiary, nInd = 5)

getCasteQtlHaplo(colony, caste = "queen")
getQueenQtlHaplo(colony)

getCasteQtlHaplo(colony, caste = "fathers")
getCasteQtlHaplo(colony, caste = "fathers", nInd = 2)
getCasteQtlHaplo(colony, caste = "fathers", nInd = 2) # random sample!
getFathersQtlHaplo(colony)
getFathersQtlHaplo(colony, nInd = 2)

getCasteQtlHaplo(colony, caste = "virginQueens")
getVirginQueensQtlHaplo(colony)

getCasteQtlHaplo(colony, caste = "workers")
getWorkersQtlHaplo(colony)

getCasteQtlHaplo(colony, caste = "drones")
getDronesQtlHaplo(colony)

getCasteQtlHaplo(apiary, caste = "queen")
getQueenQtlHaplo(apiary)

getCasteQtlHaplo(apiary, caste = "fathers")
getCasteQtlHaplo(apiary, caste = "fathers", nInd = 2)
getCasteQtlHaplo(apiary, caste = "fathers", nInd = 2) # random sample!
getFathersQtlHaplo(apiary)
getFathersQtlHaplo(apiary, nInd = 2)

getCasteQtlHaplo(apiary, caste = "virginQueens")
getVirginQueensQtlHaplo(apiary)

getCasteQtlHaplo(apiary, caste = "workers")
getWorkersQtlHaplo(apiary)

getCasteQtlHaplo(apiary, caste = "drones")
getDronesQtlHaplo(apiary)
}
\seealso{
\code{\link{getQtlHaplo}} and \code{\link{pullQtlHaplo}}
}
