% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteSegSiteGeno}
\alias{getCasteSegSiteGeno}
\alias{getQueensSegSiteGeno}
\alias{getFathersSegSiteGeno}
\alias{getVirginQueensSegSiteGeno}
\alias{getWorkersSegSiteGeno}
\alias{getDronesSegSiteGeno}
\title{Access genotypes for all segregating sites of individuals in a
  caste}
\usage{
getCasteSegSiteGeno(
  x,
  caste,
  nInd = NULL,
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getQueensSegSiteGeno(x, chr = NULL, simParamBee = NULL)

getFathersSegSiteGeno(
  x,
  nInd = NULL,
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)

getVirginQueensSegSiteGeno(x, nInd = NULL, chr = NULL, simParamBee = NULL)

getWorkersSegSiteGeno(x, nInd = NULL, chr = NULL, simParamBee = NULL)

getDronesSegSiteGeno(
  x,
  nInd = NULL,
  chr = NULL,
  dronesHaploid = TRUE,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}

\item{caste}{character, "queen", "fathers", "workers", "drones", or
"virginQueens"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{dronesHaploid}{logical, return haploid result for drones?}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
matrix with genotypes when \code{x} is \code{\link{Colony-class}} and
  list of matrices with genotypes when \code{x} is
  \code{\link{MultiColony-class}}, named by colony id when \code{x} is
  \code{\link{MultiColony-class}}
}
\description{
Level 0 function that returns genotypes for all segregating
  sites of individuals in a caste.
}
\section{Functions}{
\itemize{
\item \code{getQueensSegSiteGeno()}: Access genotype data for all segregating sites of the queen

\item \code{getFathersSegSiteGeno()}: Access genotype data for all segregating sites of fathers

\item \code{getVirginQueensSegSiteGeno()}: Access genotype data for all segregating sites of virgin queens

\item \code{getWorkersSegSiteGeno()}: Access genotype data for all segregating sites of workers

\item \code{getDronesSegSiteGeno()}: Access genotype data for all segregating sites of drones

}}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(TRUE)
SP$setTrackPed(isTrackPed = TRUE)
SP$addTraitA(10)
SP$addSnpChip(5)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(x = colony)
colony <- addVirginQueens(x = colony, nInd = 5)

apiary <- createMultiColony(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])
apiary <- buildUp(x = apiary)
apiary <- addVirginQueens(x = apiary, nInd = 5)

getCasteSegSiteGeno(colony, caste = "queen")
getQueensSegSiteGeno(colony)

getCasteSegSiteGeno(colony, caste = "fathers")
getCasteSegSiteGeno(colony, caste = "fathers", nInd = 2)
getCasteSegSiteGeno(colony, caste = "fathers", nInd = 2)
getFathersSegSiteGeno(colony)
getFathersSegSiteGeno(colony, nInd = 2)

getCasteSegSiteGeno(colony, caste = "virginQueens")
getVirginQueensSegSiteGeno(colony)

getCasteSegSiteGeno(colony, caste = "workers")
getWorkersSegSiteGeno(colony)

getCasteSegSiteGeno(colony, caste = "drones")
getDronesSegSiteGeno(colony)

getCasteSegSiteGeno(apiary, caste = "queen")
getQueensSegSiteGeno(apiary)

getCasteSegSiteGeno(apiary, caste = "fathers")
getCasteSegSiteGeno(apiary, caste = "fathers", nInd = 2)
getCasteSegSiteGeno(apiary, caste = "fathers", nInd = 2)
getFathersSegSiteGeno(apiary)
getFathersSegSiteGeno(apiary, nInd = 2)

getCasteSegSiteGeno(apiary, caste = "virginQueens")
getVirginQueensSegSiteGeno(apiary)

getCasteSegSiteGeno(apiary, caste = "workers")
getWorkersSegSiteGeno(apiary)

getCasteSegSiteGeno(apiary, caste = "drones")
getDronesSegSiteGeno(apiary)
}
\seealso{
\code{\link{getSegSiteGeno}} and \code{\link{pullSegSiteGeno}}
}
