% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCasteSegSiteHaplo}
\alias{getCasteSegSiteHaplo}
\alias{getQueensSegSiteHaplo}
\alias{getFathersSegSiteHaplo}
\alias{getVirginQueensSegSiteHaplo}
\alias{getWorkersSegSiteHaplo}
\alias{getDronesSegSiteHaplo}
\title{Access haplotypes for all segregating sites of individuals in a
  caste}
\usage{
getCasteSegSiteHaplo(
  x,
  caste,
  nInd = NULL,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)

getQueensSegSiteHaplo(x, haplo = "all", chr = NULL, simParamBee = NULL)

getFathersSegSiteHaplo(
  x,
  nInd = NULL,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)

getVirginQueensSegSiteHaplo(
  x,
  nInd = NULL,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)

getWorkersSegSiteHaplo(
  x,
  nInd = NULL,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)

getDronesSegSiteHaplo(
  x,
  nInd = NULL,
  haplo = "all",
  chr = NULL,
  simParamBee = NULL
)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{caste}{character, "queen", "fathers", "virgin_queens", "workers", or
"drones"}

\item{nInd}{numeric, number of individuals to access, if \code{NULL} all
individuals are accessed, otherwise a random sample}

\item{haplo}{character, either "all" for all haplotypes or an integer for a
single set of haplotypes, use a value of 1 for female haplotypes and a
value of 2 for male haplotypes}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
matrix with haplotypes when \code{x} is \code{\link{Colony-class}}
  and list of matrices with haplotypes when \code{x} is
  \code{\link{Colonies-class}}, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns haplotypes for all segregating
  sites of individuals in a caste.
}
\section{Functions}{
\itemize{
\item \code{getQueensSegSiteHaplo}: Access haplotype data for all segregating sites of the queen

\item \code{getFathersSegSiteHaplo}: Access haplotype data for all segregating sites of fathers

\item \code{getVirginQueensSegSiteHaplo}: Access haplotype data for all segregating sites of virgin queens

\item \code{getWorkersSegSiteHaplo}: Access haplotype data for all segregating sites of workers

\item \code{getDronesSegSiteHaplo}: Access haplotype data for all segregating sites of drones
}}

\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 2)
colony2 <- addDrones(colony2, nInd = 4)

getCasteSegSiteHaplo(colony1, caste = "queen")
getQueensSegSiteHaplo(colony1)

getCasteSegSiteHaplo(colony1, caste = "fathers")
getCasteSegSiteHaplo(colony1, caste = "fathers", nInd = 2)
getCasteSegSiteHaplo(colony1, caste = "fathers", nInd = 2)
getFathersSegSiteHaplo(colony1)
getFathersSegSiteHaplo(colony1, nInd = 2)

getCasteSegSiteHaplo(colony1, caste = "virgin_queens")
getVirginQueensSegSiteHaplo(colony1)

getCasteSegSiteHaplo(colony1, caste = "workers")
getWorkersSegSiteHaplo(colony1)

getCasteSegSiteHaplo(colony1, caste = "drones")
getDronesSegSiteHaplo(colony1)

apiary <- c(colony1, colony2)
getCasteSegSiteHaplo(apiary, caste = "queen")
getQueensSegSiteHaplo(apiary)

getCasteSegSiteHaplo(apiary, caste = "fathers")
getCasteSegSiteHaplo(apiary, caste = "fathers", nInd = 2)
getCasteSegSiteHaplo(apiary, caste = "fathers", nInd = 2)
getFathersSegSiteHaplo(apiary)
getFathersSegSiteHaplo(apiary, nInd = 2)

getCasteSegSiteHaplo(apiary, caste = "virgin_queens")
getVirginQueensSegSiteHaplo(apiary)

getCasteSegSiteHaplo(apiary, caste = "workers")
getWorkersSegSiteHaplo(apiary)

getCasteSegSiteHaplo(apiary, caste = "drones")
getDronesSegSiteHaplo(apiary)

}
\seealso{
\code{\link{getSegSiteHaplo}} and \code{\link{pullSegSiteHaplo}}
}
