% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCsdGeno}
\alias{getCsdGeno}
\title{Get genotypes from the csd locus}
\usage{
getCsdGeno(x, nInd = NULL, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}

\item{nInd}{numeric, for how many individuals; if \code{NULL} all individuals
are taken; this can be useful as a test of sampling individuals}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
matrix with genotypes when \code{x} is \code{\link{Pop-class}}, list
  of matrices with genotypes when \code{x} is \code{\link{Colony-class}}
  (list nodes named by caste) and list of a list of matrices with genotypes
  when \code{x} is \code{\link{Colonies-class}}, outer list is named by
  colony id when \code{x} is \code{\link{Colonies-class}}; \code{NULL} when
  \code{x} is \code{NULL}
}
\description{
Level 0 function that returns genotypes from the csd locus. See
  \code{\link{SimParamBee}} for more information about the csd locus.
}
\details{
The returned genotypes are spanning multiple bi-allelic SNP of
  a non-recombining csd locus / haplotype. In most cases you will want to use
  \code{\link{getCsdAlleles}}.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 3, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 2)
colony1 <- addVirginQueens(colony1, nInd = 2)
apiary <- c(colony1, colony2)

getCsdGeno(getQueen(colony1))
getCsdGeno(getFathers(colony1))
getCsdGeno(getWorkers(colony1))
getCsdGeno(getDrones(colony1))

getCsdGeno(colony1)

getCsdAlleles(getDrones(colony2))
getCsdAlleles(colony2)

getCsdGeno(apiary)
getCsdGeno(apiary, nInd = 2)

}
