% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getIbdHaplo}
\alias{getIbdHaplo}
\title{Access IBD haplotypes}
\usage{
getIbdHaplo(pop, chr = NULL, snpChip = NULL, simParam = NULL)
}
\arguments{
\item{pop}{\code{\link{Pop-class}}}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{snpChip}{integer, indicating which SNP array loci are to be retrieved,
if \code{NULL}, all sites are retrieved}

\item{simParam}{\code{\link{SimParam}}, global simulation parameters}
}
\value{
Matrix of haplotypes
}
\description{
Level 0 function that returns IBD (identity by descent)
  haplotypes.
}
\seealso{
\code{\link{pullIbdHaplo}}
}
