% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getId}
\alias{getId}
\title{Get the colony ID}
\usage{
getId(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or \code{\link{Colonies-class}}}
}
\value{
character, \code{NA} when queen not present
}
\description{
Level 0 function that returns the colony ID. This is by
  definition the ID of the queen.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)

getId(getQueen(colony1))
getId(colony1)
getId(colony2)
getId(apiary)

colony2 <- removeQueen(colony2)
getId(colony2)

}
