% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getQtlGeno}
\alias{getQtlGeno}
\title{Access QTL genotypes}
\usage{
getQtlGeno(pop, trait = 1, chr = NULL, simParam = NULL)
}
\arguments{
\item{pop}{\code{\link{Pop-class}}}

\item{trait}{numeric, indicates which trait's QTL genotype to retrieve}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{simParam}{\code{\link{SimParam}}, global simulation parameters}
}
\value{
Matrix of genotypes
}
\description{
Level 0 function that returns QTL genotypes.
}
\seealso{
\code{\link{pullQtlGeno}}
}
