% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getQueensAge}
\alias{getQueensAge}
\title{Get (calculate) the queen's age}
\usage{
getQueensAge(x, currentYear)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}

\item{currentYear}{integer, current year}
}
\value{
numeric, the age of the queen(s); named when theres is more
  than one queen; \code{NA} if queen not present
}
\description{
Level 0 function that returns the queen's age.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
apiary <- c(colony1, colony2)

queen <- getQueen(colony1)
queen <- setQueensYOB(queen, year = 2021)
getQueensAge(queen, currentYear = 2022)

colony1 <- setQueensYOB(colony1, year = 2021)
getQueensAge(colony1, currentYear = 2022)

apiary <- setQueensYOB(apiary, year = 2021)
getQueensAge(apiary, currentYear = 2022)
}
