% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getQueensAge}
\alias{getQueensAge}
\title{Get (calculate) the queen's age}
\usage{
getQueensAge(x, currentYear)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{MultiColony-class}}}

\item{currentYear}{integer, current year}
}
\value{
numeric, the age of the queen(s); named when theres is more
  than one queen; \code{NA} if queen not present
}
\description{
Level 0 function that returns the queen's age.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
apiary <- createMultiColony(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])

queen <- getQueen(colony)
queen <- setQueensYOB(queen, year = 2020)
getQueensAge(queen, currentYear = 2022)

colony <- setQueensYOB(colony, year = 2021)
getQueensAge(colony, currentYear = 2022)

apiary <- setQueensYOB(apiary, year = 2018)
getQueensAge(apiary, currentYear = 2022)
}
