% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getQueensYearOfBirth}
\alias{getQueensYearOfBirth}
\alias{getQueensYOB}
\title{Access the queen's year of birth}
\usage{
getQueensYearOfBirth(x)

getQueensYOB(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}} (one or more than one queen),
\code{\link{Colony-class}} (one colony), or
\code{\link{Colonies-class}} (more colonies)}
}
\value{
numeric, the year of birth of the queen(s); named when theres is more
  than one queen; \code{NA} if queen not present
}
\description{
Level 0 function that returns the queen's year of birth.
}
\section{Functions}{
\itemize{
\item \code{getQueensYOB()}: Access the queen's year of birth

}}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])

apiary <- createColonies(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])

queen <- getQueen(colony)
queen <- setQueensYearOfBirth(queen, year = 2022)
getQueensYearOfBirth(queen)

getQueensYearOfBirth(getQueen(colony))
colony <- setQueensYearOfBirth(colony, year = 2030)
getQueensYearOfBirth(colony)

apiary <- setQueensYearOfBirth(apiary, year = 2022)
getQueensYearOfBirth(apiary)
}
