% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getQueensYearOfBirth}
\alias{getQueensYearOfBirth}
\alias{getQueensYOB}
\title{Access the queen's year of birth}
\usage{
getQueensYearOfBirth(x)

getQueensYOB(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}} (one or more than one queen),
\code{\link{Colony-class}} (one colony), or
\code{\link{Colonies-class}} (more colonies)}
}
\value{
numeric, the year of birth of the queen(s); named when theres is more
  than one queen; \code{NA} if queen not present
}
\description{
Level 0 function that returns the queen's year of birth.
}
\section{Functions}{
\itemize{
\item \code{getQueensYOB}: Access the queen's year of birth
}}

\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)

getQueensYearOfBirth(getQueen(colony1))
getQueensYearOfBirth(c(getQueen(colony1), getQueen(colony2)))
getQueensYearOfBirth(colony1)
getQueensYearOfBirth(apiary)

queen1 <- getQueen(colony1)
queen1 <- setQueensYearOfBirth(queen1, year = 2022)
getQueensYearOfBirth(queen1)

queen2 <- getQueen(colony2)
queens <- setQueensYearOfBirth(c(queen1, queen2), year = 2023)
getQueensYearOfBirth(queens)

colony1 <- setQueensYearOfBirth(colony1, year = 2022)
getQueensYearOfBirth(colony1)

apiary <- setQueensYearOfBirth(apiary, year = 2022)
getQueensYearOfBirth(apiary)

}
