% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getSegSiteGeno}
\alias{getSegSiteGeno}
\title{Access genotypes for all segregating sites}
\usage{
getSegSiteGeno(pop, chr = NULL, simParam = NULL)
}
\arguments{
\item{pop}{\code{\link{Pop-class}}}

\item{chr}{numeric, chromosomes to retrieve, if \code{NULL}, all chromosome
are retrieved}

\item{simParam}{\code{\link{SimParam}}, global simulation parameters}
}
\value{
Matrix of genotypes
}
\description{
Level 0 function that returns genotypes for all segregating
  sites.
}
\details{
This is an alias for the AlphaSimR function to distinguish between getting
  and pulling
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
droneGroups <- pullDroneGroupsFromDCA(drones, n = 10, nDrones = nFathersPoisson)

# Create a Colony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, drones = droneGroups[[1]])
colony <- buildUp(x = colony, nWorkers = 6, nDrones = 3)

# Use getSegSiteGeno on a Population
getSegSiteGeno(getWorkers(colony))
getSegSiteGeno(getQueen(colony))
getSegSiteGeno(getDrones(colony))
}
\seealso{
\code{\link{pullSegSiteHaplo}}
}
