% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{hasCollapsed}
\alias{hasCollapsed}
\title{Test if colony has collapsed}
\usage{
hasCollapsed(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
logical, named by colony id when \code{x} is \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns colony collapse status.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 20)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony
hasCollapsed(colony)
colony <- buildUpColony(colony, nWorkers = 100)
colony
hasCollapsed(colony)
tmp <- collapseColony(colony)
tmp
hasCollapsed(tmp)

colony1 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony2 <- createColony(queen = basePop[4], fathers = drones[11:15])
apiary <- c(colony1, colony2)
apiary <- buildUpColonies(apiary, nWorkers = 100)
tmp <- collapseColonies(apiary)
tmp
hasCollapsed(tmp)
}
