% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{hasCollapsed}
\alias{hasCollapsed}
\title{Test if colony has collapsed}
\usage{
hasCollapsed(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
logical, named by colony id when \code{x} is \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns colony collapse status.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(x= colony)
colony <- addVirginQueens(colony, nInd = 5)

hasCollapsed(colony)
colony <- collapse(colony)
hasCollapsed(colony)

apiary <- createColonies(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2,3)])
apiary <- buildUp(x = apiary)

hasCollapsed(apiary)
apiary <- collapse(apiary)
hasCollapsed(apiary)
}
