% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{hasSwarmed}
\alias{hasSwarmed}
\title{Test to see if colony/colonies have swarmed in the current period}
\usage{
\method{hasSwarmed}(x)
}
\arguments{
\item{x}{Undefined argument. Can be a "Colony" class or "Colonies" class}
}
\value{

}
\description{
Test to see if colony/colonies have swarmed in the current period
}
