% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isDrone}
\alias{isDrone}
\title{Is individual a drone}
\usage{
isDrone(x, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Pop-class}}}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
logical
}
\description{
Level 0 function that tests if individuals are drones
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(x = colony)

isDrone(getQueen(colony))
isDrone(getFathers(colony, nInd = 2))
isDrone(getDrones(colony, nInd = 2))

bees <- c(
  getQueen(colony),
  getFathers(colony, nInd = 2),
  getDrones(colony, nInd = 2)
)
isDrone(bees)
}
