% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isProductive}
\alias{isProductive}
\title{Test if colony is currently productive}
\usage{
isProductive(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
logical, named by colony id when \code{x} is \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns colony production status. This can
  be used to decided if colony production can be simulated.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony
isProductive(colony)
colony <- buildUpColony(colony, nWorkers = 100)
colony
isProductive(colony)

colony1 <- createColony(queen = basePop[1], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[2], fathers = drones[6:10])
apiary <- c(colony1, colony2)
apiary <- buildUpColonies(apiary, nWorkers = 100)
isProductive(apiary)

}
