% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isProductive}
\alias{isProductive}
\title{Test if colony is currently productive}
\usage{
isProductive(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
logical, named by colony id when \code{x} is \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns colony production status. This can
  be used to decided if colony production can be simulated.
}
\examples{
founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 20)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones[1:5], nFathers = 5)
colony
isProductive(colony)
colony <- buildUpColony(colony, nWorkers = 100)
colony
isProductive(colony)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[3])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[4])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
apiary <- c(colony1, colony2)
apiary <- buildUpColonies(apiary, nWorkers = 100)
isProductive(apiary)
}
