% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isQueen}
\alias{isQueen}
\title{Is individual a queen}
\usage{
isQueen(x, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Pop-class}}}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
logical
}
\description{
Level 0 function that tests if individuals are queens
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
droneGroups <- pullDroneGroupsFromDCA(drones, n = 10, nDrones = nFathersPoisson)

# Create a Colony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, drones = droneGroups[[1]])

isQueen(getQueen(colony))
isQueen(getFathers(colony, nInd = 2))

bees <- c(
  getQueen(colony),
  getFathers(colony, nInd = 2)
)
isQueen(bees)
}
