% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isQueenMated}
\alias{isQueenMated}
\title{Is the queen mated?}
\usage{
isQueenMated(x)
}
\arguments{
\item{x}{Pop or Colony, queen or colony that will be inspected}
}
\value{
Logical
}
\description{
When queen is mated, we store drones she mated with (fathers of future progeny)
in the same object to have genetic information for generating future progeny.
This function tests if queen is mated, by testing for the presence of fathers.
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybees
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)
isQueenMated(getQueen(colony1))
isQueenMated(colony1)
isQueenMated(apiary)

colony1 <- removeQueen(colony1)
isQueenMated(colony1)

colony2 <- supersedeColony(colony2, fathers = drones[1:5])
isQueenMated(colony2)

}
