% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isQueenMated}
\alias{isQueenMated}
\title{Is the queen mated?}
\usage{
isQueenMated(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}
}
\value{
logical, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns queen's mating status.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony3 <- createColony(virginQueen = basePop[4])
apiary <- c(colony1, colony2, colony3)
isQueenMated(getQueen(colony1))
isQueenMated(colony1)
isQueenMated(apiary)

isQueenMated(removeQueen(colony1))

isQueenMated(supersedeColony(colony2))

isQueenMated(c(getQueen(colony1), getQueen(colony2), getVirginQueens(colony3)))

isQueenMated(crossVirginQueen(basePop[2], drones))

}
