% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isVirginQueen}
\alias{isVirginQueen}
\title{Is individual a virgin queen}
\usage{
isVirginQueen(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}}
}
\value{
logical
}
\description{
Level 0 function that tests if individuals are virgin queens
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony <- addWorkers(colony)
colony <- addDrones(colony)
colony <- addVirginQueens(colony)

isVirginQueen(getQueen(colony))
isVirginQueen(getWorkers(colony, nInd = 2))
isVirginQueen(getDrones(colony, nInd = 2))
isVirginQueen(getVirginQueens(colony, nInd = 2))

bees <- c(
  getQueen(colony),
  getWorkers(colony, nInd = 2),
  getDrones(colony, nInd = 2),
  getVirginQueens(colony, nInd = 2)
)
isVirginQueen(bees)
}
