% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{isWorker}
\alias{isWorker}
\title{Is individual a worker}
\usage{
isWorker(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}}
}
\value{
logical
}
\description{
Level 0 function that tests if individuals are workers
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony <- addWorkers(colony)
colony <- addDrones(colony)
colony <- addVirginQueens(colony)

isWorker(getQueen(colony))
isWorker(getWorkers(colony, nInd = 2))
isWorker(getDrones(colony, nInd = 2))
isWorker(getVirginQueens(colony, nInd = 2))

bees <- c(getQueen(colony),
          getWorkers(colony, nInd = 2),
          getDrones(colony, nInd = 2),
          getVirginQueens(colony, nInd = 2))
isWorker(bees)

}
