% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nColonies}
\alias{nColonies}
\title{Number of colonies}
\usage{
nColonies(colonies)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}
}
\value{
integer
}
\description{
Level 0 function that returns the number of colonies in a
  colonies object.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)
nColonies(apiary)
nColonies(createColonies(nCol = 10))
}
