% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nColonies}
\alias{nColonies}
\title{Number of colonies}
\usage{
nColonies(colonies)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}
}
\description{
Level 0 function that returns the number of colonies in a
  colonies object.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
apiary <- c(colony1, colony2)
nColonies(apiary)
nColonies(createColonies(n = 10))
}
