% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nColonies}
\alias{nColonies}
\alias{nNULLColonies}
\alias{nEmptyColonies}
\title{Number of colonies in a MultiColony object}
\usage{
nColonies(multicolony)

nNULLColonies(multicolony)

nEmptyColonies(multicolony)
}
\arguments{
\item{multicolony}{\code{\link{MultiColony-class}}}
}
\description{
Level 0 function that returns the number of colonies in a
  MultiColony object.
}
\section{Functions}{
\itemize{
\item \code{nNULLColonies()}: Number of \code{NULL} colonies in a MultiColony object

\item \code{nEmptyColonies()}: Number of empty colonies in a MultiColony object

}}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
droneGroups <- pullDroneGroupsFromDCA(drones, n = 10, nDrones = nFathersPoisson)
apiary <- createMultiColony(basePop[2:3], n = 2)
nColonies(apiary)
nColonies(createMultiColony(n = 10))
}
\seealso{
\code{\link{nNULLColonies}} and \code{\link{nEmptyColonies}}
}
