% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nFathers}
\alias{nFathers}
\title{Number of fathers in a colony}
\usage{
nFathers(x)
}
\arguments{
\item{x}{Colony or Colonies}
}
\value{
Integer, named by colony id when \code{x} is Colonies
}
\description{
Returns the number of nFathers (drones the queen mated with) in
a colony
}
\examples{
AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybees
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:8])
nFathers(colony1)
nFathers(colony2)

apiary <- c(colony1, colony2)
nFathers(apiary)

}
