% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nFathers}
\alias{nFathers}
\title{Number of fathers in a colony}
\usage{
nFathers(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
integer, named by colony id when \code{x} is \code{\link{Colonies-class}}
}
\description{
Returns the number of nFathers (drones the queen mated with) in a colony
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:8])
nFathers(colony1)
nFathers(colony2)

apiary <- c(colony1, colony2)
nFathers(apiary)

}
