% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nNULLColonies}
\alias{nNULLColonies}
\title{Number of NULL colonies}
\usage{
nNULLColonies(colonies)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}
}
\value{
integer
}
\description{
Level 0 function that returns the number of colonies in a
  colonies object that are in fact empty \code{NULL}.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

founderDrones <- createFounderDrones(pop = basePop[1:2], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[3], fathers = founderDrones[1:10])
colony2 <- createColony(queen = basePop[4], fathers = founderDrones[11:20])
apiary <- c(colony1, colony2)
getId(apiary)

nColonies(apiary)
nColonies(selectColonies(apiary, ID = c("3", "4")))
nColonies(selectColonies(apiary, ID = c("4", "5")))
nColonies(selectColonies(apiary, ID = c("5", "6")))

nNULLColonies(apiary)
nNULLColonies(selectColonies(apiary, ID = c("3", "4")))
nNULLColonies(selectColonies(apiary, ID = c("4", "5")))
nNULLColonies(selectColonies(apiary, ID = c("5", "6")))

}
