% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nNULLColonies}
\alias{nNULLColonies}
\title{Number of NULL colonies}
\usage{
nNULLColonies(colonies)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}
}
\value{
integer
}
\description{
Level 0 function that returns the number of colonies in a
  colonies object that are in fact empty \code{NULL}.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
apiary <- c(colony1, colony2)
getId(apiary)

nColonies(apiary)
nColonies(selectColonies(apiary, ID = c("3", "4")))
nColonies(selectColonies(apiary, ID = c("4", "5")))
nColonies(selectColonies(apiary, ID = c("5", "6")))

nNULLColonies(apiary)
nNULLColonies(selectColonies(apiary, ID = c("3", "4")))
nNULLColonies(selectColonies(apiary, ID = c("4", "5")))
nNULLColonies(selectColonies(apiary, ID = c("5", "6")))
}
