% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nNULLColonies}
\alias{nNULLColonies}
\title{Number of NULL colonies}
\usage{
nNULLColonies(colonies)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}
}
\value{
integer
}
\description{
Level 0 function that returns the number of colonies in a
  colonies object that are in fact empty \code{NULL}.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

apiary <- createColonies(basePop[2:3], n = 2)
apiary <- cross(x = apiary, fathers = fatherGroups[c(2,3)])
getId(apiary)

nColonies(apiary)
nColonies(selectColonies(apiary, ID = c("1", "2")))

nNULLColonies(apiary)
nNULLColonies(selectColonies(apiary, ID = c("1", "2")))
}
