% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nVirginQueens}
\alias{nVirginQueens}
\title{Number of virgin queens in a colony}
\usage{
nVirginQueens(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
integer, named by colony id when \code{x} is \code{\link{Colonies-class}}
}
\description{
Returns the number of virgin queens in a colony
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addVirginQueens(colony1, nInd = 3)
colony2 <- addVirginQueens(colony2, nInd = 5)
nVirginQueens(colony1)
nVirginQueens(colony2)

apiary <- c(colony1, colony2)
nVirginQueens(apiary)

}
