% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nWorkers}
\alias{nWorkers}
\title{Number of workers in a colony}
\usage{
nWorkers(x)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}
}
\value{
integer, named by colony id when \code{x} is \code{\link{Colonies-class}}
}
\description{
Returns the number of workers in a colony
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- addWorkers(x = colony, nInd = 50, exact = TRUE)
nWorkers(colony)

# Create a Multicolony class
apiary <- createColonies(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2,3)])

# If exact = TRUE, all 50 individuals are added
tmp <- addWorkers(x = apiary[1], nInd = 100, exact = TRUE)
nWorkers(tmp)
# If exact = FALSE, some of the workers are removed due to homozygosity
# on the csd (when the csd is turned on)
tmp2 <- addWorkers(x = apiary[2], nInd = 200, exact = FALSE)
nWorkers(tmp2)

apiary <- c(tmp, tmp2)
nWorkers(apiary)
}
