% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SimParamBee.R
\name{phenoQueenPlusSumOfWorkers}
\alias{phenoQueenPlusSumOfWorkers}
\title{Sample colony phenotype}
\usage{
phenoQueenPlusSumOfWorkers(colony, queenTrait = 1, workersTrait = 2)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{queenTrait}{numeric, trait that represents queen's effect on honey
yield (defined in \code{\link{SimParamBee}} - see examples)}

\item{workersTrait}{numeric, trait that represents wokers's effect on honey
yield (defined in \code{\link{SimParamBee}} - see examples)}
}
\value{
numeric matrix with a single value
}
\description{
Sample colony phenotype when \code{colonyFUN = NULL} (see
  \code{\link{SimParamBee}$phenoColony}) - this is just an example and you
  will likely want to define your own sampling functions!
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)

# Define two traits that collectively affect colony honey yield:
# 1) queen's effect on colony honey yield
# 2) workers' effect on colony honey yield
# The traits will have negative genetic correlation and heritability of 0.25
meanP <- c(20, 0)
varA <- c(1, 1 / 10)
corA <- matrix(data = c( 1.0, -0.5,
                        -0.5,  1.0), nrow = 2, byrow = TRUE)
varE <- c(3, 3 / 10)
varA / (varA + varE)
SP$addTraitA(nQtlPerChr = 100, mean = meanP, var = varA, corA = corA)
SP$setVarE(varE = varE)

basePop <- createVirginQueens(founderGenomes)
drones <- createDrones(x = basePop[1], nInd = 5)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones, nFathers = 5)
colony <- buildUpColony(colony, nWorkers = 10)

# Set phenotypes for all colony individuals
colony <- setPhenoColony(colony)

# Queen's phenotype for both traits
pheno(getQueen(colony))
# TODO: use getQueensPheno(colony, caste = "queen")
#       https://github.com/HighlanderLab/SIMplyBee/issues/26

# Workers' phenotype for both traits
pheno(getWorkers(colony))
# TODO: use getWorkersPheno(colony, caste = "queen")
#       https://github.com/HighlanderLab/SIMplyBee/issues/26

# TODO: use getColonyPheno(colony) for all individuals
#       https://github.com/HighlanderLab/SIMplyBee/issues/26

# Set phenotypes for all colony individuals AND colony
colony <- setPhenoColony(colony,
                         colonyFUN = phenoQueenPlusSumOfWorkers)
pheno(colony)
# TODO: use getColonyPheno(colony) for all individuals and/or colony
#       https://github.com/HighlanderLab/SIMplyBee/issues/26
}
\seealso{
\code{\link{SimParamBee}} field \code{phenoColony}
}
