% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{pHomBrood}
\alias{pHomBrood}
\title{Percentage of homozygous brood of a queen}
\usage{
pHomBrood(x)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
  \code{\link{Colonies-class}}

TODO: describe queen's and colony's pHomBrood}
}
\value{
numeric, named by colony id when \code{x} is
  \code{\link{Colonies-class}}
}
\description{
Level 0 function that returns the percentage of homozygous brood
  in a colony (these are non viable individuals and only their number is
  stored).
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 10)
colony2 <- addWorkers(colony2, nInd = 20)
colony1 <- addDrones(colony1, nInd = 10)
colony2 <- addDrones(colony2, nInd = 20)

# Virgin queen
pHomBrood(basePop[2])

# Mated queen
pHomBrood(crossVirginQueen(pop = basePop[2], fathers = drones[1:5]))

# Queen of the colony
pHomBrood(getQueen(colony1))

# Colony
pHomBrood(colony1)

# Colonies
apiary <- c(colony1, colony2)
pHomBrood(apiary)

}
