% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{pullColonies}
\alias{pullColonies}
\title{Pull out some colonies}
\usage{
pullColonies(colonies, ID = NULL, p = NULL)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{ID}{numeric or character, name of a colony (one or more) to be pulled
out; note that numeric value is converted to character}

\item{p}{numeric, probability of a colony being pulled out}
}
\value{
list with two \code{\link{Colonies-class}}, the \code{pulledColonies}
  and the \code{remainingColonies}
}
\description{
Level 3 function that pulls out some colonies based on colony
  ID or random selection.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

apiary <- createColonies(pop = basePop, nCol = 3)
(names <- getId(apiary))

tmp <- pullColonies(apiary, ID = names[1])
getId(tmp$pulledColonies)
getId(tmp$remainingColonies)

tmp <- pullColonies(apiary, p = 0.5)
getId(tmp$pulledColonies)
getId(tmp$remainingColonies)

tmp <- pullColonies(apiary, p = 0.5)
getId(tmp$pulledColonies)
getId(tmp$remainingColonies)

}
