% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{pullColonies}
\alias{pullColonies}
\title{Pull out some colonies}
\usage{
pullColonies(colonies, ID = NULL, n = NULL, p = NULL)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{ID}{character or numeric, name of a colony (one or more) to be pulled
out; if character (numeric) colonies are pulled out based on their name
(position)}

\item{n}{numeric, number of colonies to select}

\item{p}{numeric, probability of a colony being pulled out (takes precedence
over \code{n})}
}
\value{
list with two \code{\link{Colonies-class}}, the \code{pulledColonies}
  and the \code{remainingColonies}
}
\description{
Level 3 function that pulls out some colonies based on colony
  ID or random selection.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1:4], nInd = 100)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = 10)
apiary <- createColonies(basePop[2:5], n = 4)
apiary <- cross(apiary, fathers = fatherGroups[1:4])
getId(apiary)

tmp <- pullColonies(apiary, ID = c(1, 2))
getId(tmp$pulledColonies)
getId(tmp$remainingColonies)

tmp <- pullColonies(apiary, ID = c("5", "6"))
getId(tmp$pulledColonies)
getId(tmp$remainingColonies)

tmp <- pullColonies(apiary, ID = 5)
getId(tmp$pulledColonies)
getId(tmp$remainingColonies)

tmp <- pullColonies(apiary, ID = "9")
getId(tmp$pulledColonies)
getId(tmp$remainingColonies)

tmp <- pullColonies(apiary, n = 2)
getId(tmp$pulledColonies)
getId(tmp$remainingColonies)

tmp <- pullColonies(apiary, p = 0.75)
getId(tmp$pulledColonies)
getId(tmp$remainingColonies)
}
