% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{pullDroneGroupsFromDCA}
\alias{pullDroneGroupsFromDCA}
\title{Pulls drone groups from a Drone Congregation Area (DCA)}
\usage{
pullDroneGroupsFromDCA(DCA, n, nFathers = NULL, simParamBee = NULL)
}
\arguments{
\item{DCA}{\code{\link{Pop-class}}, population of drones;
\code{\link{isDrone}} test will be run on these individuals}

\item{n}{integer, number of drone groups to be created}

\item{nFathers}{numeric, average number of drones/fathers per group,
the actual number is sampled from a truncated Poisson with lambda of
nFathers}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
list of \code{\link{Pop-class}}
}
\description{
Level 1 function that pulls drone groups from a Drone
  Congregation Area (DCA) to use them later in mating. Number of drones per
  group is sampled from a truncated Poisson distribution parameterised with
  an average group size and truncation at zero (= zeroes are excluded).
  Drones are pulled (removed) from the DCA to reflect the fact that drones
  die after mating, so they can't be present in the DCA anymore.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addDrones(colony1, nInd = 10)
colony2 <- addDrones(colony2, nInd = 20)
apiary <- c(colony1, colony2)
DCA <- createDCA(apiary)
pullDroneGroupsFromDCA(DCA, n = 4, nFathers = 5)
}
