% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{removeColonies}
\alias{removeColonies}
\title{Remove some colonies}
\usage{
removeColonies(colonies, ID)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{ID}{character or numeric, name of a colony (one or more) to be
removed; if character (numeric) colonies are removed based on their name
(position)}
}
\value{
\code{\link{Colonies-class}} with some colonies removed
}
\description{
Level 3 function that removes some colonies based on their ID.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1:4], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[5:10], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[11:20], nFathers = 5)
colony3 <- createColony(x = basePop[4])
colony3 <- crossColony(colony3, drones = drones[21:30], nFathers = 5)
colony4 <- createColony(x = basePop[5])
colony4 <- crossColony(colony4, drones = drones[31:40], nFathers = 5)
apiary <- c(colony1, colony2, colony3, colony4)
getId(apiary)

getId(removeColonies(apiary, ID = 1))
getId(removeColonies(apiary, ID = "5"))

getId(removeColonies(apiary, ID = c(1, 2)))
getId(removeColonies(apiary, ID = c("5", "6")))

getId(removeColonies(apiary, ID = 5))
getId(removeColonies(apiary, ID = "9"))
}
