% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{removeColonies}
\alias{removeColonies}
\title{Remove some colonies}
\usage{
removeColonies(colonies, ID)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{ID}{numeric or character, name of a colony (one or more) to be removed;
note that numeric value is converted to character}
}
\value{
\code{\link{Colonies-class}} with some colonies removed
}
\description{
Level 3 function that removes some colonies based on their ID.
}
\examples{
founderGenomes <- quickHaplo(nInd = 4, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

apiary <- createColonies(pop = basePop, nCol = 3)
(names <- getId(apiary))

getId(removeColonies(apiary, ID = names[1]))
getId(removeColonies(apiary, ID = names[c(2, 3)]))

}
