% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{removeColonies}
\alias{removeColonies}
\title{Remove some colonies from the MultiColony object}
\usage{
removeColonies(multicolony, ID)
}
\arguments{
\item{multicolony}{\code{\link{MultiColony-class}}}

\item{ID}{character, ID of a colony (one or more) to be
removed}
}
\value{
\code{\link{MultiColony-class}} with some colonies removed
}
\description{
Level 3 function that removes some colonies
  from the MultiColony object based on their ID.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1:4], nInd = 100)
droneGroups <- pullDroneGroupsFromDCA(drones, n = 10, nDrones = 10)
apiary <- createMultiColony(basePop[2:5], n = 4)
apiary <- cross(apiary, drones = droneGroups[1:4])
getId(apiary)

getId(removeColonies(apiary, ID = 1))
getId(removeColonies(apiary, ID = "5"))

getId(removeColonies(apiary, ID = c(1, 2)))
getId(removeColonies(apiary, ID = c("5", "6")))

getId(removeColonies(apiary, ID = 5))
getId(removeColonies(apiary, ID = "9"))

nColonies(apiary)
apiary <- removeColonies(apiary, ID = "2")
nColonies(apiary)

}
