% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{removeColonies}
\alias{removeColonies}
\title{Remove some colonies}
\usage{
removeColonies(colonies, ID)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{ID}{character or numeric, name of a colony (one or more) to be
removed; if character (numeric) colonies are removed based on their name
(position)}
}
\value{
\code{\link{Colonies-class}} with some colonies removed
}
\description{
Level 3 function that removes some colonies based on their ID.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1:4], nInd = 100)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = 10)
apiary <- createColonies(basePop[2:5], n = 4)
apiary <- cross(apiary, fathers = fatherGroups[1:4])
getId(apiary)

getId(removeColonies(apiary, ID = 1))
getId(removeColonies(apiary, ID = "5"))

getId(removeColonies(apiary, ID = c(1, 2)))
getId(removeColonies(apiary, ID = c("5", "6")))

getId(removeColonies(apiary, ID = 5))
getId(removeColonies(apiary, ID = "9"))

nColonies(apiary)
apiary <- removeColonies(apiary, ID = "2")
nColonies(apiary)

}
