% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeDrones}
\alias{removeDrones}
\title{Remove a proportion of drones}
\usage{
removeDrones(x, p = 1, use = "rand")
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}

\item{p}{numeric, proportion to be removed}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of drones that will stay when \code{p < 1}}
}
\value{
\code{\link{Colony-class}} or \code{\link{MultiColony-class}} without drones
}
\description{
Level 2 function that removes a proportion of drones from a
Colony or MultiColony object, for example, at the end of summer.
}
\examples{
founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 100)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 5, nFathers = nFathersPoisson)

# Create and cross Colony and MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
apiary <- createMultiColony(basePop[4:5], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[3:4])

# Add drones
colony <- addDrones(colony, nInd = 10)
apiary <- addDrones(apiary, nInd = 10)

# Remove drones
nDrones(colony)
colony <- removeDrones(colony, p = 0.5)
nDrones(colony)
colony <- removeDrones(colony)
nDrones(colony)

nDrones(apiary)
apiary <- removeDrones(apiary, p = 0.5)
nDrones(apiary)
apiary <- removeDrones(apiary)
nDrones(apiary)

}
