% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeQueen}
\alias{removeQueen}
\title{Remove queen}
\usage{
removeQueen(colony)
}
\arguments{
\item{colony}{Colony}
}
\value{
Colony without a queen
}
\description{
Remove queen of a colony
}
\examples{
AlphaSimR
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybees
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony
getQueen(colony)

colony <- removeQueen(colony)
colony
getQueen(colony)

}
