% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeQueen}
\alias{removeQueen}
\title{Remove queen}
\usage{
removeQueen(colony)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}
}
\value{
\code{\link{Colony-class}} without the queen
}
\description{
Level 2 function that removes the queen of a colony.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones[1:5], nFathers = 5)
colony
getQueen(colony)

colony <- removeQueen(colony)
colony
getQueen(colony)
}
