% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeVirginQueens}
\alias{removeVirginQueens}
\title{Remove virgin queens}
\usage{
removeVirginQueens(colony)
}
\arguments{
\item{colony}{Colony}
}
\value{
Colony without virgin queens
}
\description{
Remove virgin queens of a colony
}
\examples{
AlphaSimR
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybees
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony
getVirginQueens(colony)

colony <- removeVirginQueens(colony)
colony
getVirginQueens(colony)

}
