% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeVirginQueens}
\alias{removeVirginQueens}
\title{Remove a proportion of virgin queens}
\usage{
removeVirginQueens(x, p = 1, use = "rand")
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{p}{numeric, proportion to be removed}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of virgins queens that will stay when \code{p < 1}}
}
\value{
\code{\link{Colony-class}} without virgin queens
}
\description{
Level 2 function that removes a proportion of virgin queens of a
  colony.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(queen = basePop[2], fathers = drones[1:5])
colony <- addVirginQueens(colony, nInd = 10)
getVirginQueens(colony)

colony <- removeVirginQueens(colony)
colony
getVirginQueens(colony)
}
