% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeWorkers}
\alias{removeWorkers}
\title{Remove a proportion of workers}
\usage{
removeWorkers(x, p = 1, use = "rand")
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{p}{numeric, proportion to be removed}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of workers that will stay when \code{p < 1}}
}
\value{
\code{\link{Colony-class}} or \code{\link{Colonies-class}} without workers
}
\description{
Level 2 function that removes a proportion of workers, for
  example, as a preparation for winter.
}
\examples{
founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 100)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 5, nFathers = nFathersPoisson)

# Create and cross Colony and MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
apiary <- createColonies(basePop[4:5], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[3:4])

# Add workers
colony <- addWorkers(colony, nInd = 10)
apiary <- addWorkers(apiary, nInd = 10)

# Remove workers
nWorkers(colony)
colony <- removeWorkers(x = colony, p = 0.5)
nWorkers(colony)
colony <- removeWorkers(x = colony)
nWorkers(colony)
nWorkers(apiary)
apiary <- removeWorkers(apiary, p = 0.5)
nWorkers(apiary)
apiary <- removeWorkers(apiary)
nWorkers(apiary)

}
