% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeWorkers}
\alias{removeWorkers}
\title{Remove a proportion of workers}
\usage{
removeWorkers(x, p = 1, use = "rand")
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{p}{numeric, proportion to be removed}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of workers that will stay when \code{p < 1}}
}
\value{
\code{\link{Colony-class}} without workers
}
\description{
Level 2 function that removes a proportion of workers, for
  example, as a preparation for winter.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 5)
colony <- createColony(x = basePop[2])
colony <- crossColony(colony, drones = drones, nFathers = 5)
colony

colony <- addWorkers(colony, nInd = 10)
colony
getWorkers(colony)@id

colony <- removeWorkers(x = colony, p = 0.5)
colony
getWorkers(colony)@id

colony <- removeWorkers(x = colony, p = 1.0)
colony
getWorkers(colony)
}
