% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{replaceDrones}
\alias{replaceDrones}
\title{Replace a proportion of drones with new ones}
\usage{
replaceDrones(colony, p = 1, use = "rand")
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{p}{numeric, proportion of drones to be replaced with new ones}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of drones that stay when \code{p < 1}}
}
\value{
\code{\link{Colony-class}} with replaced drones
}
\description{
Level 2 function that replaces a proportion of drones with new
  drones from the colony. Useful after events like season change, swarming,
  supersedure, etc. due to the short life span of the drones.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
colony
colony <- addDrones(colony, nInd = 10)
colony
getDrones(colony)@id
colony <- replaceDrones(colony, p = 0.5)
getDrones(colony)@id

colony <- replaceDrones(colony, p = 1.0)
getDrones(colony)@id

colony <- replaceDrones(colony, p = 1.5)
getDrones(colony)@id

}
