% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{replaceVirginQueens}
\alias{replaceVirginQueens}
\title{Replace a proportion of virgin queens with new ones}
\usage{
replaceVirginQueens(x, p = 1, use = "rand", year = NULL, simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{p}{numeric, proportion of virgin queens to be replaced with new ones}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of virgin queens that stay when \code{p < 1}}

\item{year}{numeric, year of birth for virgin queens}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colony-class}} or  or \code{\link{Colonies-class}} with
  replaced virgin queens
}
\description{
Level 2 function that replaces a proportion of virgin queens
  with new virgin queens from the colony. Useful after events like season
  change, swarming, supersedure, etc. due to the short life span of the
  virgin queens.
}
\examples{
founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 100)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 5, nFathers = nFathersPoisson)

# Create and cross Colony and MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
apiary <- createColonies(basePop[4:5], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[3:4])

# Add virgin queens
colony <- addVirginQueens(colony, nInd = 20)
apiary <- addVirginQueens(apiary, nInd = 20)

# Replace virgin queens
getVirginQueens(colony)@id
colony <- replaceVirginQueens(colony, p = 0.5)
getVirginQueens(colony)@id
colony <- replaceVirginQueens(colony, p = 1.5)
getVirginQueens(colony)@id

lapply(getVirginQueens(apiary), FUN = function(x) x@id)
apiary <- replaceVirginQueens(apiary, p = 0.5)
lapply(getVirginQueens(apiary), FUN = function(x) x@id)
apiary <- replaceVirginQueens(apiary, p = 1.5)
lapply(getVirginQueens(apiary), FUN = function(x) x@id)
}
