% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{replaceWorkers}
\alias{replaceWorkers}
\title{Replace a proportion of workers with new ones}
\usage{
replaceWorkers(colony, p = 1, use = "rand", simParamBee = NULL)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{p}{numeric, proportion of workers to be replaced with new ones}

\item{use}{character, all the options provided by \code{\link{selectInd}} -
guides selection of workers that stay when \code{p < 1}}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colony-class}} with replaced workers
}
\description{
Level 2 function that replaces a proportion of workers with new
  workers from the colony. Useful after events like season change, swarming,
  supersedure, etc. due to the short life span of the workers.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 5)
colony <- createColony(queen = basePop[2], fathers = drones)
colony
colony <- addWorkers(colony, nInd = 10)
colony
getWorkers(colony)@id
colony <- replaceWorkers(colony, p = 0.5)
getWorkers(colony)@id

colony <- replaceWorkers(colony, p = 1.0)
getWorkers(colony)@id

colony <- replaceWorkers(colony, p = 1.5)
getWorkers(colony)@id

}
