% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{reQueenColony}
\alias{reQueenColony}
\title{reQueenColony}
\usage{
\method{reQueenColony}(colony, queen)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object}

\item{queen}{Pop/number of queens to add to the queen-less colony/colonies}
}
\value{
Updated AlphaSimRBee Colony object
}
\description{
Re-queens a colony/colonies that have no queen or virgin queens present.
This function allows users to insert a virgin queen or a previously mated queen
into the queen-less colony.
}
\examples{
Create founder haplotypes
founderPop <- quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP <- SimParam$new(founderPop)

#Create population
base <- newPop(founderPop, simParam=SP)

#Create 10 mated colonies from the base population
apiary1 <- createMultipleMatedColonies(base, nColonies = 10, nAvgFathers = 15)

#Build-up the colonies
apiary1 <- buildUpColonies(apiary1, nWorkers = colonyFullSize, nDrones = colonyFullSize * 0.1)

#Split all the colonies
tmp <- splitColonies(apiary1)
apiary1 <- tmp$remnants
apiary0 <- tmp$splits

#Create 10 virgin queens from the base population
virginQueens <- base[1]

#Repopulate the split colonies with virgin queens taken from the base population
apiary0 <- reQueenColonies(apiary0, queen = virginQueens)

}
