% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{resetEvents}
\alias{resetEvents}
\title{Reset colony events}
\usage{
resetEvents(x, collapse = NULL)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{Colonies-class}}}

\item{collapse}{logical, reset the collapse event (only sensible in setting
up a new colony, which the default of \code{NULL} caters for; otherwise, a
collapsed colony should be left collapsed forever, unless you force
resetting this event with \code{collapse = TRUE})}
}
\value{
\code{\link{Colony-class}} or \code{\link{Colonies-class}} with
  events reset
}
\description{
Level 2 function that resets the slots swarm, split,
  supersedure, collapsed, and production to FALSE. Useful at the end of a
  yearly cycle to reset the events, allowing the user to track new events in
  a new year.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
colony1
apiary <- c(colony1, colony2)

(colony1 <- buildUpColony(colony1, nWorkers = 100))
resetEvents(colony1)
apiary <- buildUpColonies(apiary, nWorkers = 100)
apiary[[1]]
resetEvents(apiary)[[1]]

tmp <- splitColony(colony1)
(split <- tmp$split)
resetEvents(split)
(remnant <- tmp$remnant)
resetEvents(remnant)

tmp <- splitColonies(apiary)
(splits <- tmp$splits)
splits[[1]]
resetEvents(splits)[[1]]
(remnants <- tmp$remnants)
remnants[[1]]
resetEvents(remnants)[[1]]

tmp <- swarmColony(colony1)
(swarm <- tmp$swarm)
resetEvents(swarm)
(remnant <- tmp$remnant)
resetEvents(remnant)

tmp <- swarmColonies(apiary)
(swarms <- tmp$swarms)
swarms[[1]]
resetEvents(swarms)[[1]]
(remnants <- tmp$remnants)
remnants[[1]]
resetEvents(remnants)[[1]]

(tmp <- supersedeColony(colony1))
resetEvents(tmp)

(tmp <- supersedeColonies(apiary))
tmp[[1]]
resetEvents(tmp)[[1]]

(tmp <- collapseColony(colony1))
resetEvents(tmp)
resetEvents(tmp, collapse = TRUE)

(tmp <- collapseColonies(apiary))
tmp[[1]]
resetEvents(tmp)[[1]]
resetEvents(tmp, collapse = TRUE)[[1]]
}
