% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{selectColonies}
\alias{selectColonies}
\title{Select individual colonies from a set of colonies}
\usage{
selectColonies(colonies, ID = NULL, p = NULL)
}
\arguments{
\item{colonies}{Colonies, a set of colonies}

\item{ID}{numeric or character, name of a colony (one or more) in
\code{colonies}; note that numeric value is effectively converted to character
- see examples how to select by index}
}
\value{
Colonies, selected colonies
}
\description{
Select individual colonies from a set of colonies based on colony
index or name
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 6, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybees
founderDrones <- createFounderDrones(pop = basePop[1:3], nDronesPerQueen = 10)

colony1 <- createColony(queen = basePop[4], fathers = founderDrones[1:10])
colony2 <- createColony(queen = basePop[5], fathers = founderDrones[11:20])
colony3 <- createColony(queen = basePop[6], fathers = founderDrones[21:30])
apiary <- c(colony1, colony2, colony3)

selectColonies(apiary, ID = 4)
selectColonies(apiary, ID = "4")
selectColonies(apiary, ID = c(4, 5))
selectColonies(apiary, ID = c("4", "5"))

apiary[1]
apiary[[1]]
apiary["4"]
apiary[["4"]]
getId(apiary[c(1, 2)])
getId(apiary[c("4", "5")])
getId(apiary[c(2, 1)])
getId(apiary[c("5", "4")])

}
