% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{selectColonies}
\alias{selectColonies}
\title{Select individual colonies}
\usage{
selectColonies(colonies, ID = NULL, n = NULL, p = NULL)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{ID}{character or numeric, name of a colony (one or more) to be
selected; if character (numeric) colonies are selected based on their name
(position)}

\item{n}{numeric, number of colonies to select}

\item{p}{numeric, probability of a colony being selected (takes precedence
over \code{n})}
}
\value{
\code{\link{Colonies-class}} with selected colonies
}
\description{
Level 3 function that selects individual colonies based on
  colony ID or random selection.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1:4], nInd = 100)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = 10)
apiary <- createColonies(basePop[2:5], n = 4)
apiary <- cross(apiary, fathers = fatherGroups[1:4])
getId(apiary)

getId(selectColonies(apiary, ID = 1))
getId(selectColonies(apiary, ID = "5"))
getId(selectColonies(apiary, ID = c(1, 2)))
getId(selectColonies(apiary, ID = c("5", "6")))
getId(selectColonies(apiary, ID = 5))
getId(selectColonies(apiary, ID = "9"))
# ... alternative
getId(apiary[1])
getId(apiary[[1]])
getId(apiary["5"])
getId(apiary[["5"]])
getId(apiary[c(1, 2)])
getId(apiary[c("5", "6")])
getId(apiary[5])
getId(apiary["9"])

getId(selectColonies(apiary, p = 0.5))
getId(selectColonies(apiary, p = 0.5))
getId(selectColonies(apiary, p = 0.5))
}
