% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{selectColonies}
\alias{selectColonies}
\title{Select individual colonies}
\usage{
selectColonies(colonies, ID = NULL, n = NULL, p = NULL)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{ID}{character or numeric, name of a colony (one or more) to be
selected; if character (numeric) colonies are selected based on their name
(position)}

\item{n}{numeric, number of colonies to select}

\item{p}{numeric, probability of a colony being selected (takes precedence
over \code{n})}
}
\value{
\code{\link{Colonies-class}} with selected colonies
}
\description{
Level 3 function that selects individual colonies based on
  colony ID or random selection.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

founderDrones <- createDrones(x = basePop[1:4], nInd = 10)
colony1 <- createColony(queen = basePop[5], fathers = founderDrones[1:10])
colony2 <- createColony(queen = basePop[6], fathers = founderDrones[11:20])
colony3 <- createColony(queen = basePop[7], fathers = founderDrones[21:30])
colony4 <- createColony(queen = basePop[8], fathers = founderDrones[31:40])
apiary <- c(colony1, colony2, colony3, colony4)
getId(apiary)

getId(selectColonies(apiary, ID = 1))
getId(selectColonies(apiary, ID = "5"))
getId(selectColonies(apiary, ID = c(1, 2)))
getId(selectColonies(apiary, ID = c("5", "6")))
getId(selectColonies(apiary, ID = 5))
getId(selectColonies(apiary, ID = "9"))
# ... alternative
getId(apiary[1])
getId(apiary[[1]])
getId(apiary["5"])
getId(apiary[["5"]])
getId(apiary[c(1, 2)])
getId(apiary[c("5", "6")])
getId(apiary[5])
getId(apiary["9"])

getId(selectColonies(apiary, p = 0.5))
getId(selectColonies(apiary, p = 0.5))
getId(selectColonies(apiary, p = 0.5))

}
