% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{setColonyPheno}
\alias{setColonyPheno}
\title{Set colony phenotype}
\usage{
setColonyPheno(x, FUN = NULL, ..., simParamBee = NULL)
}
\arguments{
\item{x}{\code{\link{Colony-class}} or \code{\link{MultiColony-class}}}

\item{FUN}{function, any function that can be run on \code{colony} and
returns colony phenotypes; if \code{NULL} then
\code{\link{SimParamBee}$colonyPheno} is used - if even this is \code{NULL},
then colony phenotype is not set, but phenotypes of colony individuals are}

\item{...}{all arguments of \code{\link{setPheno}} and \code{FUN}}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colony-class}} or \code{\link{MultiColony-class}} with phenotypes
}
\description{
Level 2 function that sets phenotypes for all Colony or MultiColony individuals
  (queen, workers, drones, and virgin queens) and for the colony or each colony in
  a MultiColony object
}
\details{
When this function is called on a colony, phenotypes for all colony
  individuals is set or reset if phenotypes already exist.
}
\examples{
founderGenomes <- quickHaplo(nInd = 5, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)

# Define two traits that collectively affect colony honey yield:
# 1) queen's effect on colony honey yield
# 2) workers' effect on colony honey yield
# The traits will have negative genetic correlation and heritability of 0.25
meanP <- c(20, 0)
varA <- c(1, 1 / 10)
corA <- matrix(data = c(
  1.0, -0.5,
  -0.5, 1.0
), nrow = 2, byrow = TRUE)
varE <- c(3, 3 / 10)
varA / (varA + varE)
SP$addTraitA(nQtlPerChr = 100, mean = meanP, var = varA, corA = corA)
SP$setVarE(varE = varE)

basePop <- createVirginQueens(founderGenomes)
drones <- createDrones(x = basePop[1], nInd = 100)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 5, nFathers = 14)

# Create and cross Colony and MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(colony, fathers = fatherGroups[[1]])
colony <- buildUp(colony, nWorkers = 10)
apiary <- createMultiColony(basePop[3:5], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])
apiary <- buildUp(apiary, nWorkers = 10)

# Set phenotypes for all colony individuals
colony <- setColonyPheno(colony)
apiary <- setColonyPheno(apiary)

# Queen's phenotype for both traits
pheno(getQueen(colony))
lapply(getQueen(apiary), FUN = pheno)
# TODO: use getQueensPheno(colony, caste = "queen")
#       https://github.com/HighlanderLab/SIMplyBee/issues/26

# Workers' phenotype for both traits
pheno(getWorkers(colony))
lapply(getWorkers(apiary), FUN = pheno)
# TODO: use getWorkersPheno(colony, caste = "queen")
#       https://github.com/HighlanderLab/SIMplyBee/issues/26

# TODO: use getColonyPheno(colony) for all individuals
#       https://github.com/HighlanderLab/SIMplyBee/issues/26

# Set phenotypes for all colony individuals AND Colony
colony <- setColonyPheno(colony, FUN = phenoQueenPlusSumOfWorkers)
pheno(colony)
# Set phenotypes for all colony individuals AND MultiColony
apiary <- setColonyPheno(apiary, FUN = phenoQueenPlusSumOfWorkers)
lapply(apiary@colonies, FUN = pheno)
# TODO: use getColonyPheno(colony) for all individuals and/or colony
#       https://github.com/HighlanderLab/SIMplyBee/issues/26

# Colony phenotype - store the colony function into the SP object
SP$colonyPheno <- phenoQueenPlusSumOfWorkers
pheno(setColonyPheno(colony))
pheno(setColonyPheno(colony))
lapply(setColonyPheno(apiary)@colonies, FUN = pheno)
lapply(setColonyPheno(apiary)@colonies, FUN = pheno)
# phenotype will vary between function calls

# TODO:
# See
#     https://github.com/HighlanderLab/SIMplyBee/issues/26
#     https://github.com/HighlanderLab/SIMplyBee/issues/28
#     https://github.com/HighlanderLab/SIMplyBee/issues/32
}
