% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{setPhenoColonies}
\alias{setPhenoColonies}
\title{setPhenoColonies}
\usage{
\method{setPhenoColonies}(colonies, FUN, ...)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{...}{}
}
\value{
An updated AlphaSimRBee Colonies object
}
\description{
:
}
