% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{setPhenoColonies}
\alias{setPhenoColonies}
\title{TODO
https://github.com/HighlanderLab/SIMplyBee/issues/158}
\usage{
setPhenoColonies(colonies, FUN = NULL, ..., simParamBee = NULL)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{FUN}{TODO}

\item{...}{TODO}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
\code{\link{Colonies-class}} with phenotypes
}
\description{
Level 3 function that TODO
}
\examples{
# TODO

}
