% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{setQueensYearOfBirth}
\alias{setQueensYearOfBirth}
\title{Set the queen's year of birth}
\usage{
setQueensYearOfBirth(x, year)
}
\arguments{
\item{x}{\code{\link{Pop-class}} (one or more than one queen),
\code{\link{Colony-class}} (one colony), or
\code{\link{MultiColony-class}} (more colonies)}

\item{year}{integer, the year of the birth of the queen}
}
\value{
\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
  \code{\link{MultiColony-class}} with queens having the year of birth set
}
\description{
Level 1 function that sets the queen's year of birth.
}
\examples{
founderGenomes <- quickHaplo(nInd = 8, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
SP$setTrackRec(TRUE)
SP$setTrackPed(isTrackPed = TRUE)
SP$addTraitA(10)
SP$addSnpChip(5)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 1000)
fatherGroups <- pullDroneGroupsFromDCA(drones, n = 10, nFathers = nFathersPoisson)

# Create a Colony and a MultiColony class
colony <- createColony(x = basePop[2])
colony <- cross(x = colony, fathers = fatherGroups[[1]])
apiary <- createMultiColony(basePop[3:4], n = 2)
apiary <- cross(apiary, fathers = fatherGroups[c(2, 3)])

# Example on Colony class
getQueenYearOfBirth(colony)
getQueenYearOfBirth(apiary)

queen1 <- getQueen(colony)
queen1 <- setQueensYearOfBirth(queen1, year = 2022)
getQueenYearOfBirth(queen1)

colony <- setQueensYearOfBirth(colony, year = 2022)
getQueenYearOfBirth(colony)

apiary <- setQueensYearOfBirth(apiary, year = 2022)
getQueenYearOfBirth(apiary)
}
