% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{setQueensYearOfBirth}
\alias{setQueensYearOfBirth}
\alias{setQueensYOB}
\title{Set the queen's year of birth}
\usage{
setQueensYearOfBirth(x, year)

setQueensYOB(x, year)
}
\arguments{
\item{x}{\code{\link{Pop-class}} (one or more than one queen),
\code{\link{Colony-class}} (one colony), or
\code{\link{Colonies-class}} (more colonies)}

\item{year}{integer, the year of the birth of the queen}
}
\value{
\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
  \code{\link{Colonies-class}} with queens having the year of birth set
}
\description{
Level 1 function that sets the queen's year of birth.
}
\section{Functions}{
\itemize{
\item \code{setQueensYOB}: Set the queen's year of birth
}}

\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)

getQueensYearOfBirth(getQueen(colony1))
getQueensYearOfBirth(c(getQueen(colony1), getQueen(colony2)))
getQueensYearOfBirth(colony1)
getQueensYearOfBirth(apiary)

queen1 <- getQueen(colony1)
queen1 <- setQueensYearOfBirth(queen1, year = 2022)
getQueensYearOfBirth(queen1)

queen2 <- getQueen(colony2)
queens <- setQueensYearOfBirth(c(queen1, queen2), year = 2023)
getQueensYearOfBirth(queens)

colony1 <- setQueensYearOfBirth(colony1, year = 2022)
getQueensYearOfBirth(colony1)

apiary <- setQueensYearOfBirth(apiary, year = 2022)
getQueensYearOfBirth(apiary)

}
