% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{setQueensYearOfBirth}
\alias{setQueensYearOfBirth}
\alias{setQueensYOB}
\title{Set the queen's year of birth}
\usage{
setQueensYearOfBirth(x, year)

setQueensYOB(x, year)
}
\arguments{
\item{x}{\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
\code{\link{Colonies-class}}}

\item{year}{integer, the year of the birth of the queen}
}
\value{
\code{\link{Pop-class}}, \code{\link{Colony-class}}, or
  \code{\link{Colonies-class}}
}
\description{
Level 1 function that sets the queen's year of birth.
}
\section{Functions}{
\itemize{
\item \code{setQueensYOB}: Set the queen's year of birth
}}

\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
apiary <- c(colony1, colony2)

getQueensYOB(getQueen(colony1))
getQueensYOB(colony1)
getQueensYOB(apiary)

queen <- getQueen(colony1)
queen <- setQueensYOB(queen, year = 2021)
getQueensYOB(queen)

colony1 <- setQueensYOB(colony1, year = 2021)
getQueensYOB(colony1)

apiary <- setQueensYOB(apiary, year = 2021)
getQueensYOB(apiary)

queen <- setQueensYearOfBirth(queen, year = 2022)
getQueensYearOfBirth(queen)

colony1 <- setQueensYearOfBirth(colony1, year = 2022)
getQueensYearOfBirth(colony1)

apiary <- setQueensYearOfBirth(apiary, year = 2022)
getQueensYearOfBirth(apiary)

}
