% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{simulateHoneyBeeGenomes}
\alias{simulateHoneyBeeGenomes}
\title{Simulate the Honey bee genome}
\usage{
simulateHoneyBeeGenomes(
  nMelN = 0L,
  nMelS = 0L,
  nCar = 0L,
  nLig = 0L,
  ploidy = 2L,
  nChr = 16L,
  nSegSites = 100000L,
  Ne = 170000L,
  nBp = 225200000/16,
  genLen = 3.199121,
  mutRate = 3.4e-09,
  recRate = 2.3e-07,
  nThreads = NULL
)
}
\arguments{
\item{nChr}{number of chromosomes to simulate}

\item{nSegSites}{number of segregating sites to keep per chromosome}

\item{Ne}{effective population size}

\item{nBp}{base pair length of chromosome}

\item{genLen}{genetic length of chromosome in Morgans}

\item{mutRate}{per base pair mutation rate}

\item{nThreads}{if OpenMP is available, this will allow for simulating
chromosomes in parallel. If the value is NULL, the number of threads is
automatically detected}

\item{nInd}{number of individuals to simulate}

\item{histNe}{effective population size in previous generations}

\item{histGen}{number of generations ago for effective population sizes given
in histNe}

\item{split}{an optional historic population split in terms of generations
ago}
}
\value{
\code{\link{MapPop-class}}
}
\description{
Level 0 function that returns simulated honeybee genomes
}
\examples{
founderGenomes <- simulateHoneyBeeGenomes(
  nInd = 10, nChr = 1,
  nSegSites = 2, Ne = 10
)
}
