% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{simulateHoneyBeeGenomes}
\alias{simulateHoneyBeeGenomes}
\title{Simulate the Honey bee genome}
\usage{
simulateHoneyBeeGenomes(
  nMelN = 0L,
  nMelS = 0L,
  nCar = 0L,
  nLig = 0L,
  ploidy = 2L,
  nChr = 16L,
  nSegSites = 100L,
  nBp = 225200000/16,
  genLen = 3.199121,
  mutRate = 3.4e-09,
  recRate = 2.3e-07,
  nThreads = NULL
)
}
\arguments{
\item{nMelN}{integer, number of Apis mellifera mellifera North individuals to simulate}

\item{nMelS}{integer, number of Apis mellifera mellifera South individuals to simulate}

\item{nCar}{integer, number of Apis mellifera carnica individuals to simulate}

\item{nLig}{integer, number of Apis mellifera ligustica individuals to simulate}

\item{ploidy}{integer, the ploidy of the individuals}

\item{nChr}{integer, number of chromosomes to simulate}

\item{nSegSites}{integer, number of segregating sites to keep per chromosome}

\item{nBp}{integer, base pair length of chromosome}

\item{genLen}{numeric, genetic length of chromosome in Morgans}

\item{mutRate}{numeric, per base pair mutation rate}

\item{recRate}{numeric, per base pair recombination rate}

\item{nThreads}{integer, if OpenMP is available, this will allow for simulating
chromosomes in parallel. If the value is NULL, the number of threads is
automatically detected}
}
\value{
\code{\link{MapPop-class}}
}
\description{
Level 0 function that returns simulated honeybee genomes
}
\examples{
# founderGenomes <- simulateHoneyBeeGenomes(
#   nInd = 10, nChr = 1,
#   nSegSites = 2, Ne = 10
# )
}
