% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{simulateHoneyBeeGenomes}
\alias{simulateHoneyBeeGenomes}
\title{Simulate the Honey bee genome}
\usage{
simulateHoneyBeeGenomes(
  nInd = NULL,
  nChr = 16,
  nSegSites = 1000,
  Ne = 157598,
  nBp = 225200000/16,
  genLen = 34.5/16,
  mutRate = 9e-09,
  histNe = Ne,
  histGen = 1,
  split = NULL,
  nThreads = NULL
)
}
\arguments{
\item{nInd}{number of individuals to simulate}

\item{nChr}{number of chromosomes to simulate}

\item{nSegSites}{number of segregating sites to keep per chromosome}

\item{Ne}{effective population size}

\item{nBp}{base pair length of chromosome}

\item{genLen}{genetic length of chromosome in Morgans}

\item{mutRate}{per base pair mutation rate}

\item{histNe}{effective population size in previous generations}

\item{histGen}{number of generations ago for effective population sizes given
in histNe}

\item{split}{an optional historic population split in terms of generations
ago}

\item{nThreads}{if OpenMP is available, this will allow for simulating
chromosomes in parallel. If the value is NULL, the number of threads is
automatically detected}
}
\value{
\code{\link{MapPop-class}}
}
\description{
Level 0 function that returns simulated honeybee genomes
}
\examples{
founderGenomes <- simulateHoneyBeeGenomes(nInd = 2)

}
