% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{splitColonies}
\alias{splitColonies}
\title{Split colony in two colonies for all given colonies}
\usage{
splitColonies(colonies, p = 0.3, year = NULL)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{p}{numeric, percentage of workers that will go to the split colony}

\item{year}{numeric, year of birth for virgin queens}
}
\value{
list with two \code{\link{Colonies-class}}, the \code{splits} and the
  \code{remnants} (see the description of \code{\link{splitColony}} what each
  node holds!)
}
\description{
Level 3 function that does the same as \code{\link{splitColony}}
  but for all given colonies.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony1 <- buildUpColony(colony1, nWorkers = 100)
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony2 <- buildUpColony(colony2, nWorkers = 100)
apiary <- c(colony1, colony2)
apiary
apiary[[1]]
apiary[[2]]

tmp <- splitColonies(apiary)
tmp$splits
tmp$splits[[1]]
tmp$splits[[2]]
hasSplit(tmp$splits)
tmp$remnants
tmp$remnants[[1]]
tmp$remnants[[2]]
hasSplit(tmp$remnants)

}
