% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{splitColonies}
\alias{splitColonies}
\title{Split colony in two colonies for all given colonies}
\usage{
splitColonies(colonies, p = NULL, year = NULL, simParamBee = NULL)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}

\item{p}{numeric, proportion of workers that will go to the split colony; if
\code{NULL} then \code{\link{SimParamBee}$splitP} is used}

\item{year}{numeric, year of birth for virgin queens}

\item{simParamBee}{\code{\link{SimParamBee}}, global simulation parameters}
}
\value{
list with two \code{\link{Colonies-class}}, the \code{splits} and the
  \code{remnants} (see the description of \code{\link{splitColony}} what each
  node holds!)
}
\description{
Level 3 function that does the same as \code{\link{splitColony}}
  but for all given colonies.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- createVirginQueens(founderGenomes)

drones <- createDrones(x = basePop[1], nInd = 10)
colony1 <- createColony(x = basePop[2])
colony1 <- crossColony(colony1, drones = drones[1:5], nFathers = 5)
colony1 <- buildUpColony(colony1, nWorkers = 100)
colony2 <- createColony(x = basePop[3])
colony2 <- crossColony(colony2, drones = drones[6:10], nFathers = 5)
colony2 <- buildUpColony(colony2, nWorkers = 100)
apiary <- c(colony1, colony2)
apiary
apiary[[1]]
apiary[[2]]

tmp <- splitColonies(apiary)
tmp$splits
tmp$splits[[1]]
tmp$splits[[2]]
hasSplit(tmp$splits)
tmp$remnants
tmp$remnants[[1]]
tmp$remnants[[2]]
hasSplit(tmp$remnants)
}
