% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{supersedeColonies}
\alias{supersedeColonies}
\title{Supersede colony for all given colonies}
\usage{
supersedeColonies(colonies)
}
\arguments{
\item{colonies}{\code{\link{Colonies-class}}}
}
\value{
\code{\link{Colonies-class}} with superseded colonies
}
\description{
Level 3 function that does the same as
  \code{\link{supersedeColony}} but for all given colonies.
}
\examples{
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- newPop(founderGenomes)

drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony1 <- buildUpColony(colony1, nWorkers = 100)
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony2 <- buildUpColony(colony2, nWorkers = 100)
apiary <- c(colony1, colony2)
apiary
apiary[[1]]
apiary[[2]]

apiary <- supersedeColonies(apiary)
apiary
apiary[[1]]
apiary[[2]]
hasSuperseded(apiary)

}
