% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{supersedeColony}
\alias{supersedeColony}
\title{Supersede colony}
\usage{
supersedeColony(colony, year = NULL)
}
\arguments{
\item{colony}{\code{\link{Colony-class}}}

\item{year}{numeric, year of birth for virgin queens}
}
\value{
\code{\link{Colony-class}} with the supersede event set to
  \code{TRUE}
}
\description{
Level 2 function that supersedes colony - an event where the
  queen dies. The workers and drones stay unchanged, but workers raise virgin
  queens, of which only one prevails.
}
\examples{
founderGenomes <- quickHaplo(nInd = 2, nChr = 1, segSites = 100)
SP <- SimParamBee$new(founderGenomes)
basePop <- asVirginQueen(newPop(founderGenomes))

drones <- createDrones(x = basePop[1], nInd = 10)
colony <- createColony(queen = basePop[2], fathers = drones)
(colony <- buildUpColony(colony, nWorkers = 100))

supersedeColony(colony)

}
