% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{supersedeColony}
\alias{supersedeColony}
\title{Replicates a supersedure of the colony and replaces the queen with a virgin queen.}
\usage{
\method{supersedureColony}(colony, crossVirginQueen, fathers, nWorkers, nDrones)
}
\arguments{
\item{colony}{Colony class. AlphaSimRBee Colony object from the \code{createColony(...)} call}

\item{crossVirginQueen}{Logical. Whether a virgin queen is to be mated}

\item{fathers}{AlphaSimR population object. Number of fathers pulled from the DCA}

\item{pWorkers}{Numeric, proportion of workers that are replaced with the workers from the new queen}

\item{pDrones}{Numeric, proportion of drones that are replaced with the drones from the new queen}
}
\value{
Single AlphaSim population object of superseded colony
}
\description{
Replicates the process of supersedure, where the
queen is replaced by a new virgin queen. The workers and the drones stay
in the colony. If no fathers are present, mating of the virgin queen does not occur.
}
\examples{
inst/examples/examples_supersedeColony.R
}
\seealso{
\code{\link[??????]{supersedure}}
}
